% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_utils.R
\name{add.barcodes}
\alias{add.barcodes}
\title{Add barcodes in the colData}
\usage{
add.barcodes(sfe)
}
\arguments{
\item{sfe}{The SpatialFeaturesExperiment object.}
}
\value{
The modified SpatialFeaturesExperiment object with added barcode
columns.
}
\description{
IMPORTANT: this function is not exported to be used alone.
A function to add the barcodes in the \code{colData} slot of a
\code{SpatialFeaturesExperiment} (SFE) object. In case of multiple samples
the \code{read10xVisiumSFE} function adds a suffix in the barcodes to make
them unique so that they can be used as col/rownames. This function takes
care of this by removing the suffix when transferring the barcodes to a
column.
}
\details{
This function adds a new column named "Barcode" to the \code{colData} slot
of the given \code{SpatialFeaturesExperiment} (SFE) object. The function
moves the existing row names to this new column. If the barcodes in the
\code{colData} have a suffix in the format "-(0-9)-(0-9)", which is added
by the \code{read10xVisiumSFE} function to make them unique for multiple
samples, this function removes the suffix from the barcodes in the new
column. Additionally, the function extracts the capture area index from the
barcodes and adds it as a new column named "Capt_area" in the \code{colData}.
}
\seealso{
\code{\link{read10xVisiumSFE}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
