% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.gTruth.R
\name{add.gTruth}
\alias{add.gTruth}
\title{Add ground truth annotation to SpatialFeaturesExperiment object}
\usage{
add.gTruth(sfe, gtruth)
}
\arguments{
\item{sfe}{The SpatialFeaturesExperiment object.}

\item{gtruth}{A dataframe containing the ground truth for the dataset.
It should have at least three columns: "Barcode" (spot barcodes matching
the colnames of the SFE object), "sample_id" (sample ID for each spot,
important for multiple samples), and the ground truth annotation column.}
}
\value{
The modified SpatialFeaturesExperiment object with the ground truth
annotation added to the colData slot.
}
\description{
IMPORTANT: this function is not exported to be used alone.
A function to add the ground truth annotation in the colData slot of a
SpatialFeaturesExperiment (SFE) object. Multiple annotations can be imported
using this approach, but note that if not all spots are annotated, NAs will
be imported. The SFE object has other slots for adding extra annotations that
may not cover the entire dataset.
}
\details{
This function adds the ground truth annotation to the colData slot of a
SpatialFeaturesExperiment object. It performs a merge between the colData of
the SFE object and the provided ground truth dataframe based on "Barcode" and
"sample_id". The merge includes all spots, and any missing annotations are
represented as NAs in the merged colData. The row names of the colData are
temporarily stored in a separate column before merging and then set back as
row names after the merge operation.
}
\seealso{
\code{\link{add.index}}: A function to add an index to the colData slot of a
SpatialFeaturesExperiment object.
\code{\link{add.barcodes}}: A function to add barcodes to the colData slot of
a SpatialFeaturesExperiment object.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
