% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_utils.R
\name{add.index}
\alias{add.index}
\title{Add index to SpatialFeaturesExperiment object}
\usage{
add.index(sfe)
}
\arguments{
\item{sfe}{The SpatialFeaturesExperiment object.}
}
\value{
The modified SpatialFeaturesExperiment object with the index added to
the colData slot.
}
\description{
IMPORTANT: this function is not exported to be used alone.
A function to add the index in the colData slot of a
SpatialFeaturesExperiment (SFE) object. The index serves as a unique
identifier for every spot, even if there are multiple samples in the SFE
object.
}
\details{
This function adds an index column to the colData slot of a
SpatialFeaturesExperiment object. The index is generated as "spot_" followed
by a numeric sequence, starting from 1 and incrementing for each spot.
The index provides a unique identifier for every spot, ensuring each spot can
be uniquely identified even in the presence of multiple samples.
}
\seealso{
\code{\link{add.barcodes}}: A function to add barcodes to the colData slot of
a SpatialFeaturesExperiment object.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
