% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_utils.R
\name{add.spotCntd}
\alias{add.spotCntd}
\title{Generate centroids from spot coordinates in a SpatialFeaturesExperiment
object}
\usage{
add.spotCntd(sfe, sample_id)
}
\arguments{
\item{sfe}{The SpatialFeaturesExperiment object.}

\item{sample_id}{A character string specifying unique sample identifiers, one
or each directory specified via \code{samples} when you loaded the SFE object
using the \code{read10xVisiumSFE} function.}
}
\value{
The modified SpatialFeaturesExperiment object with the centroid
coordinates stored in the colGeometries slot.
}
\description{
IMPORTANT: this function is not exported to be used alone.
A function to generate centroids from the spot coordinates. It takes the
coordinates out of the \code{spatialCoords} slot of a
SpatialFeaturesExperiment (SFE) object and stores it inside the colGeometries
slot.
}
\details{
This function extracts the spot coordinates from the \code{spatialCoords}
slot of the SFE object, generates centroids from the coordinates, and stores
the resulting centroids in the \code{colGeometries} slot of the SFE object.
The centroid coordinates are represented as an \code{sf} object, with each
centroid corresponding to a spot in the SFE object. The row names of the
\code{cntds} object are set to match the column names of the SFE object for
easy reference and linking between the two.
}
\seealso{
\code{\link{add.index}}: A function to add an index in the colData slot of a
SpatialFeaturesExperiment object.
\code{\link{add.barcodes}}: A function to add barcodes in the colData slot of
a SpatialFeaturesExperiment object.
\code{\link{add.gTruth}}: A function to add ground truth annotation in the
colData slot of a SpatialFeaturesExperiment object.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
