% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_utils.R
\name{add.spotHex}
\alias{add.spotHex}
\title{Add Hexagonal geometries}
\usage{
add.spotHex(
  sfe,
  samples,
  sample_id,
  res = c("lowres", "hires", "fullres"),
  flipped
)
}
\arguments{
\item{sfe}{The \code{SpatialFeaturesExperiment} object.}

\item{samples}{The samples' folder path as provided at
\code{\link{read10xVisiumSFE}}.}

\item{sample_id}{A character string specifying unique sample identifiers, one
or each directory specified via \code{samples} when you loaded the SFE object
using the \code{read10xVisiumSFE} function.}

\item{res}{The desired resolution. Can take one of the following values:
"lowres", "hires", "fullres".}
}
\description{
IMPORTANT: this function is not exported to be used alone.
This function adds hexagons in the
\code{SpatialFeatureExperiment} (SFE) object utilizing a perimeter of
off-tissue spots that surround the Visium slide spots. This perimeter assists
the Voronoi tessellation process by removing any instances where on-tissue
spots at the edges end up without a polygon due to the way the tessellation
is calculated. The function also provides options to adjust the perimeter
addition based on the tilt of pixel coordinates.
}
\seealso{
\code{\link{read10xVisiumSFE}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
