% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistMat.R
\name{addDistMat}
\alias{addDistMat}
\title{Add Distance Matrix}
\usage{
addDistMat(msfe, p, sample_id = TRUE, ...)
}
\arguments{
\item{msfe}{\code{MetaSpatialFeatureExperiment} (Meta-SFE) object.}

\item{p}{Numeric scalar specifying the power parameter for the Minkowski
distance. See details below for more information.}

\item{sample_id}{character string or TRUE specifying sample/image
identifier(s); here, TRUE is equivalent to all samples/images.}

\item{...}{Additional arguments to be passed to \code{gw.dist} function.}
}
\value{
An updated SpatialFeatureExperiment object with the distance matrix
stored in the metadata.
}
\description{
Calculate and store the distance matrix based on spatial coordinates using
the specified metric.
}
\details{
This function calculates the distance matrix based on the spatial coordinates
of the provided SpatialFeatureExperiment object. The distance matrix is
computed using the specified metric, where p = 2 corresponds to Euclidean
distance, p = 1 corresponds to Manhattan distance, and p > 2 corresponds to
Minkowski distance with the specified power. The resulting distance matrix is
stored in the metadata of the SpatialFeatureExperiment object.
}
\examples{
# Load sfe object
data(sfe)

# Add distance matrix to sfe
sfe <- addDistMat(sfe, p = 2)


}
\seealso{
\code{\link[=getDistMat]{getDistMat()}} for a function to get the distance matrix from within
an SFE object.
}
\author{
Eleftherios (Lefteris) Zormpas
}
