% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metaSFE.R
\name{addMultipleSFE}
\alias{addMultipleSFE}
\title{Method to add an SFE object with multiple samples to the
MetaSpatialFeatureExperiment}
\usage{
addMultipleSFE(x, sfe)
}
\arguments{
\item{x}{An instance of the MetaSpatialFeatureExperiment class.}

\item{sfe}{An instance of the SpatialFeatureExperiment class with multiple
samples.}
}
\value{
An updated MetaSpatialFeatureExperiment object with the added SFEs.
}
\description{
Method to add an SFE object with multiple samples to the
MetaSpatialFeatureExperiment
}
