% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPerLocQC.R
\name{addPerLocQC}
\alias{addPerLocQC}
\title{Add location-related QC metrics}
\usage{
addPerLocQC(
  sfe,
  gTruth = NULL,
  assay = "counts",
  in_tissue = TRUE,
  MARGIN,
  ...
)
}
\arguments{
\item{sfe}{The SpatialFeatureExperiment object.}

\item{gTruth}{A dataframe that contains the ground truth for your dataset.
It needs to have at least 3 columns. One column named "Barcode" with the
spot Barcodes (these need to match the colnames of the SFE object), one
column named "sample_id" which includes the sample ID for each spot (this is
important when you import multiple samples) and one column with the ground
truth annotation.}

\item{assay}{the name of the assay to use. Defaults to 'counts'.}

\item{MARGIN}{1 = Features, 2 = Locations. Indicates for which aspect the
sparsity will be calculated for; features (genes) or locations.}

\item{...}{further arguments passed to \code{addPerCellQC}, to pass to
\code{perCellQCMetrics.}}
}
\description{
A function to add a series of location (spot)-related QC metrics..
}
\examples{
# Load data
data(sfe_raw)
data(gTruth)

# Calculate location QC stats

sfe <- addPerLocQC(sfe_raw, gTruth = gTruth, assay = "counts", MARGIN = 2)


}
\seealso{
\code{\link{addPerCellQC}}, \code{\link{add.barcodes}},
\code{\link{add.gTruth}}, \code{\link{add.index}},
\code{\link{get.QC.Sparsity}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
