% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metaSFE.R
\name{addSFE}
\alias{addSFE}
\title{Method to add a SpatialFeatureExperiment to the MetaSpatialFeatureExperiment}
\usage{
addSFE(x, sfe, sample_id = TRUE)
}
\arguments{
\item{x}{An instance of the MetaSpatialFeatureExperiment class.}

\item{sfe}{An instance of the SpatialFeatureExperiment class to add.}

\item{sample_id}{Currently only accepts TRUE. Internally performs automatic
selection of the sample ID in the SFE object. Assumes that the SFE object
includes only one sample inside. If multiple samples are stored in the same
SFE object, please use the \code{addMultipleSFE} instead.}
}
\value{
An updated MetaSpatialFeatureExperiment object with the added SFE.
}
\description{
Method to add a SpatialFeatureExperiment to the MetaSpatialFeatureExperiment
}
