% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotateDataFrame}
\alias{annotateDataFrame}
\title{Annotate a data frame}
\usage{
annotateDataFrame(
  data,
  biomart,
  add,
  id = 1,
  column = NA,
  check.names = FALSE,
  output = "data.frame",
  annot.col = NULL
)
}
\arguments{
\item{data}{A data frame with an ENSGID column or rownames as input.}

\item{biomart}{Biomart annotations of any organism. Can be generated using
the \code{createBiomart} function.}

\item{add}{a vector of biomart columns to add to the input data frame.
Defaults to gene_name, biotype and description.}

\item{id}{position or name of column in biomart that matches the ENSGIDs
column or the rows names in input the data frame.}

\item{column}{IT MUST BE AN INTEGER: the number of the ENSGIDs column from
the input data frame. If left to NA then the function will look to match the
rownames. If no row names with ENSGIDs exist then returns an error.}

\item{check.names}{argument for \code{data.frame} function at the end.}

\item{output}{takes one of two strings; "data.frame" (default) and "vector".
The "data.frame" option will output the input data with the annotation
columns added. The "vector" option will output a vector with a single
annotation type for the genes present in the input data.}

\item{annot.col}{the annotation column to be returned as vector. Only needed
if output = "vector"}
}
\value{
An annotated data frame or vector based on the specified \code{output}
argument.
}
\description{
A function to annotate any data frame of data or biological data that has a
column of ENSGene IDs.
}
\details{
This function annotates a data frame (\code{data}) using a biomart database
(\code{biomart}). It matches the rows of the data frame with the corresponding
entries in the biomart based on the provided identifier (\code{id}) in the
biomart. If the data frame has a column with ENGene IDs then that specific
column (\code{column}) must be specified as an integer.

If no specific biomart columns are provided in the \code{add} argument, the
function will add the following default columns: "gene_name", "biotype",
"description". If the biomart contains a column named "human_homolog", it
will be included in the annotation as well.

If there is a successful match between the data frame and the biomart, the
annotated data is returned as a data frame or a vector based on the
specified \code{output} argument. The output data frame includes the original
data, the matching biomart columns, and an identifier column (\code{id})
containing the row names of the input data frame.

If some rows in the data frame are missing from the biomart table, a warning
message is displayed, indicating that the biomart might have been created
using a different Ensembl version. Users are advised to verify the Ensembl
version used for annotation and ensure consistency with the biomart table.
}
\examples{
library(SpatialFeatureExperiment)

# Load data
data(sfe)
df <- rowData(sfe)[,c(5:8)]

# Create a biomart
biomart <- createBiomart(organism = "human")

# Annotate tha dataframe
annotated_df <- annotateDataFrame(df, biomart = biomart)

}
\seealso{
\code{\link{createBiomart}}, \code{\link[tibble]{as_tibble}} to
convert the output to a tibble if necessary. \code{\link[biomaRt]{useMart}}
and \code{\link[biomaRt]{getBM}} to access and retrieve data from the biomart
database.
}
\author{
Eleftherios (Lefteris) Zormpas
}
