% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotateVector}
\alias{annotateVector}
\title{Annotate a vector}
\usage{
annotateVector(
  data,
  biomart,
  add,
  id = 1,
  check.names = FALSE,
  output = "data.frame",
  annot.col = NULL
)
}
\arguments{
\item{data}{A vector with ENSGIDs as input.}

\item{biomart}{Biomart annotations of any organism. Can be generated using
the \code{\link{createBiomart}} function from this package or the native
functions from the \code{biomaRt} package.}

\item{add}{A vector of biomart columns to add to the input data frame.
Defaults to "gene_name", "biotype", and "description".}

\item{id}{Position or name of the column in biomart that matches the ENSGIDs
column or the row names in the input data frame.}

\item{check.names}{Argument for \code{data.frame} function at the end.}

\item{output}{Takes one of two strings: "data.frame" (default) and "vector".
The "data.frame" option will output the input vector alongside the annotation
columns in a data.frame. The "vector" option will output a vector with a
single annotation type for the genes present in the input vector.}

\item{annot.col}{The annotation column to be returned as a vector. Only
needed if \code{output = "vector"}.}
}
\description{
A function to annotate any vector of ENSGene IDs.
}
\examples{
data <- c("ENSG00000139618", "ENSG00000141510", "ENSG00000198746")
biomart <- createBiomart(organism = "human")
annotated_vector <- annotateVector(data, biomart = biomart)

}
\seealso{
\code{\link{createBiomart}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
