% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_applyQCthresh.R
\name{applyQCthresh_feat}
\alias{applyQCthresh_feat}
\title{Apply Quality Control Thresholds to MetaSpatialFeatureExperiment}
\usage{
applyQCthresh_feat(msfe)
}
\arguments{
\item{msfe}{A \code{MetaSpatialFeatureExperiment} object.}
}
\value{
A new \code{MetaSpatialFeatureExperiment} object with features that
pass the QC thresholds.
}
\description{
This function applies quality control (QC) thresholds to a
\code{MetaSpatialFeatureExperiment} object. It filters out features (genes
or other genomic elements) that do not meet the specified QC criteria,
creating a new \code{MetaSpatialFeatureExperiment} with the filtered
features.
}
\details{
This function performs quality control on a
\code{MetaSpatialFeatureExperiment} object, filtering out features that do
not meet the specified QC criteria. The QC criteria are defined based on the
internal \code{.int_discardFeat} function.
}
\examples{
\dontrun{# Apply QC thresholds to a MetaSpatialFeatureExperiment object
msfe <- # A metaSFE object for which filters have been defined in rowData
msfe_filtered <- applyQCthresh_feat(msfe)
}

}
\seealso{
\code{\link{.int_discardFeat}} for details on the internal QC
filtering function.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{feature}
\keyword{quality-control}
\keyword{spatial}
\keyword{transcriptomics}
