% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyQCthresh_.R
\name{applyQCthresh_loc}
\alias{applyQCthresh_loc}
\alias{applyQCthresh}
\title{Apply Quality Control Thresholds to SpatialFeatureExperiment}
\usage{
applyQCthresh_loc(sfe, sample_id = TRUE)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} object.}

\item{sample_id}{A character string, \code{TRUE}, or \code{NULL} specifying
sample/image identifier(s). If \code{TRUE}, all samples/images are
considered. If \code{NULL}, the first available entry is considered.}
}
\value{
A new \code{SpatialFeatureExperiment} object with features that pass
the QC thresholds.
}
\description{
This function applies quality control (QC) thresholds to a
\code{SpatialFeatureExperiment} object. It filters out features (genes or
other genomic elements) that do not meet the specified QC criteria, creating
a new \code{SpatialFeatureExperiment} with the filtered features.
}
\details{
This function performs quality control on a
\code{SpatialFeatureExperiment} object, filtering out features that do not
meet the specified QC criteria. The QC criteria are defined based on the
internal \code{.int_filterQCDiscard} function.
}
\examples{
\dontrun{
# Apply QC thresholds to a SpatialFeatureExperiment object
sfe  # An SFE object for which filters have been defined in colData
sfe_filtered <- applyQCthresh_loc(sfe, sample_id = TRUE)
}

}
\seealso{
\code{\link{.int_filterQCDiscard}} for details on the internal QC
filtering function.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{feature}
\keyword{quality-control}
\keyword{spatial}
\keyword{transcriptomics}
