% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLibSizeFactors.R
\name{computeLibSizeFactors}
\alias{computeLibSizeFactors}
\alias{.int_sizeFactCalc}
\alias{.int_summaryCalc}
\alias{.int_summarySizeFact}
\title{Compute Library Size Factors in a SpatialFeatureExperiment}
\usage{
computeLibSizeFactors(sfe, type = c("inter", "intra"), ...)

.int_sizeFactCalc(id, sfe, ...)

.int_summaryCalc(id, sfe)

.int_summarySizeFact(sfe, ids)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} object.}

\item{type}{Character string, one of "inter" or "intra", indicating whether
to calculate size factors within each sample ("intra") or across all samples
("inter"). The default is "intra".}

\item{...}{Additional arguments passed to \code{\link[scater]{computeLibraryFactors}}.}

\item{id}{Character string, specifying the sample/image identifier.}

\item{ids}{A character vector specifying sample/image identifiers.}
}
\value{
A modified \code{SpatialFeatureExperiment} object with computed
library size factors.

A modified \code{SpatialFeatureExperiment} object with computed library size factors.

A summary of library size factors for the specified sample.

A data frame summarizing library size factors for multiple samples.
}
\description{
This function computes library size factors in a
\code{SpatialFeatureExperiment} object. Size factors represent a scaling
factor for each sample, adjusting for differences in sequencing depth.

Compute Library Size Factors for a Specific Sample

Summary of Library Size Factors for a Specific Sample

This function computes a summary of library size factors for multiple
samples in a spatial transcriptomics experiment.
}
\details{
This function calculates library size factors for each sample
individually ("intra") or for all samples together ("inter"). The type of
calculation is determined by the \code{type} argument.

This function calculates library size factors for a specific sample identified by
the \code{id} argument using the \code{\link[scater]{computeLibraryFactors}} function.

This function provides a summary of library size factors for a
specific sample identified by the \code{id} argument. It extracts the size
factors using the \code{\link[scater]{librarySizeFactors}} function and
summarises them.

The function internally uses the
\code{\link[scater]{librarySizeFactors}} function for calculating library
size factors and the \code{\link{.int_summaryCalc}} function to generate
summaries for each sample. The results are then combined into a data frame.

This function efficiently computes a summary of library size factors for
multiple samples in a spatial transcriptomics experiment. It leverages the
capabilities of the internal function \code{.int_summaryCalc}, ensuring a
systematic and comprehensive overview of library size factors.

The data is organized into a data frame with rows corresponding to sample
identifiers and columns representing summary statistics for library size
factors.

The function uses the \code{\link[scater]{librarySizeFactors}} function for
size factor calculations and the internal function \code{.int_summaryCalc}
for generating individual sample summaries. This function is a crucial step
in understanding the distribution and characteristics of library size
factors in spatial transcriptomics data, providing valuable insights for
downstream analyses.
}
\examples{
\dontrun{# Compute library size factors within each sample
sfe # An SFE object
sfe <- computeLibSizeFactors(sfe, type = "intra")

# Compute library size factors across all samples
sfe # An SFE object
sfe <- computeLibSizeFactors(sfe, type = "inter")
}

}
\seealso{
\code{\link[scater]{computeLibraryFactors}}

\code{\link[scater]{computeLibraryFactors}}

\code{\link[scater]{librarySizeFactors}}

\code{\link[scater]{librarySizeFactors}},
\code{\link{.int_summaryCalc}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{factors}
\keyword{factors,}
\keyword{library}
\keyword{size}
\keyword{spatial}
\keyword{summary}
\keyword{transcriptomics}
\keyword{transcriptomics,}
