% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{createBiomart}
\alias{createBiomart}
\title{Create a BioMart table}
\usage{
createBiomart(
  organism = "human",
  attributes,
  version = NULL,
  patch = FALSE,
  mirror = NULL,
  ...
)
}
\arguments{
\item{organism}{The first letter of the genus and full species name, e.g.,
"hsapiens" for human. A list of common names can be found in
\code{listEnsembl}, or you can use accepted names for human, mouse, rat,
zebrafish, fruitfly, pig, worm, and yeast.}

\item{attributes}{A vector of column names to pass to \code{getBM}. The
default attributes include ensembl_gene_id, external_gene_name, gene_biotype,
chromosome_name, start_position, end_position, strand, description, and
transcript_count.}

\item{version}{Ensembl version for previous releases.}

\item{patch}{Keep features on patches starting with CHR_. Default is FALSE.}

\item{mirror}{Specify an Ensembl mirror to connect to. Valid options are
'www', 'uswest', 'useast', 'asia'. If no mirror is specified, the primary
site at www.ensembl.org will be used. Mirrors are not available for the
Ensembl Genomes databases.}

\item{...}{Additional options like filters and values passed to \code{getBM}
or \code{listAttributes}.}
}
\value{
A data frame containing the biomart table.
}
\description{
A function that creates biomart tables containing Ensembl
annotations using the biomaRt package.
}
\details{
The function downloads the biomart for the specified organism and
version, or the latest release if no version is provided. It performs a
default search using the attributes specified or the default attributes.
The resulting biomart table is returned as a data frame.
}
\examples{
# Create Human BioMart
bm <- createBiomart(organism = "human", version = NULL)
head(bm)

# Create Mouse BioMart
bm <- createBiomart(organism = "mouse",
attributes = c("ensembl_gene_id","external_gene_name",
"gene_biotype", "chromosome_name"),
version = NULL)
head(bm)

}
\seealso{
\code{\link{listEnsembl}}, \code{\link{useEnsembl}},
\code{\link{getBM}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
