% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.QC.Sparsity.R
\name{.get.Global.Sparsity}
\alias{.get.Global.Sparsity}
\title{Fetch global sparsity stats}
\usage{
.get.Global.Sparsity(sfe)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object}
}
\description{
A function to fetch the dataset sparsity attributes from an assay.
}
\keyword{internal}
