% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.int_addLeadingZero}
\alias{.int_addLeadingZero}
\title{Internal: Format with leading zero}
\usage{
.int_addLeadingZero(x)
}
\arguments{
\item{x}{the number}
}
\description{
This function appends the number \code{0} only for the numbers 1 to 9.
}
\author{
Eleftherios (Lefteris) Zormpas
}
