% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{.int_addToFGWCout}
\alias{.int_addToFGWCout}
\title{Internal: Add annotation and geom hex info to fgwc output}
\usage{
.int_addToFGWCout(fgwc, sfe, geom_type)
}
\arguments{
\item{fgwc}{the fgwc output}

\item{sfe}{the sfe object used as input for \code{fgwcSTE}}
}
\value{
description an updated \code{fgwc$finaldata} data frame.
}
\description{
Adding annotation and/or geom hex info allows easier plotting later.
}
\author{
Eleftherios (Lefteris) Zormpas
}
