% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC_.R
\name{.int_annotColours}
\alias{.int_annotColours}
\title{Internal Function: .int_annotColours}
\usage{
.int_annotColours(sfe, colours)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object.}

\item{colours}{Vector of colours for annotations. If NULL, default colours
are used.}
}
\value{
Returns a vector of annotation colours.
}
\description{
Fetches annotation colours for a SpatialFeatureExperiment object.
}
\details{
This function retrieves annotation colours based on the provided
colours vector or generates default colours if none are provided.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
