% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{.int_applyPseudocount}
\alias{.int_applyPseudocount}
\title{Internal Function: Apply Pseudocount}
\usage{
.int_applyPseudocount(values, pseudocount)
}
\arguments{
\item{values}{A numeric vector to which the pseudocount will be added.}

\item{pseudocount}{A numeric value representing the pseudocount to be added
(default is \code{1e-12}).}
}
\value{
A numeric vector with the pseudocount added.
}
\description{
This function adds a pseudocount to a numeric vector to prevent zero values.
}
\details{
The pseudocount is used to avoid zero values that can lead to errors in
subsequent calculations, such as division by zero or logarithm of zero.
}
\seealso{
\code{\link[stats]{log}} for logarithmic operations.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
\keyword{pseudocount}
