% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwcSTE.R
\name{.int_arrangeByType}
\alias{.int_arrangeByType}
\title{Arrange cell type Markers by Type}
\usage{
.int_arrangeByType(markers)
}
\arguments{
\item{markers}{A data frame containing information about cell type
markers, including gene names and types.}
}
\value{
A data frame with markers arranged by their types in ascending order.
}
\description{
This function arranges cell type markers by their types in
ascending order.
}
\details{
The function uses the \code{dplyr::arrange} function to sort the markers
based on the cell types in ascending order, resulting in a data frame where
markers are organised by their types for improved readability.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{arrangement,}
\keyword{cell}
\keyword{markers,}
\keyword{type,}
\keyword{types}
