% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{.int_calculate_nmf}
\alias{.int_calculate_nmf}
\title{Perform NMF on cell-level data}
\usage{
.int_calculate_nmf(
  x,
  ncomponents = 2,
  ntop = 500,
  subset_row = NULL,
  scale = FALSE,
  transposed = FALSE,
  seed = 1,
  ...
)
}
\arguments{
\item{x}{For \code{calculateNMF}, a numeric matrix of log-expression values where rows are features and columns are cells.
Alternatively, a \linkS4class{SummarizedExperiment} or \linkS4class{SingleCellExperiment} containing such a matrix.

For \code{runNMF}, a \linkS4class{SingleCellExperiment} object.}

\item{...}{For the \code{calculateNMF} generic, additional arguments to pass to specific methods.
For the ANY method, additional arguments to pass to \code{\link[Rtsne]{Rtsne}}.
For the SummarizedExperiment and SingleCellExperiment methods, additional arguments to pass to the ANY method.}
}
\value{
For \code{calculateNMF}, a numeric matrix is returned containing the NMF coordinates for each cell (row) and dimension (column).

For \code{runNMF}, a modified \code{x} is returned that contains the NMF coordinates in \code{\link{reducedDim}(x, name)}.

In both cases, the matrix will have the attribute \code{"basis"} containing the gene-by-factor basis matrix.
}
\description{
Perform non-negative matrix factorization (NMF) for the cells, based on the data in a SingleCellExperiment object.
}
\details{
The function \code{\link[RcppML]{nmf}} is used internally to compute the NMF.
Note that the algorithm is not deterministic, so different runs of the function will produce differing results.
Users are advised to test multiple random seeds, and then use \code{\link{set.seed}} to set a random seed for replicable results.
}
\seealso{
\code{\link[RcppML]{nmf}}, for the underlying calculations.

\code{\link{plotNMF}}, to quickly visualize the results.
}
\author{
Aaron Lun
}
