% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{.int_calculateKLdivergence}
\alias{.int_calculateKLdivergence}
\title{Internal Function: Calculate KL Divergence}
\usage{
.int_calculateKLdivergence(x, y, pseudocount = 1e-12)
}
\arguments{
\item{x}{A numeric vector representing the first probability distribution.}

\item{y}{A numeric vector representing the second probability distribution.}

\item{pseudocount}{A numeric value representing the pseudocount to be added
to both distributions (default is \code{1e-12}).}
}
\value{
A numeric value representing the KL divergence between the two
distributions.
}
\description{
This function calculates the Kullback-Leibler (KL) divergence between two
numeric vectors.
}
\details{
The function calculates the KL divergence between two probability
distributions \code{x} and \code{y}. A pseudocount is added to avoid
division by zero or logarithm of zero.
}
\seealso{
\code{\link[stats]{log}} for logarithmic operations.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
\keyword{kl-divergence}
