% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{.int_calculateMKL}
\alias{.int_calculateMKL}
\title{Internal Function: Calculate Mean Kullback-Leibler Divergence (MKL)}
\usage{
.int_calculateMKL(A.hat, A)
}
\arguments{
\item{A.hat}{A numeric matrix representing the estimated values.}

\item{A}{A numeric matrix representing the true values.}
}
\value{
A numeric value representing the MKL divergence between the two matrices.
}
\description{
This function calculates the Mean Kullback-Leibler (MKL) divergence between
two matrices.
}
\details{
The function computes the MKL divergence between the estimated matrix
(\code{A.hat}) and the true matrix (\code{A}) by taking the mean KL
divergence over all rows.
}
\seealso{
\code{\link[stats]{mean}} for mean calculations.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
\keyword{kl-divergence}
\keyword{mkl}
