% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{.int_calculateMSE}
\alias{.int_calculateMSE}
\title{Internal Function: Calculate Mean Squared Error (MSE)}
\usage{
.int_calculateMSE(A.hat, A)
}
\arguments{
\item{A.hat}{A numeric matrix representing the estimated values.}

\item{A}{A numeric matrix representing the true values.}
}
\value{
A numeric value representing the MSE between the two matrices.
}
\description{
This function calculates the Mean Squared Error (MSE) between two matrices.
}
\details{
The function computes the MSE between the estimated matrix (\code{A.hat})
and the true matrix (\code{A}).
}
\seealso{
\code{\link[stats]{mean}} for mean calculations.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
\keyword{mse}
