% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_spotHex.R
\name{dot-int_char_to_numeric}
\alias{dot-int_char_to_numeric}
\alias{.int_char_to_numeric}
\title{Internal: Convert a Character to a Numeric Value}
\usage{
.int_char_to_numeric(x)
}
\arguments{
\item{x}{A character value to be converted to numeric.}
}
\value{
If the conversion is successful, a numeric value is returned. If
the input cannot be converted, NA (Not Available) is returned.
}
\description{
IMPORTANT: This function takes a character input and attempts to convert it
into a numeric value. If the conversion is successful, it returns the numeric
value; otherwise, it returns NA.
}
\examples{
char_to_numeric("123")  # Returns: 123
char_to_numeric("abc")  # Returns: NA

}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
