% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metaSFE.R
\name{.int_checkAndOutput}
\alias{.int_checkAndOutput}
\title{Internal: Check and Output MSFE or SFE Object}
\usage{
.int_checkAndOutput(m_sfe, sfe, sample_id)
}
\arguments{
\item{m_sfe}{An object of class SpatialFeatureExperiment or
MetaSpatialFeatureExperiment.}

\item{sfe}{An object of class SpatialFeatureExperiment.}

\item{sample_id}{A character vector specifying the sample ID for the sample
we used.}
}
\value{
An object of class SpatialFeatureExperiment. If the input is already
an SFE, it returns the original object. If the input is a MetaSFE, it
returns a subset based on the provided sample IDs.
}
\description{
An internal function to check the class type of the input object
(\code{m_sfe}) and return either the original SpatialFeatureExperiment (SFE)
object or, in the case of a MetaSpatialFeatureExperiment (MetaSFE), add a
subset based on the provided sample IDs (\code{sample_id}).
}
\seealso{
\code{\link{addSFE}}
}
