% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_STE.R
\name{.int_checkDMat}
\alias{.int_checkDMat}
\title{Internal: select and obtain the distance matrix}
\usage{
.int_checkDMat(dMat, sfe)
}
\arguments{
\item{dMat}{A matrix representing the distance matrix, or a string
indicating the type of distance matrix ('euclidean', 'manhattan', or
'minkowski').}

\item{sfe}{The spatial feature experiment object from which to obtain the
distance matrix if not provided explicitly.}
}
\value{
A distance matrix.
}
\description{
This function selects and obtains the distance matrix (\code{dMat}) to be
used. It first checks if dMat is provided, then tries to obtain it from the
metadata of a spatial feature object (\code{sfe}), and finally attempts to
obtain it based on the specified type. If all attempts fail, it raises an
error asking the user to add a distance matrix using the \code{addDistMat}
function.
}
\keyword{internal}
