% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{.int_checkFGWCClass}
\alias{.int_checkFGWCClass}
\title{Internal: Check if an Object is of Class 'fgwc'}
\usage{
.int_checkFGWCClass(object)
}
\arguments{
\item{object}{The object to check.}
}
\value{
A logical value indicating whether the object is of class 'fgwc'.
}
\description{
This function checks if a given object is of class 'fgwc'.
}
\author{
Eleftherios (Lefteris) Zormpas
}
