% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{.int_checkMappings}
\alias{.int_checkMappings}
\title{Internal: Check the validity of mapping inputs for pie and donut plots}
\usage{
.int_checkMappings(mapping)
}
\arguments{
\item{mapping}{A ggplot2 \code{aes} function output containing the mapping
information for the pie and donut plots. The names should be 'pie' and
'donut', and the values should be the column names from the colData.}
}
\value{
If the mappings are valid, the function returns \code{NULL} invisibly.
If the mappings are invalid, the function throws an error with a
descriptive message.
}
\description{
This function checks if the provided mapping inputs for the pie and donut
plots are valid. It ensures that the mapping names are among the allowed
options, which are 'annotation', 'cluster', and 'factors'.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
