% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_plot.R
\name{.int_checkSACompatible}
\alias{.int_checkSACompatible}
\title{Internal Function: Check compatibility between selected statistic and
statistical significance method}
\usage{
.int_checkSACompatible(localRes, pVal)
}
\arguments{
\item{localRes}{A character string with the \code{localResult} name.}

\item{pVal}{Numeric. The p-value.}
}
\value{
This function returns a value for the \code{pVal} argument.
}
\description{
This internal function checks the compatibility between selected statistic
and statistical significance method. Geary's C in particular returns no
Z-score statistic. As a result, when someone is plotting Geary's C the
\code{pVal} argument is not used. A message will be printed to
inform for this and also to suggest using \code{permutation} instead.
}
\details{
This function is intended for internal use to validate the compatibility
of the input parameters for plotting local spatial autocorrelation
statistics. It checks if \code{geary} is provided with \code{z-score}, if yes, then
sets the \code{pVal} argument to \code{NULL}. Otherwise returns the \code{pVal} argument as
is.
}
\author{
Eleftherios (Lefteris) Zormpas
}
