% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_geary.R
\name{.int_checkSAInput}
\alias{.int_checkSAInput}
\title{Internal Function: Check Spatial Autocorrelation Input}
\usage{
.int_checkSAInput(x, listw)
}
\arguments{
\item{x}{A numeric vector, the same length as the neighbours list in listw.}

\item{listw}{A \code{listw} object created, for example, by \code{nb2listw}.}
}
\value{
This function does not return any value. It stops and throws an error if the
input is invalid.
}
\description{
This internal function checks the validity of spatial autocorrelation input.
}
\details{
This function is intended for internal use to validate the input parameters
for spatial autocorrelation calculations. It checks if 'x' is numeric, if
its length matches the length of neighbors in 'listw', and if 'listw' is of
class 'listw'.
}
\author{
Eleftherios (Lefteris) Zormpas
}
