% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metaSFE.R
\name{.int_checkSampleNames}
\alias{.int_checkSampleNames}
\title{Internal: Check if sample names are correct}
\usage{
.int_checkSampleNames(x, sample_ids)
}
\arguments{
\item{x}{An instance of the MetaSpatialFeatureExperiment class.}

\item{sample_ids}{Character string. The sample IDs to be checked.}
}
\value{
Logical vector indicating whether each sample ID is correct.
}
\description{
An internal function to check that the sample names provided for extraction
are correct.
}
