% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_plot.R
\name{.int_checkVarNames}
\alias{.int_checkVarNames}
\title{Internal: Check Variable Names in SpatialFeatureExperiment}
\usage{
.int_checkVarNames(sfe, var)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} object.}

\item{var}{A character string representing the variable name to be checked.}
}
\value{
Returns the variable name if it is found in \code{colData}. If not, it
attempts to match and return the corresponding ENSG ID.
}
\description{
This internal function checks if a specified variable name exists in the
\code{colData} of a \code{SpatialFeatureExperiment} object. If the variable is not found
in \code{colData}, it attempts to match the variable to an ENSG (Ensembl Gene)
ID.
}
\author{
Eleftherios (LEFteris) Zormpas
}
\keyword{internal}
