% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_applyQCthresh.R
\name{.int_cleanMetaData}
\alias{.int_cleanMetaData}
\title{Internal Function: Clean Up Duplicate Entries in the Metadata Slot of a
SpatialFeatureExperiment}
\usage{
.int_cleanMetaData(sfeOut)
}
\arguments{
\item{sfeOut}{A \code{SpatialFeatureExperiment} object.}
}
\value{
A modified \code{SpatialFeatureExperiment} object with duplicate
entries removed from the metadata slot.
}
\description{
This internal function cleans up duplicate entries in the metadata slot of a
\code{SpatialFeatureExperiment} object.
}
\details{
This internal function ensures that the metadata slot of the
\code{SpatialFeatureExperiment} object does not contain duplicate entries.
It makes the names of the metadata unique and keeps the slots with the
original names.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
\keyword{metadata}
\keyword{spatial}
\keyword{transcriptomics}
