% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{.int_compositeScore}
\alias{.int_compositeScore}
\title{Internal: Determine the best number of clusters using multiple indexes}
\usage{
.int_compositeScore(index_data)
}
\arguments{
\item{index_data}{A data frame containing the results of multiple indexes.
Columns should include \code{k}, \code{FPC}, \code{CE}, \code{SC}, \code{XB}.}
}
\value{
The index_data dataframe with the CompositeScore column added.
}
\description{
Internal: Determine the best number of clusters using multiple indexes
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{composite}
\keyword{score}
