% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{.int_computeErrors}
\alias{.int_computeErrors}
\title{Internal Function: Compute Errors for NMF Factor Selection}
\usage{
.int_computeErrors(k, A, A.rand, n_cores, max.iter, loss, seed)
}
\arguments{
\item{k}{An integer specifying the number of factors to use for NMF.}

\item{A.rand}{A numeric matrix representing the randomized version of the
data (\code{A}).}

\item{n_cores}{An integer indicating the number of cores to use for parallel
processing.}

\item{max.iter}{An integer specifying the maximum number of iterations for
NMF.}

\item{loss}{Character, either \code{"mse"} (mean squared error) or
\code{"mkl"} (mean Kullback-Leibler divergence) (default is \code{"mse"}).}

\item{seed}{An integer specifying the random seed for reproducibility}

\item{.A}{A numeric matrix representing the data to be factorized.}
}
\value{
A numeric vector with two elements:
\describe{
\item{\code{err}}{The reconstruction error for the input matrix
(\code{A}).}
\item{\code{err.rand}}{The reconstruction error for the randomized matrix
(\code{A.rand}).}
}
}
\description{
This function computes the reconstruction errors for a given number of
factors (\code{k}) using both the input matrix (\code{A}) and a randomized
version (\code{A.rand}).
}
\details{
The function uses the \code{nnmf} function from the \code{NNLM} package to
perform Non-negative Matrix Factorization (NMF) and calculates the
reconstruction errors using either Mean Squared Error (MSE) or Mean
Kullback-Leibler (MKL) divergence.
}
\seealso{
\code{\link[NNLM]{nnmf}} for NMF calculations.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{error-calculation}
\keyword{internal}
\keyword{nmf}
