% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metaSFE.R
\name{.int_convertToDgCMatrix}
\alias{.int_convertToDgCMatrix}
\title{Internal: Convert assay matrix to dgCMatrix}
\usage{
.int_convertToDgCMatrix(sfe, nm)
}
\arguments{
\item{sfe}{An instance of the SummarizedExperiment class.}

\item{nm}{Name of the assay matrix.}
}
\value{
SummarizedExperiment with the specified assay matrix converted to
dgCMatrix.
}
\description{
An internal function to convert any assay matrix class to dgCMatrix class.
}
