% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_STE.R
\name{.int_countElements}
\alias{.int_countElements}
\title{Internal: Function to count elements in each string}
\usage{
.int_countElements(string)
}
\arguments{
\item{string}{A character string containing elements separated by "+".}
}
\value{
The number of elements in the string.
}
\description{
This internal function splits a character string by the "+" symbol and counts
the number of elements in the resulting list.
}
\author{
Eleftherios Zormpas
}
