% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setQCthresh.R
\name{.int_custom_1}
\alias{.int_custom_1}
\title{Apply Custom QC Metric to Rows in a SpatialFeatureExperiment object}
\usage{
.int_custom_1(sfe, qcMetric, qc_name)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment containing spatial
transcriptomics experiment data.}

\item{qcMetric}{A vector of TRUE and FALSE for rows (genes)
that are flagged for discarding.}

\item{qc_name}{The name of the custom QC metric to be added to the rowData.}
}
\value{
An updated SpatialFeatureExperiment object with the custom QC metric
added to the rowData.
}
\description{
This function adds a custom QC metric to rows (genes) in a
spatial transcriptomics experiment.
}
\details{
The function takes a vector \code{qcMetric} containing TRUE and FALSE
values, representing rows that should be flagged for discarding. It adds a
new column with the specified name \code{qc_name} in the rowData of the
SpatialFeatureExperiment object, marking rows based on the custom QC metric.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{QC}
\keyword{SpatialFeatureExperiment}
\keyword{custom}
\keyword{metric,}
