% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_setQCthresh.R
\name{.int_custom_2}
\alias{.int_custom_2}
\title{Apply Custom QC Metric to Columns in a Spatial Transcriptomics Experiment}
\usage{
.int_custom_2(sfe, qcMetric, qc_name)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object containing spatial
transcriptomics experiment data.}

\item{qcMetric}{A vector of TRUE and FALSE for columns (locations) that are
flagged for discarding.}

\item{qc_name}{The name of the custom QC metric to be added to the colData.}
}
\value{
An updated SpatialFeatureExperiment object with the custom QC metric
added to the colData.
}
\description{
This function adds a custom QC metric to columns in a spatial
transcriptomics experiment.
}
\details{
The function takes a vector \code{qcMetric} containing TRUE and FALSE
values, representing columns (spots) that should be flagged for discarding.
It adds a new column with the specified name \code{qc_name} in the colData of the
spatial transcriptomics experiment, marking spots based on the custom QC
metric.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{QC}
\keyword{SpatialFeatureExperiment,}
\keyword{custom}
\keyword{features}
\keyword{metric,}
\keyword{spatial}
