% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC_.R
\name{.int_dataToPlot}
\alias{.int_dataToPlot}
\title{Internal Function: .int_dataToPlot}
\usage{
.int_dataToPlot(sfe, ids, fill)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object.}

\item{ids}{Vector of sample IDs to include in the plot.}

\item{fill}{Variable used for filling the spots in the plot.}
}
\value{
Returns a data frame suitable for generating spatial plots.
}
\description{
Creates a data frame for generating spatial plots based on specified fill
variable.
}
\details{
This function creates a data frame with sample metadata and spatial
coordinates based on the specified fill variable.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
