% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{.int_determineCluster}
\alias{.int_determineCluster}
\title{Internal: Determine Highest Cluster Membership from Fuzzy Membership Matrix}
\usage{
.int_determineCluster(data, uij)
}
\arguments{
\item{data}{A numeric matrix or data frame where each row is a data point
and each column is a feature.}

\item{uij}{A matrix where each row corresponds to a data point and each
column corresponds to a cluster. The values are membership scores.}
}
\value{
A data frame or matrix similar to 'data' but with an additional
column 'cluster' that indicates the cluster to which each data point
has been assigned. The cluster index is based on the maximum membership
score for each point.
}
\description{
This function assigns each data point to a cluster based on the highest
membership value from the fuzzy membership matrix 'uij'.
}
\details{
This function is an adaptation of the \code{determine_cluster} function from the
\code{naspaclust} R package.
}
\author{
Eleftherios (Lefteris) Zormpas
}
