% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyQCthresh_.R
\name{.int_discardFeat}
\alias{.int_discardFeat}
\title{Internal Function: Discard Features in a SpatialFeatureExperiment Based on
Quality Control}
\usage{
.int_discardFeat(sfe)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} object.}
}
\value{
A modified \code{SpatialFeatureExperiment} object with discarded
features removed.
}
\description{
This internal function discards features in a \code{SpatialFeatureExperiment}
object based on the quality control information stored in the feature
metadata (rowData).
}
\details{
This internal function discards features in the
\code{SpatialFeatureExperiment} object where the quality control information
in the feature metadata, specifically the \code{qc_discard} column, is
\code{TRUE}. Features marked for discard are removed from the object.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{control}
\keyword{internal}
\keyword{quality}
\keyword{spatial}
\keyword{transcriptomics}
