% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{.int_EnsgIDtoName}
\alias{.int_EnsgIDtoName}
\title{Internal Function: match EnsgIDs to gene names without the use of a biomart}
\usage{
.int_EnsgIDtoName(sfe, genes)
}
\arguments{
\item{sfe}{An SFE object.}

\item{genes}{A vector of EnsgIDs.}
}
\value{
a vector of Gene Names
}
\description{
This function uses the SFE object to translate a vector of ENSGIDs into gene
names.
}
\author{
Eleftherios (Lefteris) Zormpas
}
