% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setQCthresh.R
\name{.int_featToDiscard}
\alias{.int_featToDiscard}
\title{Feature Grouping: Mark Features to Discard}
\usage{
.int_featToDiscard(sfe, filters)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment containing spatial
transcriptomics experiment data.}

\item{filters}{TRUE selects all calculated QC metrics. Alternatively,
provide a character vector with the column names of the QC metrics you want
to consider for filtering. The QC metric columns begin with "qc_".}
}
\value{
An updated SpatialFeatureExperiment experiment object with a new
column "qc_discard" in the rowData, indicating whether each feature should
be discarded based on the selected QC metrics.
}
\description{
This function marks features (genes) in a spatial transcriptomics experiment
to be discarded based on specified QC metrics.
}
\details{
The function uses the internal function \code{.int_selectFilters} to
obtain the selected QC metrics. It then applies a logical operation to mark
features to be discarded based on the presence of TRUE values in at least
one of the selected QC metrics. The result is added as a new column
"qc_discard" in the rowData of the SpatialFeatureExperiment object.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{QC}
\keyword{SpatialFeatureExperiment}
\keyword{discard,}
\keyword{feature}
\keyword{grouping,}
\keyword{metrics,}
