% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_plot.R
\name{.int_fetchAnnotColours}
\alias{.int_fetchAnnotColours}
\title{Internal: Get Colours}
\usage{
.int_fetchAnnotColours(annot)
}
\arguments{
\item{annot}{annotations vector}
}
\value{
a named vector of colours with annotation labels as names.
}
\description{
This function returns colours for the annotations. It is intended to work in
an \code{lapply} setup.
}
\author{
Eleftherios (Lefteris) Zormpas
}
