% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwcSTE.R
\name{.int_fgwcMarkerCounts}
\alias{.int_fgwcMarkerCounts}
\title{Internal Function: Retrieve Normalised Counts for gene Markers}
\usage{
.int_fgwcMarkerCounts(sfe, markers)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object containing normalised counts.}

\item{markers}{A data frame containing information about FGWC markers,
including gene names and corresponding ENSG.IDs.}
}
\value{
A data frame with normalised counts for the specified FGWC markers.
}
\description{
This function retrieves normalised counts for a specified set of gene
markers from a SpatialFeatureExperiment object.
}
\details{
The function extracts normalised counts for a set of FGWC markers
from the provided SpatialFeatureExperiment object. It filters the counts
based on the ENSG.IDs of the specified markers, rearranges columns for
better readability, and provides a data frame with gene names as row names
and observations as variables.
}
\seealso{
\code{\link[colData]{colData}}, \code{\link{as.data.frame}}, \code{\link{dplyr::filter}}, \code{\link{dplyr::left_join}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{FGWC,}
\keyword{SpatialFeatureExperiment}
\keyword{counts,}
\keyword{markers,}
\keyword{normalised}
