% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{.int_fgwcPlotDataViolin}
\alias{.int_fgwcPlotDataViolin}
\title{Internal Function: Generate Data for FGWC Violin Plot}
\usage{
.int_fgwcPlotDataViolin(fgwc, sfe)
}
\arguments{
\item{fgwc}{An object containing FGWC results, typically obtained from the
\code{fgwcSTE} function.}

\item{sfe}{A SpatialFeatureExperiment object containing spatial coordinates
and features.}
}
\value{
A data frame containing information for generating FGWC plots. In
single-cluster mode, the data frame includes cluster information for each
spot. In multi-cluster mode, the entire data frame is returned.
}
\description{
This function generates data suitable for plotting FFGWC results as violins.
}
\details{
The function generates data suitable for plotting FGWC results as
violin plots.
}
\seealso{
\code{\link[SpatialFeatureExperiment]{colGeometry}},
\code{\link[SpatialFeatureExperiment]{colData}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{FGWC,}
\keyword{clustering,}
\keyword{coordinates}
\keyword{spatial}
