% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{.int_fillArgs}
\alias{.int_fillArgs}
\title{Internal Function: .int_fillArgs}
\usage{
.int_fillArgs(fill_args, annot_cols)
}
\arguments{
\item{fill_args}{List of fill arguments.}

\item{annot_cols}{Vector of annotation colours.}
}
\value{
Returns a list of fill arguments with defaults if not provided.
}
\description{
Sets default fill arguments if not provided.
}
\details{
This function sets default fill arguments, such as values and
na.value, if the input list is empty.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
