% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_utils.R
\name{.int_filterGenes}
\alias{.int_filterGenes}
\title{Internal Function: filter genes based on Global SA statistic}
\usage{
.int_filterGenes(data, colStat, colPVal, pVal, stat_thresh)
}
\arguments{
\item{data}{the SFE rowData.}

\item{colStat}{the SA statistic's values column name.}

\item{colPVal}{the SA statistic's p-values column name.}

\item{pVal}{the SA statistic's p-value threshold. MUST be a numeric.}

\item{stat_thresh}{the SA statistic threshold. MUST be a numeric.}
}
\description{
This function uses p-value and/or SA statistic values as thresholds to
filter and select genes. Returns a data frame of selected genes from the
SFE object's rowData.
}
\author{
Eleftherios (Lefteris) Zormpas
}
