% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_filterInTissue.R
\name{.int_filterInTissue}
\alias{.int_filterInTissue}
\title{Internal Function: Filter In-Tissue Spots for a Specific Sample}
\usage{
.int_filterInTissue(id, sfe)
}
\arguments{
\item{id}{A character string specifying the sample/image identifier.}

\item{sfe}{A \code{SpatialFeatureExperiment} object.}
}
\value{
A \code{SpatialFeatureExperiment} object with in-tissue spots
filtered for the specified sample.
}
\description{
This internal function filters in-tissue spots for a specific sample in
spatial transcriptomics data. It takes a character string (\code{id})
specifying the sample/image identifier and a \code{SpatialFeatureExperiment}
object (\code{sfe}). The logic is implemented using the .int_filterInTissue
function, and the resulting \code{SpatialFeatureExperiment} object contains
only in-tissue spots for the specified sample.
}
\details{
This internal function efficiently filters in-tissue spots for a specific
sample in spatial transcriptomics data. It takes a character string
(\code{id}) specifying the sample/image identifier and a
\code{SpatialFeatureExperiment} object (\code{sfe}). The logic is based on
the .int_filterInTissue function, and the resulting
\code{SpatialFeatureExperiment} object contains only in-tissue spots for the
specified sample.
}
\keyword{function,}
\keyword{in-tissue}
\keyword{internal}
\keyword{spatial}
\keyword{spots}
\keyword{transcriptomics,}
