% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyQCthresh_.R
\name{.int_filterQCDiscard}
\alias{.int_filterQCDiscard}
\title{Internal Function: Filter QC Discard Status for a Specific Sample in
SpatialFeatureExperiment}
\usage{
.int_filterQCDiscard(id, sfe)
}
\arguments{
\item{id}{A character string specifying the sample identifier.}

\item{sfe}{A \code{SpatialFeatureExperiment} object.}
}
\value{
A new \code{SpatialFeatureExperiment} object with QC-discarded
features removed for the specified sample.
}
\description{
This internal function filters out features marked for quality control (QC)
discard for a specific sample in a \code{SpatialFeatureExperiment} object.
}
\details{
This internal function filters out features that have been marked
for quality control (QC) discard for a specific sample. It returns a
modified \code{SpatialFeatureExperiment} object with the QC-discarded
features removed.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
\keyword{quality-control}
\keyword{spatial}
\keyword{transcriptomics}
