% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{.int_findElbowPointAngle}
\alias{.int_findElbowPointAngle}
\title{Internal Function: Find Elbow Point Using Angles}
\usage{
.int_findElbowPointAngle(index_data)
}
\arguments{
\item{index_data}{A data frame containing columns \code{k} (number of clusters)
and \code{IndexValue} (metric value).}
}
\value{
An integer representing the optimal number of clusters (\code{k}).
}
\description{
This function identifies the elbow point using the angle method.
}
\details{
This function takes the \code{index_data} data frame, created internally by a
function like \code{fgwc_findOptimumK}, sorts it by \code{k} (just to ensure it's
in the correct order), and then calculates the angle between each pair of
adjacent points. The point with the largest angle is considered the elbow
point. The function returns the \code{k} value corresponding to this elbow point.

The logic behind finding the maximum angle is that the "elbow" in a curve
is where there is a significant change in the slope, typically from steep
to shallow. This point often represents a balance between the complexity
and variance explained, making it a useful heuristic in cluster analysis
or other optimisation tasks.
}
\seealso{
\code{\link{fgwc_findOptimumK}} for finding the optimal number of
clusters.
}
\keyword{elbow-point}
\keyword{internal}
