% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.int_findFileWithRegex}
\alias{.int_findFileWithRegex}
\title{Internal: Find a file in a directory using regex}
\usage{
.int_findFileWithRegex(directory, pattern)
}
\arguments{
\item{directory}{character string. The directory.}

\item{pattern}{character string. The regex pattern.}
}
\description{
This function returns the pathway for a file using regex
}
\author{
Eleftherios (Lefteris) Zormpas
}
