% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{.int_fpcoeff}
\alias{.int_fpcoeff}
\title{Internal: Calculate the Fuzzy Partition Coefficient (FPC)}
\usage{
.int_fpcoeff(u)
}
\arguments{
\item{u}{The FGWC membership matrix. A 2D matrix (N, C), where C is the
number of clusters and N is the number of data points (locations). Each
element represents the membership degree of a data point to a particular
cluster.}
}
\value{
A number representing the Fuzzy Partition Coefficient.
}
\description{
Computes the Fuzzy Partition Coefficient (FPC) relative to a fuzzy
c-partitioned matrix \code{u} (essentially the FGWC membership matrix). Measures
'fuzziness' in partitioned clustering.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{coefficient}
\keyword{partition}
