% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_gearyC.R
\name{.int_geary}
\alias{.int_geary}
\title{Internal Function: Calculate Geary's C}
\usage{
.int_geary(gene, sfe, assay, listw, n, n1, S0, zero.policy)
}
\description{
This internal function calculates the C statistic for a vector of gene
expression. For the parameter arguments check the
\code{link[STExplorer]{gearyGlobalC}} function.
}
\author{
Eleftherios (Lefteris) Zormpas
}
