% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_gearyC.R
\name{.int_gearyCPerm}
\alias{.int_gearyCPerm}
\title{Internal Function: Calculate Geary's C with Permutation testing}
\usage{
.int_gearyCPerm(
  gene,
  sfe,
  assay,
  listw,
  nsim,
  alternative,
  spChk,
  adjust.n,
  return_boot,
  zero.policy
)
}
\description{
This internal function calculates the C statistic for a vector of gene
expression with Monte-Carlo permutation testing of significance. For the
parameter arguments check the \code{link[STExplorer]{gearyGlobalCPerm}}
function.
}
\author{
Eleftherios (Lefteris) Zormpas
}
