% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_gearyC.R
\name{.int_gearyLocalC}
\alias{.int_gearyLocalC}
\title{Internal Function: Calculate local Geary's C}
\usage{
.int_gearyLocalC(gene, sfe, listw, zero.policy, ...)
}
\description{
This internal function calculates the local Ci statistic for a vector of
gene expression. For the parameter arguments check the
\code{link[STExplorer]{gearyLocalC}} function.
}
\author{
Eleftherios (Lefteris) Zormpas
}
