% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_gearyC.R
\name{.int_gearyLocalCPerm}
\alias{.int_gearyLocalCPerm}
\title{Internal Function: Calculate local Geary's C with statistical testing}
\usage{
.int_gearyLocalCPerm(
  gene,
  sfe,
  assay,
  listw,
  nsim,
  alternative,
  ...,
  zero.policy,
  iseed,
  no_repeat_in_row
)
}
\description{
This internal function calculates the local Ci statistic for a vector of
gene expression alongside a statistical significance testing. For the
parameter arguments check the \code{link[STExplorer]{gearyLocalCPerm}}
function.
}
\author{
Eleftherios (Lefteris) Zormpas
}
