% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_plot.R
\name{.int_generatePlot}
\alias{.int_generatePlot}
\title{Internal Function: Generate Spatial Plot for a Principal Component}
\usage{
.int_generatePlot(pc.No, cutoff, lead.item, fill, legend.position, size)
}
\arguments{
\item{pc.No}{The principal component number to plot.}

\item{cutoff}{The cutoff threshold for displaying legend labels.}

\item{lead.item}{The lead item containing spatial information.}

\item{fill}{The label for the fill aesthetic in the plot.}

\item{legend.position}{The position of the legend in the plot.}

\item{size}{The size of legend text in the plot.}
}
\value{
A ggplot2 object representing the spatial plot for the specified
principal component.
}
\description{
This function generates a spatial plot for a specified principal component
(PC) using the given parameters.
}
\details{
The function generates a spatial plot using the specified principal
component, lead item, and other parameters. It utilises ggplot2 for creating
the plot and allows customization of legend labels, colours, and other
visual elements.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{component,}
\keyword{ggplot2}
\keyword{plot,}
\keyword{principal}
\keyword{spatial}
