% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_FunctionalClust.R
\name{.int_generateGeneList}
\alias{.int_generateGeneList}
\title{Internal Function: Generate Gene List Based on Score Type}
\usage{
.int_generateGeneList(inputGSEA, X, scoreType)
}
\arguments{
\item{inputGSEA}{The input gene expression data for the specified spatial
location.}

\item{X}{The index of the spatial location for which the gene list is
generated.}

\item{scoreType}{The type of score used for sorting genes. Valid options
are "std" for standard GSEA and "pos" for positive enrichment.}
}
\value{
A vector containing the sorted gene list.
}
\description{
This function generates a gene list based on the specified score type from
the input gene expression data for a single spatial location. It is an
internal function used within the main analysis function and is not intended
for direct use.
}
\details{
This function internally sorts genes in descending order based on the
specified score type ("std" or "pos"). For "std", genes are sorted based on
their original loading scores, while for "pos", genes are sorted based on
the absolute value of their expression values.
}
\examples{
# This function is an internal function used within the main function.
# Refer to the main analysis function for example usage.

}
\seealso{
\code{\link{.int_sortGenes}}

Other Spatial Transcriptomics Analysis: 
\code{\link{.int_runGSEA}()},
\code{\link{gwpca_FunctionalClustering}()},
\code{\link{plotGWPCA_FuncCLust}()}
}
\concept{Spatial Transcriptomics Analysis}
\keyword{data}
\keyword{expression}
\keyword{gene}
\keyword{internal}
\keyword{list}
\keyword{score}
\keyword{sorting}
\keyword{type}
