% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{.int_generateInitialCentroids}
\alias{.int_generateInitialCentroids}
\title{Internal: Generate Initial Centroids for Clustering}
\usage{
.int_generateInitialCentroids(data, ncluster, gendist, randomN)
}
\arguments{
\item{data}{A numeric matrix or data frame where rows are observations
and columns are variables (features).}

\item{ncluster}{An integer specifying the number of clusters (centroids) to
generate.}

\item{gendist}{A character string specifying the type of distribution to use
for initialising the centroids. This can be "normal" for a normal
distribution or "uniform" for a uniform distribution.}

\item{randomN}{An integer used to set the seed for random number generation
to ensure reproducibility.}
}
\value{
A numeric matrix where each row represents a centroid and each column
corresponds to a feature in the input data. The dimensions of this matrix
are \code{ncluster} by \code{ncol(data)}.
}
\description{
This function generates initial centroids for clustering based on the
specified distribution. It supports both normal and uniform distributions to
initialise the centroid positions. The centroids are generated for each
feature (column) in the data matrix.
}
\details{
This function is an adaptation of the \code{uij} function from the
\code{naspaclust} R package.
}
\author{
Eleftherios (Lefteris) Zormpas
}
