% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_setQCthresh.R
\name{.int_get_qcMetric_for_sample}
\alias{.int_get_qcMetric_for_sample}
\title{Internal: Get the qcMetric for the sample}
\usage{
.int_get_qcMetric_for_sample(qcMetric, id = NULL)
}
\description{
Internal: Get the qcMetric for the sample
}
\author{
Eleftherios (Lefteris) Zormpas
}
