% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{.int_getAnnotsAndColoursFGWC}
\alias{.int_getAnnotsAndColoursFGWC}
\title{Internal: Get Annotations and Colours}
\usage{
.int_getAnnotsAndColoursFGWC(fgwc, loc_annot, order_rows)
}
\arguments{
\item{fgwc}{gwc class object as generated by \code{fgwcSTE}}

\item{order_rows}{order rows based on annotation or cluster}

\item{loc-annot}{which annotations to add}
}
\value{
a list
}
\description{
This function returns annotations label and colours for the heatmap
}
\author{
Eleftherios (Lefteris) Zormpas
}
