% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{.int_getColSubset}
\alias{.int_getColSubset}
\title{Get a Subset of Colours from a Palette}
\usage{
.int_getColSubset(no1, no2)
}
\arguments{
\item{no1}{The starting index for the subset.}

\item{no2}{The ending index for the subset.}
}
\value{
A vector of colours representing the subset from the palette.
}
\description{
This function retrieves a subset of colours from a palette based on
specified indices.
}
\details{
The function uses the \code{getColours} function to obtain a palette of
colours. It then extracts a subset of colours starting from the specified
index \code{no2} up to the end of the palette.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{colours,}
\keyword{palette}
\keyword{subset,}
