% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_utils.R
\name{.int_getColumnNames}
\alias{.int_getColumnNames}
\title{Internal Function: Construct and export column names for Global SA results}
\usage{
.int_getColumnNames(statistic, test)
}
\arguments{
\item{statistic}{A character string specifying the type of spatial
autocorrelation results ("moran", "geary", "getis") to look
for in the m_sfe object for the specified sample.}

\item{test}{A character string specifying the statistical test method
("z-score" or "permutation") to look for in the m_sfe object
for the specified sample.}
}
\description{
This function aims to construct column names based on the statistic and test.
}
\author{
Eleftherios (Lefteris) Zormpas
}
