% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.int_getDistMat}
\alias{.int_getDistMat}
\title{Internal Function: Get Distance Matrix from SpatialFeatureExperiment}
\usage{
.int_getDistMat(sfe, dMetric)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object containing a distance matrix in
its metadata.}

\item{dMetric}{The name of the distance metric used for the stored matrix
(Euclidean, Manhattan, or Minkowski).}
}
\value{
The distance matrix retrieved from the metadata of the
SpatialFeatureExperiment object.
}
\description{
This function retrieves the distance matrix stored in the metadata of a
SpatialFeatureExperiment object.
}
\details{
The function retrieves the distance matrix stored in the metadata
of the SpatialFeatureExperiment object. Users need to specify the distance
metric (Euclidean, Manhattan, or Minkowski) used for the stored matrix.
}
\seealso{
\code{\link[SpatialFeatureExperiment]{colData}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{SpatialFeatureExperiment}
\keyword{distance}
\keyword{matrix,}
\keyword{metadata,}
