% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{.int_getFactorData}
\alias{.int_getFactorData}
\title{Internal: Extract Factor Data}
\usage{
.int_getFactorData(fgwc, subset_row)
}
\arguments{
\item{fgwc}{fgwc class object as generated by \code{fgwcSTE}}

\item{subset_row}{subset rows (locations) using a boolean vector of length
equal to the number of locations.}
}
\value{
a data frame of NMF factors
}
\description{
This function extracts factor data from FGWC output
}
\author{
Eleftherios (Lefteris) Zormpas
}
