% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_plot.R
\name{.int_getFillLabel}
\alias{.int_getFillLabel}
\title{Internal Function: Get Fill Label for Spatial Autocorrelation Plot}
\usage{
.int_getFillLabel(statistic, plot = c("localStat", "cluster"))
}
\arguments{
\item{statistic}{A character string specifying the type of spatial
autocorrelation used ("moran", "geary", "getis").}

\item{plot}{A character string specifying the type of plot ("localStat" or
"cluster").}
}
\value{
A character string representing the fill label for the plot.
}
\description{
Internal Function: Get Fill Label for Spatial Autocorrelation Plot
}
\author{
Eleftherios (Lefteris) Zormpas
}
