% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_plot.R
\name{.int_getGeneHeatmapData}
\alias{.int_getGeneHeatmapData}
\title{Internal: Extract Membership Data}
\usage{
.int_getGeneHeatmapData(sfe, assay, subset_row, subset_col)
}
\arguments{
\item{subset_row}{subset rows (genes) using a boolean vector of length
equal to the number of genes.}

\item{subset_col}{subset cols (locations) using a boolean vector of length
equal to the number of locations.}

\item{fgwc}{fgwc class object as generated by \code{fgwcSTE}}
}
\value{
a data frame of cluster membership \%
}
\description{
This function extracts membership data from FGWC output
}
\author{
Eleftherios (Lefteris) Zormpas
}
