% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.int_getGradient}
\alias{.int_getGradient}
\title{Internal: Generate a Gradient of Colours for a Single Colour}
\usage{
.int_getGradient(colour, steps)
}
\arguments{
\item{colour}{A character string representing the colour to use for the gradient.}

\item{steps}{An integer specifying the number of steps in the gradient.}
}
\value{
A vector of colour values representing the gradient.
}
\description{
This function creates a gradient of colours from a single colour to white,
with a specified number of steps.
}
\keyword{internal}
