% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_STE.R
\name{.int_getGWRdata}
\alias{.int_getGWRdata}
\title{Internal: Prepare Data for Geographically Weighted Regression (GWR)}
\usage{
.int_getGWRdata(sfe, formula, type, assay)
}
\arguments{
\item{sfe}{An object of class SpatialFeatureExperiment.}

\item{formula}{The formula for the regression.}

\item{type}{A character string specifying the type of geometry to use for GWR.
Options are "spot" (spot geometry), "hex" (hexagon geometry), or "centroid"
(centroid geometry).}

\item{assay}{the counts assay to use}
}
\value{
A SpatialPointsDataFrame or SpatialPolygonsDataFrame object
containing the prepared data for GWR.
}
\description{
An internal function to prepare data for Geographically Weighted Regression
(GWR) by retrieving geometries and converting data to the appropriate format.
}
\details{
This function prepares the data for Geographically Weighted Regression
(GWR) by retrieving the specified type of geometry from a SpatialFeatureExperiment
object and converting the data to the appropriate format. The available options
for the type parameter are "spot" (spot geometry), "hex" (hexagon geometry),
and "centroid" (centroid geometry).
}
\author{
Eleftherios Zormpas
}
