% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_STE.R
\name{.int_getGWRdataGene}
\alias{.int_getGWRdataGene}
\title{Internal: Retrieve Gene Data for GWR}
\usage{
.int_getGWRdataGene(sfe, variables, assay)
}
\arguments{
\item{sfe}{An object of class SpatialFeatureExperiment.}

\item{variables}{A character vector of gene identifiers (e.g., ENSG IDs) to
be included in the GWR.}

\item{assay}{A character string specifying the assay to use for gene
expression data.}
}
\value{
A data frame with barcodes as row names and the specified gene
expression values as columns.
}
\description{
An internal function to extract gene expression data for use in
Geographically Weighted Regression (GWR).
}
\details{
This function retrieves gene expression data from the specified
assay of a SpatialFeatureExperiment object, selecting only the genes
specified in the variables parameter. The data is transposed to have
barcodes as row names and gene expressions as columns, suitable for further
processing in GWR.
}
\author{
Eleftherios Zormpas
}
