% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_STE.R
\name{.int_getGWRdataLoc}
\alias{.int_getGWRdataLoc}
\title{Internal: Retrieve Location Data for GWR}
\usage{
.int_getGWRdataLoc(sfe, variables)
}
\arguments{
\item{sfe}{An object of class SpatialFeatureExperiment.}

\item{variables}{A character vector of column names from colData to be
included in the GWR.}
}
\value{
A data frame with barcodes as row names and the specified location
variables as columns.
}
\description{
An internal function to extract location-specific variables for use in
Geographically Weighted Regression (GWR).
}
\details{
This function extracts location-specific data from the colData of
a SpatialFeatureExperiment object. It selects only the variables specified
in the variables parameter. The resulting data frame is formatted with
barcodes as row names, ready for integration into GWR analysis.
}
\author{
Eleftherios Zormpas
}
