% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_STE.R
\name{.int_getGWRdataNonGeneAssay}
\alias{.int_getGWRdataNonGeneAssay}
\title{Internal: Retrieve Non-Gene Assay Data for GWR}
\usage{
.int_getGWRdataNonGeneAssay(sfe, variables, assay)
}
\arguments{
\item{sfe}{An object of class SpatialFeatureExperiment.}

\item{variables}{A character vector of non-gene assay identifiers to
be included in the GWR.}

\item{assay}{A character string specifying the assay to use for non-gene
data.}
}
\value{
A data frame with barcodes as row names and the specified non-gene
assay values as columns.
}
\description{
An internal function to extract non-gene assay data for use in
Geographically Weighted Regression (GWR).
}
\details{
This function retrieves non-gene data from the specified
assay of a SpatialFeatureExperiment object, selecting only the variables
specified in the variables parameter. The data is transposed to have
barcodes as row names and non-gene assay values as columns, suitable for further
processing in GWR.
}
