% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_plot.R
\name{.int_getGWRdataToPlot}
\alias{.int_getGWRdataToPlot}
\title{Internal Function: Prepare GWR Data for Plotting}
\usage{
.int_getGWRdataToPlot(gwr, Var)
}
\arguments{
\item{gwr}{A GWR object containing the fitted GWR model results.}

\item{Var}{A character string specifying the variable whose coefficients
are to be plotted.}
}
\value{
A Simple Features (sf) object containing the coefficients, their
significance, and the geometry for plotting.
}
\description{
This internal function prepares the data from a Geographically Weighted
Regression (GWR) model for plotting. It extracts and formats the coefficients
and their significance for visualization.
}
\details{
The function adds a column \code{signifCoef} to indicate the significance of
the coefficients based on the t-value threshold of 1.96. It also prepares
a \code{geom_signif} column to facilitate the plotting of significant areas.

This function is used internally by the \code{plotGWR} function to prepare
spatial data for plotting.
}
\seealso{
\code{\link{plotGWR}}, \code{\link{gwr_toSF}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
