% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{.int_getImgDt}
\alias{.int_getImgDt}
\title{Internal Function: .int_getImgDt}
\usage{
.int_getImgDt(sfe, sample_id = NULL, image_id = "lowres")
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object.}

\item{sample_id}{A character string, \code{TRUE}, or \code{NULL} specifying
sample/image identifier(s). If \code{TRUE}, all samples/images are
considered. If \code{NULL}, the first available entry is considered.}

\item{image_id}{Character vector specifying the image type ("lowres" by
default).}
}
\value{
Returns a list of raster objects based on image data.
}
\description{
Fetches image data and transforms it to raster for plotting.
}
\details{
This function fetches image data, transforms it to raster, and
organises it into a list with sample IDs as names.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
