% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{.int_getImgLims}
\alias{.int_getImgLims}
\title{Internal Function: .int_getImgLims}
\usage{
.int_getImgLims(sfe)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object.}
}
\value{
Returns a list of x and y limits for image plotting.
}
\description{
Extracts image limits for plotting based on spot locations.
}
\details{
This function extracts spot locations and sets x and y limits for image plotting.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
