% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC_.R
\name{.int_getImgLims2}
\alias{.int_getImgLims2}
\title{Internal Function: .int_getImgLims2}
\usage{
.int_getImgLims2(sfe, sample_id = NULL)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object.}

\item{sample_id}{A character string, \code{TRUE}, or \code{NULL} specifying
sample/image identifier(s). If \code{TRUE}, all samples/images are
considered. If \code{NULL}, the first available entry is considered.}
}
\value{
Returns a list of x and y limits for image plotting, organized by
sample IDs.
}
\description{
Extracts image limits for plotting based on spot locations for each sample.
}
\details{
This function extracts spot locations and sets x and y limits for
image plotting for each specified sample ID.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
