% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC_.R
\name{.int_getImgSmplIDs}
\alias{.int_getImgSmplIDs}
\title{Internal Function: .int_getImgSmplIDs}
\usage{
.int_getImgSmplIDs(sfe, sample_id = NULL)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object.}

\item{sample_id}{A character string, \code{TRUE}, or \code{NULL} specifying
sample/image identifier(s). If \code{TRUE}, all samples/images are
considered. If \code{NULL}, the first available entry is considered.}
}
\value{
Returns a character vector of sample IDs for image plotting.
}
\description{
Fetches sample IDs for image plotting.
}
\details{
This function fetches sample IDs from image data based on the input
parameters, providing flexibility for customising the sample IDs for image
plotting.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
