% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{.int_getIndexes}
\alias{.int_getIndexes}
\alias{.int_PartitionCoefficient}
\alias{.int_ClassificationEntropy}
\alias{.int_SilhouetteCoefficient}
\alias{.int_SeparationIndex}
\alias{.int_XieBeniIndex}
\alias{.int_IFV}
\alias{.int_Kwon}
\title{Internal: Calculate Validation indexes}
\usage{
.int_getIndexes(data, cluster, uij, vi, m, a = exp(1))

.int_PartitionCoefficient(uij)

.int_ClassificationEntropy(uij, a)

.int_SilhouetteCoefficient(data, cluster, uij, vi, m)

.int_SeparationIndex(data, uij, vi)

.int_XieBeniIndex(data, uij, vi, m)

.int_IFV(data, uij, vi, m)

.int_Kwon(data, uij, vi, m)
}
\arguments{
\item{data}{an object of data with d>1. Can be \code{matrix} or
\code{data.frame}. If your data is univariate, bind it with \code{1} to
get 2 columns.}

\item{uij}{membership matrix}

\item{vi}{centroid matrix}

\item{m}{degree of fuzziness or fuzzifier. Default is 2.}
}
\description{
A set of functions to calculate FGWC validation indexes
}
\details{
These functions are copied and/or adapted from the \code{naspaclust} R package.
}
\author{
Eleftherios (Lefteris) Zormpas
}
