% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_getisG.R
\name{.int_getisGTest}
\alias{.int_getisGTest}
\title{Internal Function: Calculate Getis & Ord's Global G with statistical
significance}
\usage{
.int_getisGTest(
  gene,
  sfe,
  listw,
  zero.policy,
  alternative,
  spChk,
  adjust.n,
  B1correct,
  adjust.x,
  Arc_all_x
)
}
\description{
This internal function calculates the G statistic and the statistical
significance for a vector of gene expression. For the parameter arguments
check the \code{link[STExplorer]{getisGlobalGTest}} function.
}
\author{
Eleftherios (Lefteris) Zormpas
}
