% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_getisG.R
\name{.int_getisLocal}
\alias{.int_getisLocal}
\alias{.int_getisLocalPerm}
\title{Internal Function: Calculate Getis & Ord's G}
\usage{
.int_getisLocal(
  gene,
  sfe,
  assay,
  listw,
  zero.policy,
  spChk,
  GeoDa,
  alternative,
  return_internals
)

.int_getisLocalPerm(
  gene,
  sfe,
  assay,
  listw,
  nsim,
  zero.policy,
  spChk,
  alternative,
  iseed,
  fix_i_in_Gstar_permutations,
  no_repeat_in_row
)
}
\description{
This internal function calculates the G statistic for a vector of gene
expression. For the parameter arguments check the
\code{link[STExplorer]{getisGlobalG}} function.

This internal function calculates the C statistic for a vector of gene
expression. For the parameter arguments check the
\code{link[STExplorer]{gearyGlobalC}} function.
}
\author{
Eleftherios (Lefteris) Zormpas
}
