% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{.int_getLabelsNo}
\alias{.int_getLabelsNo}
\title{Get the Number of Unique Labels in a Vector}
\usage{
.int_getLabelsNo(vector)
}
\arguments{
\item{vector}{A vector for which to calculate the number of unique labels.}
}
\value{
An integer representing the number of unique labels in the vector.
}
\description{
This function calculates the number of unique labels in a vector, excluding
missing values.
}
\details{
The function utilises the \code{unique} function to extract unique l
abels from the vector. It then counts the non-missing labels and returns the
total number.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{counting}
\keyword{labels,}
\keyword{unique}
\keyword{vector,}
