% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.int_getListSmplIDs}
\alias{.int_getListSmplIDs}
\title{INTERNAL: Get Sample IDs from a named List}
\usage{
.int_getListSmplIDs(list, sample_id)
}
\arguments{
\item{list}{A list object containing sample information.}

\item{sample_id}{Either a logical vector indicating the samples to include
(if TRUE, all samples are included), or a character vector specifying the
sample IDs to include. It is suggested to use a character vector to specify
a specific sample, or NULL to select the first available sample.}
}
\value{
A character vector containing the selected sample IDs.
}
\description{
This internal function extracts sample IDs from a list, based on the
provided sample ID specification.
}
\keyword{internal}
