% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{.int_getMapping}
\alias{.int_getMapping}
\title{Internal: Extract and Process Mapping Value}
\usage{
.int_getMapping(mapping, varname)
}
\arguments{
\item{mapping}{A named list or vector representing the mapping.}

\item{varname}{The name or index of the variable to extract from the
mapping.}
}
\value{
The processed value associated with the specified variable name. If
the processed value contains comma-separated elements, it is returned as a
vector of strings. Otherwise, it is returned as a single string.
}
\description{
This function extracts the value associated with a specified variable name
from a mapping and performs various string processing operations on the
extracted value.
}
\details{
The function performs the following steps:
\enumerate{
\item Extracts the value associated with the specified variable name from the
mapping.
\item If the version of the 'ggplot2' package is greater than "2.2.1", it
removes all occurrences of the tilde character (~) from the extracted
value.
\item Removes all occurrences of "c(", ")", and white space characters from the
extracted value using a regular expression.
\item If the processed value contains a comma, it splits the value into a
vector of strings using the comma as the separator.
\item Returns the processed value.
}
}
