% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{.int_getMembershipData}
\alias{.int_getMembershipData}
\title{Internal: Extract Membership Data}
\usage{
.int_getMembershipData(fgwc, subset_row)
}
\arguments{
\item{fgwc}{fgwc class object as generated by \code{fgwcSTE}}

\item{subset_row}{subset rows (locations) using a boolean vector of length
equal to the number of locations.}
}
\value{
a data frame of cluster membership \%
}
\description{
This function extracts membership data from FGWC output
}
\author{
Eleftherios (Lefteris) Zormpas
}
