% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_STE.R
\name{.int_getMetadataDMat}
\alias{.int_getMetadataDMat}
\title{Internal: Get the distance matrix from spatial feature object's metadata}
\usage{
.int_getMetadataDMat(dMat, sfe)
}
\arguments{
\item{dMat}{expected to be NULL}

\item{sfe}{The spatial feature experiment object.}
}
\value{
A distance matrix if found in metadata, otherwise \code{NULL}.
}
\description{
This function attempts to obtain the distance matrix from the metadata of a
spatial feature experiment object (\code{sfe}) when the dMat argument
provided in the exported function is left to NULL. If so, it returns the
first distance matrix; otherwise, it returns \code{NULL}.
}
\keyword{internal}
