% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{.int_getMetageneData}
\alias{.int_getMetageneData}
\title{Internal: Extract Metagene Data}
\usage{
.int_getMetageneData(fgwc)
}
\arguments{
\item{fgwc}{fgwc class object as generated by \code{fgwcSTE}}
}
\value{
a data frame of NMF factors
}
\description{
This function extracts metagene data from FGWC output
}
\author{
Eleftherios (Lefteris) Zormpas
}
