% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_plot.R
\name{.int_getNBdata}
\alias{.int_getNBdata}
\title{Internal Function: Extract Neighbour Graph Data}
\usage{
.int_getNBdata(sfe)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object.}
}
\value{
A Simple Features (sf) object representing the neighbour graph data.
}
\description{
This internal function extracts neighbour graph data from a
SpatialFeatureExperiment.
}
\details{
This function is an internal utility to extract neighbour graph data from a
SpatialFeatureExperiment. It utilizes the colGraph function to obtain the
neighbours and spatial coordinates, and then converts the data to a Simple
Features (sf) object.
}
\seealso{
\code{\link{colGraph}}, \code{\link{spatialCoords}}, \code{\link{nb2lines}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{graph}
\keyword{neighbour}
\keyword{sf}
\keyword{spatial}
