% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{.int_getPieDonutData}
\alias{.int_getPieDonutData}
\title{Internal: Prepare data for the pie-doughnut chart}
\usage{
.int_getPieDonutData(fgwc, sfe, pies, donuts)
}
\arguments{
\item{fgwc}{an object of class \code{fgwc}}

\item{sfe}{an object of class \code{SpatialFeatureExperiment}}

\item{pies}{pie mapping}

\item{donut}{donut mapping}
}
\value{
A data frame.
}
\description{
This function prepares data for the pie-doughnut chart
}
\author{
Eleftherios (Lefteris) Zormpas
}
