% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_STE.R
\name{.int_getProvidedDMat}
\alias{.int_getProvidedDMat}
\title{Internal: Get the provided distance matrix}
\usage{
.int_getProvidedDMat(dMat)
}
\arguments{
\item{dMat}{A matrix representing the distance matrix.}
}
\value{
A distance matrix if provided, otherwise \code{NULL}.
}
\description{
This function checks if a distance matrix (\code{dMat}) is provided. If
provided, it returns the distance matrix; otherwise, it returns \code{NULL}.
}
\keyword{internal}
