% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_plot.R
\name{.int_getSAClustColours}
\alias{.int_getSAClustColours}
\title{Internal Function: Get Colours for Spatial Autocorrelation Clusters}
\usage{
.int_getSAClustColours(clust_col, clusters)
}
\arguments{
\item{clust_col}{A named character vector of colours for plotting clusters.
The names must match cluster names.}

\item{clusters}{A character vector of cluster names.}
}
\value{
A character vector of colours for spatial autocorrelation clusters.
}
\description{
Internal Function: Get Colours for Spatial Autocorrelation Clusters
}
\author{
Eleftherios (Lefteris) Zormpas
}
