% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_utils.R
\name{.int_getSAdata}
\alias{.int_getSAdata}
\title{Internal: Retrieve SpatialFeatureExperiment Data for Spatial Autocorrelation}
\usage{
.int_getSAdata(sfe, var, assay)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} object.}

\item{var}{A character string representing the variable name to be retrieved.}

\item{assay}{A character string specifying the assay to be used for data
extraction.}
}
\value{
Returns a data vector corresponding to the specified variable, which
could be from gene expression assays or location-specific metadata.
}
\description{
This internal function retrieves data from a \code{SpatialFeatureExperiment}
object based on the type of variable specified. The function determines
whether the variable corresponds to Ensembl Gene IDs, location-specific
values, or other non-gene assay data.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
