% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_plot.R
\name{.int_getSAFeatSuffix}
\alias{.int_getSAFeatSuffix}
\title{Internal Function: Get Feature Suffix for Spatial Autocorrelation}
\usage{
.int_getSAFeatSuffix(feature, statistic)
}
\arguments{
\item{feature}{A character string indicating the feature of interest.}

\item{statistic}{A character string specifying the type of spatial
autocorrelation used ("moran", "geary", "getis").}
}
\value{
A character vector of feature suffixes for the specified statistic.
}
\description{
Internal Function: Get Feature Suffix for Spatial Autocorrelation
}
\author{
Eleftherios (Lefteris) Zormpas
}
