% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_plot.R
\name{.int_getSAStatColumns}
\alias{.int_getSAStatColumns}
\title{Internal Function: Select global SA stat columns to plot}
\usage{
.int_getSAStatColumns(colnames, statistic, ylab)
}
\arguments{
\item{colnames}{The rowData column names from an SFE object.}

\item{statistic}{A character vector specifying which spatial autocorrelation
statistics to plot. Valid options are "moran", "getis", and
"geary". Default is c("moran", "getis", "geary").}

\item{ylab}{A character string indicating the label for the y-axis. It can
be "gene_name" to use gene names or any other value to use
Ensemble Gene IDs. Default is "gene_name".}
}
\value{
A character string representing the name of the local results.
}
\description{
This function is fetching the required columns from the SFE's rowData.
}
\author{
Eleftherios (Lefteris) Zormpas
}
