% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{.int_getSubClusts}
\alias{.int_getSubClusts}
\title{Get Subclusters from Heatmap}
\usage{
.int_getSubClusts(heatmap, k)
}
\arguments{
\item{heatmap}{Data containing clustering information.}

\item{k}{Number of subclusters.}
}
\value{
A data frame with barcode and corresponding subcluster assignment.
}
\description{
Assigns subclusters to barcodes based on the input heatmap and the specified
number of clusters.
}
