% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_plot.R
\name{.int_getTitle}
\alias{.int_getTitle}
\title{Internal Function: Get Title for Spatial Autocorrelation Plot}
\usage{
.int_getTitle(sfe, feature, title)
}
\arguments{
\item{sfe}{An object of class 'SpatialFeatureExperiment'
containing spatial data.}

\item{feature}{A character string indicating the feature of interest.}

\item{title}{A character string specifying the title ("ENSGID" or
"name").}
}
\value{
A character string representing the title for the plot.
}
\description{
Internal Function: Get Title for Spatial Autocorrelation Plot
}
\author{
Eleftherios (Lefteris) Zormpas
}
