% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_STE.R
\name{.int_getTypedDMat}
\alias{.int_getTypedDMat}
\title{Internal: Get the distance matrix based on the specified type}
\usage{
.int_getTypedDMat(sfe, type)
}
\arguments{
\item{sfe}{The spatial feature experiment object.}

\item{type}{The type of distance matrix ('euclidean', 'manhattan', or
'minkowski').}
}
\value{
A distance matrix if found based on the specified type, otherwise
\code{NULL}.
}
\description{
This function attempts to obtain the distance matrix from the metadata of a
spatial feature object (\code{sfe}) based on the specified type. If found,
it returns the distance matrix; otherwise, it returns \code{NULL}.
}
\keyword{internal}
