% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{.int_initialise}
\alias{.int_initialise}
\title{Internal Function: Initialize Parameters for NMF Optimization}
\usage{
.int_initialise(A, k_range, loss, seed)
}
\arguments{
\item{A}{A numeric matrix representing the data to be factorized.}

\item{k_range}{A numeric vector specifying the range of factor numbers to
evaluate.}

\item{loss}{A character string, either \code{"mse"} (mean squared error) or
\code{"mkl"} (mean Kullback-Leibler divergence).}

\item{seed}{An integer specifying the random seed for reproducibility
(default is \code{NULL}).}
}
\value{
Nothing. Initializes the parameters and sets the random seed if
provided.
}
\description{
This function initializes parameters for Non-negative Matrix Factorization (NMF) optimization.
}
\details{
The function checks that the loss function is valid (\code{"mse"} or
\code{"mkl"}), warns about large datasets, sets the random seed if
specified, and ensures that the \code{k_range} sequence contains at least
three values.
}
\seealso{
\code{\link[base]{set.seed}} for setting the random seed.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{initialise}
\keyword{internal}
\keyword{nmf}
