% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setQCthresh.R
\name{.int_isZero}
\alias{.int_isZero}
\title{Identify Genes with Zero Expression}
\usage{
.int_isZero(sfe)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment containing spatial
transcriptomics experiment data.}
}
\value{
An updated SpatialFeatureExperiment object with a new column
"qc_isZero" in the rowData, indicating whether each gene has zero expression.
}
\description{
This function identifies genes with zero expression in a spatial
transcriptomics experiment
}
\details{
The function checks if the total expression for each gene is equal
to zero and adds a new column "qc_isZero" in the rowData of the
SpatialFeatureExperiment to mark genes with zero expression.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{SpatialFeatureExperiment}
\keyword{expression,}
\keyword{genes,}
\keyword{zero}
