% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setQCthresh.R
\name{.int_logLowMean}
\alias{.int_logLowMean}
\title{Identify Genes with Low Logarithmic Mean Expression}
\usage{
.int_logLowMean(sfe, threshold)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment containing spatial
transcriptomics experiment data.}

\item{threshold}{The threshold value for determining low logarithmic mean
expression.}
}
\value{
An updated SpatialFeatureExperiment object with a new column
"qc_isLogLow" in the rowData, indicating whether each gene has a logarithmic
mean expression below the specified threshold.
}
\description{
This function identifies genes with low logarithmic mean expression in a
spatial transcriptomics experiment
}
\details{
The function calculates the logarithmic mean expression for each
gene and checks if it is below the specified threshold. It then adds a new
column "qc_isLogLow" in the rowData of the single-cell experiment to mark
genes with low logarithmic mean expression.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{SpatialFeatureExperiment}
\keyword{expression,}
\keyword{genes,}
\keyword{logarithmic}
\keyword{mean}
