% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{.int_logTransformCounts}
\alias{.int_logTransformCounts}
\title{Internal: Only Log2-transform gene counts}
\usage{
.int_logTransformCounts(sfe, assay.type = "counts")
}
\arguments{
\item{sfe}{A A SpatialFeatureExperiment object.}

\item{assay.type}{Character string. The counts assay name to use. Defaults
to "counts"}
}
\value{
An SFE object with a new counts table named "unNormLogCounts".
}
\description{
This function is transforming gene counts using log2(counts + 1). It does
NOT use library sizes to normalise the data.
}
\author{
Eleftherios (Lefteris) Zormpas
}
