% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{.int_matchNamesToEnsgID}
\alias{.int_matchNamesToEnsgID}
\title{Internal Function: match gene names to EnsgIDs}
\usage{
.int_matchNamesToEnsgID(sfe, genes)
}
\arguments{
\item{sfe}{An SFE object.}

\item{genes}{A vector of EnsgIDs or gene names. A mixture of both throws an
error.}
}
\value{
a vector of EnsgIDs
}
\description{
This internal function takes an SFE object and a vector of gene names or
EnsgIDs, checks if there are gene names and returns their EnsgIDs from the
SFE object.
}
\author{
Eleftherios (Lefteris) Zormpas
}
