% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_moranI.R
\name{.int_moran}
\alias{.int_moran}
\title{Internal Function: Calculate Moran's I}
\usage{
.int_moran(gene, sfe, listw, n, S0, zero.policy, NAOK)
}
\description{
This internal function calculates the I statistic for a vector of gene
expression. For the parameter arguments check the
\code{link[STExplorer]{moranGlobalI}} function.
}
\author{
Eleftherios (Lefteris) Zormpas
}
