% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_moranI.R
\name{.int_moranIPerm}
\alias{.int_moranIPerm}
\title{Internal Function: Calculate Moran's I with Permutation testing}
\usage{
.int_moranIPerm(
  gene,
  sfe,
  assay,
  listw,
  nsim,
  alternative,
  na.action,
  spChk,
  adjust.n,
  return_boot,
  zero.policy
)
}
\description{
This internal function calculates the I statistic for a vector of gene
expression with Monte-Carlo permutation testing of significance. For the
parameter arguments check the \code{link[STExplorer]{moranGlobalIPerm}}
function.
}
\author{
Eleftherios (Lefteris) Zormpas
}
