% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_moranI.R
\name{.int_moranLocalI}
\alias{.int_moranLocalI}
\title{Internal Function: Calculate local Moran's I}
\usage{
.int_moranLocalI(
  gene,
  sfe,
  listw,
  zero.policy,
  na.action,
  conditional,
  alternative,
  mlvar,
  spChk,
  adjust.x
)
}
\description{
This internal function calculates the local Ii statistic for a vector of
gene expression. For the parameter arguments check the
\code{link[STExplorer]{moranLocalI}} function.
}
\author{
Eleftherios (Lefteris) Zormpas
}
