% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_moranI.R
\name{.int_moranLocalIPerm}
\alias{.int_moranLocalIPerm}
\title{Internal Function: Calculate local Moran's I with statistical testing}
\usage{
.int_moranLocalIPerm(
  gene,
  sfe,
  assay,
  listw,
  nsim,
  alternative,
  na.action,
  zero.policy,
  mlvar,
  spChk,
  adjust.x,
  sample_Ei,
  iseed,
  no_repeat_in_row
)
}
\description{
This internal function calculates the local Ii statistic for a vector of
gene expression alongside a statistical significance testing. For the
parameter arguments check the \code{link[STExplorer]{moranLocalIPerm}}
function.
}
\author{
Eleftherios (Lefteris) Zormpas
}
