% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_getSubset.R
\name{.int_msfeGetSubSetCol}
\alias{.int_msfeGetSubSetCol}
\title{INTERNAL: Retrieve Subset of Rows based on Regular Expression from a
SpatialFeatureExperiment}
\usage{
.int_msfeGetSubSetCol(sfe, col_name, subset)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object.}

\item{col_name}{The name of the column to be used for subsetting.}

\item{subset}{A regular expression string to be used for subsetting.}
}
\value{
A logical vector indicating the subset of features based on the
criteria.
}
\description{
INTERNAL: Retrieve Subset of Rows based on Regular Expression from a
SpatialFeatureExperiment
}
