% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{.int_nmfPlotDataMap}
\alias{.int_nmfPlotDataMap}
\title{Internal Function: generate data for NMF factor scores}
\usage{
.int_nmfPlotDataMap(nmf, sfe)
}
\arguments{
\item{nmf}{An object containing NMF results, typically obtained from the
\code{fgwc_nmf} function.}

\item{sfe}{A SpatialFeatureExperiment object containing spatial coordinates
and features.}
}
\value{
A data frame containing information for generating NMF plots.
}
\description{
This function generates data suitable for plotting FFGWC results as violins.
}
\details{
The function generates data suitable for plotting NMF results as
map.
}
\seealso{
\code{\link[SpatialFeatureExperiment]{colGeometry}},
\code{\link[SpatialFeatureExperiment]{colData}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{FGWC,}
\keyword{clustering,}
\keyword{coordinates}
\keyword{spatial}
