% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_genes.R
\name{.int_perGeneLogMean}
\alias{.int_perGeneLogMean}
\title{Internal Function: Calculate Per-Gene Logarithmic Mean for a Specific Sample
in a MetaSpatialFeatureExperiment object}
\usage{
.int_perGeneLogMean(id, sfe)
}
\arguments{
\item{id}{The unique identifier of the sample.}

\item{sfe}{The SpatialFeatureExperiment object for the specified sample.}
}
\value{
The SpatialFeatureExperiment object with per-gene logarithmic means
calculated for the specified sample.
}
\description{
This internal function calculates the per-gene logarithmic mean for a
specific sample in a MetaSpatialFeatureExperiment object.
}
\details{
The function calculates the per-gene logarithmic mean for a
specific sample in a SpatialFeatureExperiment. It considers the log counts
and the number of locations a gene is expressed in. The resulting log mean
is saved in the rowData of the SpatialFeatureExperiment object.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{SpatialFeatureExperiment}
\keyword{experiment,}
\keyword{function,}
\keyword{internal}
\keyword{log}
\keyword{mean,}
\keyword{per-gene}
\keyword{spatial}
\keyword{transcriptomics}
