% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{.int_pheatInput}
\alias{.int_pheatInput}
\title{Internal Function: Prepare Input Data for Plotting a Heatmap with Feature
Grouping}
\usage{
.int_pheatInput(marker_clusts, cluster_no, markers)
}
\arguments{
\item{marker_clusts}{A data frame containing information about marker genes
and their cluster assignments.}

\item{cluster_no}{Numeric, the cluster or subcluster number for which to
prepare the input data.}

\item{markers}{A data frame containing information about FGWC markers,
including gene names.}
}
\value{
A scaled and ordered data frame suitable for generating a heatmap
with cell type.
}
\description{
This function prepares input data for plotting a heatmap with feature
grouping based on cluster or subcluster results.
}
\details{
The function extracts information about marker genes and their
cluster or subcluster assignments from the provided data frame. It scales
the data, sets NaN values to zero, orders the data based on FGWC marker
genes, and drops any introduced NAs, resulting in a data frame ready for
plotting a heatmap with cell type.
}
\seealso{
\code{\link{scale}}, \code{\link[dplyr]{filter}},
\code{\link[dplyr]{select}}, \code{\link[tidyr]{drop_na}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{cell}
\keyword{clustering}
\keyword{heatmap,}
\keyword{type,}
