% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_plot.R
\name{.int_plotGWRCoeff}
\alias{.int_plotGWRCoeff}
\title{Internal Function: Plot GWR Coefficients}
\usage{
.int_plotGWRCoeff(gwr_sf, i, Var)
}
\arguments{
\item{gwr_sf}{A Simple Features (sf) object containing the spatial data and
the GWR coefficients. This object should include columns for the
coefficients and their significance, as well as the geometry information.}

\item{i}{An integer.}

\item{Var}{A character string indicating the variable to be selected for
plotting.}
}
\value{
A \code{ggplot} object representing the spatial distribution of the
specified GWR coefficient, with significant areas highlighted.
}
\description{
This internal function generates a spatial plot for the coefficients
obtained from a Geographically Weighted Regression (GWR) model. The function
highlights areas where the coefficients are statistically significant and
provides a visual representation of the intercept or a specified predictor
variable.
}
\details{
This function is an internal utility used to plot the spatial distribution
of GWR model coefficients. Depending on the value of the \code{i} parameter, the
function plots either the intercept or the specified predictor variable. The
plot uses a colour gradient to represent the magnitude of the coefficients
and overlays the significant areas with a thin black outline.

The function is designed to be used within the \code{plotGWR} function and is not
intended for direct use by end-users.
}
\seealso{
\code{\link{plotGWR}}, \code{\link{gwr_toSF}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{GWR}
\keyword{coefficients}
\keyword{ggplot}
\keyword{internal}
\keyword{plot}
\keyword{spatial}
