% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNeighbourGraph.R
\name{.int_plotNBgraph}
\alias{.int_plotNBgraph}
\title{Internal Function: Plot Neighbour Graph with Optional Tissue Image}
\usage{
.int_plotNBgraph(id, plotImage, image_list, nb_data_list, limits_list)
}
\arguments{
\item{id}{The sample ID.}

\item{plotImage}{Logical, indicating whether to plot the tissue image.
Default is TRUE.}

\item{image_list}{List containing image data for each sample.}

\item{nb_data_list}{List containing neighbour graph data for each sample.}

\item{limits_list}{List containing capture area limits for each sample.}
}
\description{
This internal function plots the neighbour graph with an optional overlay
of tissue image for a selected sample.
}
\details{
This function is an internal utility to visualize the neighbour graph for a
selected sample. It allows the option to overlay the tissue image on the
graph. The capture area limits are considered for proper spatial alignment.
}
\seealso{
\code{\link{ggplot}}, \code{\link{tidyterra::geom_spatraster_rgb}},
\code{\link{ggplot2::geom_sf}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{ggplot2}
\keyword{plotting}
\keyword{spatial}
\keyword{visualization}
