% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_plot.R
\name{.int_prepareDataMap}
\alias{.int_prepareDataMap}
\title{Internal: prepare Data Map}
\usage{
.int_prepareDataMap(sfe, assay, genes, minmax, type)
}
\arguments{
\item{sfe}{An object of class 'SpatialFeatureExperiment' containing spatial
data.}

\item{assay}{Character vector specifying the type of assay data
to use. Options are "counts" for raw counts and
"logcounts" for log2-normalized counts.}

\item{genes}{A vector of gene names or ENSG_IDs to be plotted.}

\item{minmax}{Numeric vector of length 2 specifying the minimum
and maximum expression values to include in the
plot.}

\item{type}{Character vector specifying the type of spatial data
to use. Options are "spot" for spot geometry and "hex"
for hexagon geometry.}
}
\value{
A data frame containing gene expression data prepared for
plotting on spatial feature experiment objects.
}
\description{
A function to prepare data for plotting gene expression on spatial
feature experiment (SFE) objects.
}
