% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_plot.R
\name{.int_prepareGlobalSAPlotData}
\alias{.int_prepareGlobalSAPlotData}
\title{Internal Function: Prepare Global SA summary data from plotting}
\usage{
.int_prepareGlobalSAPlotData(data, moran_thresh, geary_thresh, getis_thresh)
}
\arguments{
\item{data}{A subset of SFE's rowData}

\item{moran_thresh}{A numeric vector of length 2 specifying the threshold
range for Moran's I values to be considered significant.
Default is c(-0.5, 0.5).}

\item{geary_thresh}{A numeric vector of length 2 specifying the threshold
range for Geary's C values to be considered significant.
Default is c(0.5, 1.5).}

\item{getis_thresh}{A numeric value specifying the threshold for Getis-Ord's
G statistic to be considered significant. Default is 2.}
}
\value{
A long format data frame.
}
\description{
Internal Function: Prepare Global SA summary data from plotting
}
\author{
Eleftherios (Lefteris) Zormpas
}
