% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_plot.R
\name{.int_prepareSAPlotData}
\alias{.int_prepareSAPlotData}
\title{Internal Function: Prepare Data for Spatial Autocorrelation Plot}
\usage{
.int_prepareSAPlotData(
  sfe,
  feature,
  localRes,
  pVal,
  type,
  plot = c("localStat", "cluster")
)
}
\arguments{
\item{sfe}{An object of class 'SpatialFeatureExperiment' containing spatial
data.}

\item{feature}{A character string indicating the feature of interest.}

\item{localRes}{A character string with the name of the local results.}

\item{pVal}{A numeric value representing the statistical significance
threshold.}

\item{type}{A character string specifying the type of plot ("spot" or "hex").}

\item{plot}{A character string specifying the type of plot ("localStat" or
"cluster").}
}
\value{
A data frame containing the data for spatial autocorrelation plot.
}
\description{
Internal Function: Prepare Data for Spatial Autocorrelation Plot
}
\author{
Eleftherios (Lefteris) Zormpas
}
