% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_plot.R
\name{.int_prepDtMapGWPCA}
\alias{.int_prepDtMapGWPCA}
\alias{.int_arrayToListGWPCA}
\title{Internal: prepare Data Map for GWPCA}
\usage{
.int_prepDtMapGWPCA(gwpca, locs, genes, comps, type, gene_names, absolute)

.int_arrayToListGWPCA(x)
}
\arguments{
\item{gwpca}{An object of class 'gwpca' containing GWPCA data.}

\item{locs}{locations to select -- currently out of use --> select all}

\item{genes}{A vector of gene names or ENSG_IDs to be plotted.}

\item{comps}{the components to be plotted.}

\item{type}{Character vector specifying the type of spatial data
to use. Options are "spot" for spot geometry and "hex"
for hexagon geometry.}

\item{gene_names}{The gene names}

\item{absolute}{TRUE/FALSE. Absolute leading scores or not?}
}
\value{
A data frame containing leading score gene data prepared for
plotting on a map.
}
\description{
A function to prepare data for plotting leading scores for specific genes
from a GWPCA object.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{component,}
\keyword{ggplot2,}
\keyword{internal}
\keyword{leading}
\keyword{plot,}
\keyword{principal}
\keyword{scores}
\keyword{spatial}
