% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{.int_removeDuplicates}
\alias{.int_removeDuplicates}
\title{Remove Duplicates from cell type Markers}
\usage{
.int_removeDuplicates(markers, column)
}
\arguments{
\item{markers}{A data frame containing information about cell type
markers, including gene names and types.}

\item{column}{The column based on which duplicates are identified
and removed.}
}
\value{
A data frame with duplicate entries removed based on the specified
column.
}
\description{
This function removes duplicate entries from cell type markers based
on a specified column.
}
\details{
The function uses the \code{BiocGenerics::duplicated} function to
identify and tag duplicate entries in the specified column. It then filters
out those entries, resulting in a data frame with unique cell type
markers based on the chosen column.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{cell}
\keyword{duplicates,}
\keyword{markers,}
\keyword{removal}
\keyword{type,}
