% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_utils.R
\name{dot-int_resSwitch}
\alias{dot-int_resSwitch}
\alias{.int_resSwitch}
\title{Internal: switch resolution name}
\usage{
.int_resSwitch(.res)
}
\arguments{
\item{.res}{resolution name}
}
\description{
IMPORTANT: this function is not exported to be used alone.
A function that selects the name of the spot diameter from the required
resolution which is stored inside SpatialFeatureExperiment's metadata.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
