% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_FunctionalClust.R
\name{.int_runGSEA}
\alias{.int_runGSEA}
\title{Internal Function: Run Gene Set Enrichment Analysis (GSEA) for a Single
Location}
\usage{
.int_runGSEA(
  X,
  inputGSEA,
  minGSSize,
  pvalueCutoff,
  TERM2GENE,
  pAdjustMethod,
  verbose,
  scoreType,
  nPermSimple,
  ...
)
}
\arguments{
\item{X}{The index of the spatial location for which GSEA is performed.}

\item{inputGSEA}{The input gene expression data for the specified spatial
location.}

\item{minGSSize}{The minimum gene set size to be considered in GSEA.}

\item{pvalueCutoff}{The p-value cutoff for identifying significant gene sets
in GSEA.}

\item{TERM2GENE}{The term-to-gene mapping for gene sets.}

\item{pAdjustMethod}{The method for multiple testing correction in GSEA.}

\item{verbose}{A logical value indicating whether to print verbose output
during GSEA.}

\item{scoreType}{The GSEA scoring type.}

\item{nPermSimple}{The number of permutations for simple GSEA.}

\item{...}{Additional arguments to be passed to the GSEA function.}
}
\value{
A dataframe containing the results of GSEA for the specified spatial
location.
}
\description{
This function performs GSEA for a single spatial location using a given set
of parameters. It is an internal function used within the main analysis
function and is not intended for direct use.
}
\details{
This function is used internally to perform Gene Set Enrichment Analysis
(GSEA) for a single spatial location based on the provided gene expression
data and parameters. The results are returned as a dataframe.
}
\examples{
# This function is an internal function used within the main function.
# Refer to the main analysis function for example usage.

}
\seealso{
\code{\link{gwpca_FunctionalClustering}}, \code{\link[clusterProfiler]{GSEA}}

Other Spatial Transcriptomics Analysis: 
\code{\link{gwpca_FunctionalClustering}()},
\code{\link{plotGWPCA_FuncCLust}()}
}
\concept{Spatial Transcriptomics Analysis}
\keyword{GSEA}
\keyword{analysis}
\keyword{enrichment}
\keyword{gene}
\keyword{internal}
\keyword{set}
\keyword{spatial}
\keyword{transcriptomics}
