% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_utils.R
\name{.int_SAgeneCheck}
\alias{.int_SAgeneCheck}
\alias{dot-int_SAgeneCheck}
\title{Internal Function: check genes input}
\usage{
.int_SAgeneCheck(sfe, genes)
}
\description{
This internal function checks for the genes input in the main function.
}
\author{
Eleftherios (Lefteris) Zormpas
}
