% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_setQCthresh.R
\name{.int_selectFilters}
\alias{.int_selectFilters}
\title{Select QC Metrics for Filtering}
\usage{
.int_selectFilters(sfe, filters)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object containing spatial
transcriptomics experiment data.}

\item{filters}{TRUE selects all calculated QC metrics. Alternatively,
provide a character vector with the column names of the QC metrics you want
to consider for filtering. The QC metric columns begin with "qc_".}
}
\value{
A character vector representing the selected QC metrics for
filtering.
}
\description{
This function selects QC metrics for filtering based on specified criteria.
}
\details{
The function provides flexibility in selecting QC metrics for
filtering. If \code{filters} is set to TRUE, all QC metrics are selected. If it
is a character vector, the function validates and selects the specified QC
metrics. The selected columns are then returned for further processing.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{QC}
\keyword{filtering,}
\keyword{metrics,}
\keyword{selection}
