% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{.int_selectGeom}
\alias{.int_selectGeom}
\title{Internal: Select geometry}
\usage{
.int_selectGeom(sfe, type = c("spot", "hex", "centroid"))
}
\arguments{
\item{sfe}{An object of class SpatialFeatureExperiment.}

\item{type}{A character string specifying the type of geometry to select.
Options are "spot" (spot geometry), "hex" (hexagon geometry), or "centroid"
(centroid geometry). Default is "spot".}
}
\value{
A geometry object representing the selected geometry type.
}
\description{
An internal function to select geometries from a SpatialFeatureExperiment
object based on the specified type.
}
\details{
This function retrieves the specified type of geometry from a
SpatialFeatureExperiment object. The available options for the type
parameter are "spot" (spot geometry), "hex" (hexagon geometry), and
"centroid" (centroid geometry).
}
\author{
Eleftherios Zormpas
}
