% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metaSFE.R
\name{.int_sfeORmsfe}
\alias{.int_sfeORmsfe}
\title{Internal Function: Get SpatialFeatureExperiment or
MetaSpatialFeatureExperiment}
\usage{
.int_sfeORmsfe(m_sfe, sample_id)
}
\arguments{
\item{m_sfe}{An object of class SpatialFeatureExperiment or
MetaSpatialFeatureExperiment.}

\item{sample_id}{A character vector specifying the sample IDs to include
(only relevant for MetaSpatialFeatureExperiment).}
}
\value{
An object of class SpatialFeatureExperiment. If the input is already
an SFE, it returns the original object. If the input is a MetaSFE, it
returns a subset based on the provided sample IDs.
}
\description{
This internal function checks the class type of the input object and returns
either the original SpatialFeatureExperiment (SFE) object or, in the case of
a MetaSpatialFeatureExperiment (MetaSFE), retrieves a subset based on the
provided sample IDs.
}
