% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_FunctionalClust.R
\name{.int_sortGenes}
\alias{.int_sortGenes}
\title{Internal Function: Sort Genes in Descending Order}
\usage{
.int_sortGenes(data)
}
\arguments{
\item{data}{The gene expression data to be sorted.}
}
\value{
A vector containing the sorted gene expression values.
}
\description{
This function sorts genes in descending order based on their expression
values. It is an internal function used within other functions and is not
intended for direct use.
}
\details{
This function sorts genes in descending order based on their expression
values. It is used internally within other functions to sort genes before
further processing.
}
\examples{
# This function is an internal function used within other functions.
# Refer to the documentation of the calling function for example usage.

}
\seealso{
\code{\link{.int_generateGeneList}}
}
\keyword{data}
\keyword{expression}
\keyword{gene}
\keyword{internal}
\keyword{list}
\keyword{sorting}
