% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.QC.Sparsity.R
\name{dot-int_sparsity_2}
\alias{dot-int_sparsity_2}
\alias{.int_sparsity_2}
\title{Internal: calculate location sparsity}
\usage{
.int_sparsity_2(sfe, assay)
}
\arguments{
\item{sfe}{a SpatialFeaturesExperiment objects.}

\item{assay}{the name of the assay to use.}
}
\description{
IMPORTANT: this function is not exported to be used alone.
A function to calculate the sparsity of the dataset and get the
sparsity of each location.
}
\keyword{internal}
