% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_utils.R
\name{dot-int_spotHex_minMax}
\alias{dot-int_spotHex_minMax}
\alias{.int_spotHex_minMax}
\title{Internal: get min and max array col/row}
\usage{
.int_spotHex_minMax(.data)
}
\arguments{
\item{.data}{A data frame containing the spot coordinates.}
}
\value{
A list with the following elements:
\describe{
\item{xmax}{The maximum x-coordinate value.}
\item{xmin}{The minimum x-coordinate value.}
\item{ymax}{The maximum y-coordinate value.}
\item{ymin}{The minimum y-coordinate value.}
}
}
\description{
IMPORTANT: this function is not exported to be used alone.
This function calculates the minimum and maximum values of x and y
coordinates for the spots in the provided data.
}
\details{
This function is used internally to calculate the minimum and maximum x and y
coordinates for the spots in the provided data frame.
}
\author{
Eleftherios (Lefteris) Zormpas
}
