% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_setQCthresh.R
\name{.int_summaryTable}
\alias{.int_summaryTable}
\title{Generate Summary Table for QC Metrics}
\usage{
.int_summaryTable(sfe, filters)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment containing spatial
transcriptomics experiment data.}

\item{filters}{TRUE selects all calculated QC metrics. Alternatively,
provide a character vector with the column names of the QC metrics you want
to consider for filtering. The QC metric columns begin with "qc_".}
}
\value{
A data frame representing a summary table for the selected QC
metrics, showing the sum of each metric across all samples.
}
\description{
This function generates a summary table for selected QC metrics in a spatial
transcriptomics experiment.
}
\details{
The function uses the internal function \code{.int_selectFilters} to
obtain the selected QC metrics. It then creates a matrix with one row and
columns corresponding to the selected metrics. The function calculates the
sum of each metric across all samples and populates the summary table
accordingly.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{QC}
\keyword{SpatialFeatureExperiment}
\keyword{metrics,}
\keyword{summary}
\keyword{table,}
