% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{.int_tissueImgPlot}
\alias{.int_tissueImgPlot}
\title{Internal Function: .int_tissueImgPlot}
\usage{
.int_tissueImgPlot(
  image,
  image_name,
  limits_list,
  sfe,
  type,
  annotate,
  colours,
  fill_args,
  alpha,
  ...
)
}
\arguments{
\item{image}{List of raster objects.}

\item{image_name}{Character vector specifying the image name.}

\item{limits_list}{List of x and y limits for image plotting.}

\item{sfe}{A SpatialFeatureExperiment object.}

\item{type}{Character vector specifying the plot type ("none", "spot", or
"hex").}

\item{annotate}{Logical. If TRUE, includes annotations in the plot.}

\item{colours}{Vector of colours for annotations.}

\item{fill_args}{List of fill arguments.}

\item{alpha}{Numeric. Transparency level for image plotting.}

\item{...}{Additional arguments to pass to ggplot2::scale_fill_manual.}
}
\value{
Returns a ggplot2 plot object for tissue images with optional
annotations.
}
\description{
Generates a ggplot2 plot for tissue images with optional annotations.
}
\details{
This function generates a ggplot2 plot for tissue images with
optional annotations, allowing customisation of various plot elements.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
