% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{.int_validateColumnNames}
\alias{.int_validateColumnNames}
\title{Validate Column Names in a Data Frame}
\usage{
.int_validateColumnNames(df, expected_columns)
}
\arguments{
\item{df}{A data frame to validate.}

\item{expected_columns}{A character vector specifying the expected column
names.}
}
\value{
NULL if validation is successful; otherwise, an error is raised.
}
\description{
This function validates the column names in a data frame, ensuring they
match the expected column names.
}
\details{
The function checks if the actual column names in the provided data
frame match the expected column names. If there is a mismatch, it raises an
error indicating the expected and actual column names for user correction.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{column}
\keyword{data}
\keyword{frame,}
\keyword{names}
\keyword{validation,}
