% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{.int_validateInput}
\alias{.int_validateInput}
\title{Internal Function: Validate Input}
\usage{
.int_validateInput(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{Another numeric vector.}
}
\value{
Nothing. Stops execution if the input vectors are not of the same
length.
}
\description{
This function checks if two numeric vectors have the same length.
}
\details{
The function ensures that both input vectors are of the same length. If not,
it raises an error.
}
\seealso{
\code{\link[base]{stopifnot}} for input validation.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
\keyword{validation}
