% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{fgwc_findOptimumK}
\alias{fgwc_findOptimumK}
\title{Identify the Optimal Number of Clusters Using FGWC}
\usage{
fgwc_findOptimumK(
  fgwc_in,
  k_range = 2:10,
  index_type = "FPC",
  elbow_method = "angle",
  m_sfe,
  sample_id = NULL,
  algorithm = "classic",
  distMat = NULL,
  dMetric = NULL,
  parameters,
  plot = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{fgwc_in}{The matrix (locations x factors/components) that will be used
for the actual FGWC as generated by the \code{fgwc_nmf} function.}

\item{k_range}{An integer vector specifying the range of \code{k} values to
evaluate (default is \code{2:10}).}

\item{index_type}{A character string indicating which index type to use for
evaluation. Must be one of \code{"composite"}, \code{"FPC"}, \code{"CE"},
\code{"SC"}, or \code{"XB"} (default is \code{"FPC"}).}

\item{elbow_method}{A character string specifying the method for detecting
the elbow point. Options are \code{"angle"} or \code{"knee"} (default is
\code{"angle"}).}

\item{m_sfe}{The \code{SpatialFeatureExperiment} or
\code{MetaSpatialFeatureExperiment} object containing spatial expression
data.}

\item{sample_id}{A character string indicating the name of the sample. It is
required when the \code{m_sfe} argument is provided with a MetaSFE object.}

\item{algorithm}{A character string indicating the algorithm used for FGWC.
Currently accepting one of these two options: \code{"classic"} or
\code{"abc"}.}

\item{distMat}{An \code{n*n} distance matrix between regions.}

\item{dMetric}{Character string specifying the distance metric.}

\item{parameters}{A vector that consists of FGWC parameters
(see \code{\link{fgwc_params}} for parameter details).}

\item{plot}{Logical, indicating whether to plot the index values with the
optimal \code{k} highlighted (default is \code{TRUE}).}

\item{verbose}{Logical, indicating whether to print progress messages
(default is \code{FALSE}).}

\item{...}{Additional arguments passed to the \code{fgwcSTE} function.}
}
\value{
An integer representing the optimal number of clusters.
}
\description{
This function identifies the optimal number of clusters (\code{k}) for Fuzzy
Geographically Weighted Clustering (FGWC) by evaluating different clustering
indices over a range of \code{k} values and finding the elbow point.
}
\details{
The function evaluates different clustering indices over a range of \code{k}
values and identifies the elbow point using either the angle or knee method.

If the \code{index_type} is set to \code{"composite"}, a composite score is
calculated based on multiple indices, and the elbow point is determined
using the composite score.
}
\examples{
\dontrun{
# Assuming `fgwc_input` is a matrix generated by the `fgwc_nmf` function
# and `sfe` is your SpatialFeatureExperiment object
optimal_k <- fgwc_findOptimumK(fgwc_in = fgwc_input, m_sfe = sfe)
}

}
\seealso{
\code{\link{fgwcSTE}}, \code{\link{fgwc_params}}
}
\keyword{clustering}
\keyword{fgwc}
\keyword{spatial-expression}
