% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{fgwc_nmf}
\alias{fgwc_nmf}
\title{Perform Non-negative Matrix Factorisation (NMF)}
\usage{
fgwc_nmf(
  m_sfe,
  sample_id,
  assay = "logcounts",
  top_hvgs,
  ncomponents = 2,
  ntop = 600,
  subset_row = NULL,
  scale = FALSE,
  ...
)
}
\arguments{
\item{m_sfe}{The \code{SpatialFeatureExperiment} or
\code{MetaSpatialFeatureExperiment} object containing spatial expression
data.}

\item{sample_id}{The sample ID or index for which to perform clustering.}

\item{assay}{The name of the assay in \code{m_sfe} to use for clustering
(default is "logcounts").}

\item{top_hvgs}{A character vector of gene names representing the top highly
variable genes. The gene names are expected to be ENSGene IDs.}

\item{ncomponents}{The number of components (clusters) to identify
(default is 2). The components are the 'factors' calculated mby NMF.}

\item{ntop}{The maximum number of top features (genes) to use in the
clustering (default is 600).}

\item{subset_row}{An optional vector specifying the subset of rows to use in
the clustering.}

\item{scale}{Logical, indicating whether to scale the input data
(default is FALSE).}

\item{...}{Additional arguments to be passed to \code{scater::calculateNMF}.}
}
\value{
An NMF result object with factors representing the reduced dimensions
of the original expression matrix.
}
\description{
This function performs Non-negative Matrix Factorisation (NMF) on the gene
expression matrix present inside a SpatialFeatureExperiment or a
MetaSpatialFeatureExperiment.
}
\examples{
\dontrun{
# Assuming `sfe` is your SpatialFeatureExperiment object
nmf_result <- fgwc_nmf(m_sfe = sfe, sample_id = "ID1", top_hvgs = top_genes)
}

}
\seealso{
\code{\link[scater]{calculateNMF}} for additional details on NMF clustering.

Other clustering functions: 
\code{\link{fgwc_nmfFactorNumber}()},
\code{\link{fgwc_params}()}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\concept{clustering functions}
\keyword{clustering}
\keyword{nmf}
\keyword{spatial-expression}
