% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwcSTE.R
\name{fgwc_params}
\alias{fgwc_params}
\title{Generate Parameters for Fuzzy Geographically Weighted Clustering}
\usage{
fgwc_params(
  algorithm = c("classic", "abc", "fpa", "gsa", "hho", "ifa", "pso", "tlbo"),
  ncluster,
  kind = "u",
  m = 2,
  distance = "manhattan",
  order = 1,
  alpha = 0.5,
  a = 1,
  b = 1,
  max.iter = 500,
  error = 1e-05,
  randomN = 1,
  uij = NA,
  vi = NA
)
}
\arguments{
\item{algorithm}{The type of FGWC or optimization algorithm to use. Choose
from "classic", "abc", "fpa", "gsa", "hho", "ifa", "pso", "tlbo".}

\item{ncluster}{The number of clusters.}

\item{kind}{Use \code{'u'} for the membership approach and \code{'v'} for
the centroid approach.}

\item{m}{Degree of fuzziness or fuzzifier (default is 2).}

\item{distance}{The distance metric between data and centroid (default is
"euclidean").}

\item{order}{Minkowski order (default is 1).}

\item{alpha}{The old membership effect with \link{0,1}. If \code{alpha} equals 1,
it will be the same as Fuzzy C-Means. If 0, it equals the
neighborhood effect.}

\item{a}{Spatial magnitude of distance (default is 1).}

\item{b}{Spatial magnitude of population (default is 1).}

\item{max.iter}{Maximum iteration (default is 500).}

\item{error}{Error tolerance (default is 1e-5).}

\item{randomN}{Random seed for initialization (if \code{uij} or \code{vi}
is NA, default is 1).}

\item{uij}{Membership matrix initialization.}

\item{vi}{Centroid matrix initialization.}
}
\value{
A named vector of FGWC parameters or optimization algorithm
parameters.
}
\description{
This function generates parameters for Fuzzy Geographically Weighted
Clustering (FGWC) or optimization algorithms for FGWC.
}
\examples{
\dontrun{
# Example usage:
params <- fgwc_params(algorithm = "classic", ncluster = 3, kind = "u", m = 2)
}

}
\seealso{
\code{\link{fgwcuv}}, \code{\link{abcfgwc}}, \code{\link{fpafgwc}},
\code{\link{gsafgwc}}, \code{\link{hhofgwc}}, \code{\link{ifafgwc}},
\code{\link{psofgwc}}, \code{\link{tlbofgwc}}

Other clustering functions: 
\code{\link{fgwc_nmf}()}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\concept{clustering functions}
\keyword{clustering}
\keyword{fgwc}
\keyword{fuzzy-logic}
\keyword{optimization}
