% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwcSTE.R
\name{fgwcSTE}
\alias{fgwcSTE}
\title{#' Fuzzy Geographicaly Weighted Clustering}
\usage{
fgwcSTE(
  m_sfe,
  sample_id,
  data,
  pop = NULL,
  distMat = NULL,
  dMetric = NULL,
  algorithm = "classic",
  fgwc_param = NULL,
  opt_param = NULL
)
}
\arguments{
\item{m_sfe}{The \code{SpatialFeatureExperiment} or
\code{MetaSpatialFeatureExperiment} object containing spatial expression
data.}

\item{sample_id}{A character string indicating the name of the sample. It is
required when the \code{m_sfe} argument is provided with a MetaSFE object.}

\item{data}{an object of data with d>1. Can be \code{matrix} or
\code{data.frame}. If your data is univariate, bind it with \code{1} to get
a data frame with 2 columns.}

\item{pop}{an n*1 vector contains population.}

\item{distMat}{an n*n distance matrix between regions.}

\item{dMetric}{Character string specifying the distance metric.}

\item{algorithm}{algorithm used for FGWC}

\item{fgwc_param}{a vector that consists of FGWC parameter
(see \code{\link{fgwcuv}} for parameter details)}

\item{opt_param}{a vector that consists of optimization algorithm parameter
(see \code{\link{fgwcuv}} for parameter details)}
}
\value{
an object of class \code{"fgwc"}.\cr
An \code{"fgwc"} object contains as follows:
\itemize{
\item \code{converg} - the process convergence of objective function
\item \code{f_obj} - objective function value
\item \code{membership} - membership matrix
\item \code{centroid} - centroid matrix
\item \code{validation} - validation indices (there are partition
coefficient (\code{PC}), classification entropy (\code{CE}),
SC index (\code{SC}), separation index (\code{SI}), Xie and Beni's index
(\code{XB}), IFV index (\code{IFV}), and Kwon index (Kwon))
\item \code{max.iter} - Maximum iteration
\item \code{cluster} - the cluster of the data
\item \code{finaldata} - The final data (with the cluster)
\item \code{call} - the syntax called previously
\item \code{time} - computational time.
}
}
\description{
Wrapper around classic fuzzy clustering with addition of
spatial configuration of membership matrix. It supports optimisation
algorithms.
}
\details{
Fuzzy Geographically Weighted Clustering (FGWC) was developed by
\code{naspaclust} by adding neighbourhood effects and population to
configure the membership matrix in Fuzzy C-Means. There are two kinds of
options in doing classical FGWC. The first is using \code{"u"} (default)
for membership optimization and \code{"v"} for centroid optimisation.

There are seven optimisation algorithms that are currently provided in this
package, mainly from the \code{naspaclust}. The optimization algorithm uses
the centroid as the parameter to be optimised.

Here are the algorithms that can be used:
\itemize{
\item \code{"classic"} - The classical algorithm of FGWC based on
Mason and Jacobson (2007) for centroid optimisation
and Runkler and Katz (2006) for membership optimization.
\item \code{"abc"} - Optimization using artificial bee colony algorithm
based on Karaboga and Basturk (2007) (see also Wijayanto and Purwarianti
2014 and Wijayanto et al. 2016 for FGWC implementation).
\item \code{"fpa"} - Optimization using flower pollination algorithm based
on (Yang 2012).
\item \code{"gsa"} - Optimization using gravitational search algorithm based
on Rashedi et al. (2009) and Li and Dong (2017) (see also Pamungkas and
Pramana 2019 for FGWC implementation).
\item \code{"hho"} - Optimization using harris-hawk optimization with
\code{"heidari"} (Heidari et al. 2019) (default), and \code{"bairathi"}
(Bairathi and Gopalani 2018).
\item \code{"ifa"} - Optimization using intelligent firefly algorithm based
on Yang (2009), as well as the intelligent improvement by Fateen and
Bonilla-Petriciolet (2013) (see also Nasution et al. 2020 for FGWC
implementation).
\item \code{"pso"} - Optimization using particle swarm optimization based on
Runkler and Katz (2006) and Bansal et al. (2011) for inertia option (see also
Wijayanto and Purwarianti 2014; Putra and Kurniawan 2017; Abdussamad 2020
for FGWC implementation).
\item \code{"tlbo"} - Optimization using teaching - learning based
optimization based on Rao et al. (2012) and elitism improvement by Rao and
Patel (2012).
}
}
\references{
\itemize{
\item Abdussamad S (2020). "Evaluation of Implementation Context Based
Clustering In Fuzzy Geographically Weighted Clustering-Particle Swarm
Optimization Algorithm." Jurnal EECCIS, 14(1), 10--15. ISSN 2460-8122,
\url{https://jurnaleeccis.ub.ac.id/index.php/eeccis/article/view/609}.
\item Bairathi D, Gopalani D (2018). "A Novel Swarm Intelligence Based
Optimization Method: Harris' Hawk Optimization." In Advances in Intelligent
Systems and Computing, 832--842. Springer International Publishing.
doi: 10.1007/978-3-030-16660-1_81,
\url{https://doi.org/10.1007/978-3-030-16660-1_81}.
\item Bansal JC, Singh PK, Saraswat M, Verma A, Jadon SS, Abraham A (2011).
"Inertia Weight strategies in Particle Swarm Optimization." In 2011 Third
World Congress on Nature and Biologically Inspired Computing.
doi: 10.1109/nabic.2011.6089659,
\url{https://doi.org/10.1109/nabic.2011.6089659}.
\item Fateen SK, Bonilla-Petriciolet A (2013). "Intelligent Firefly Algorithm
for Global Optimization." Cuckoo Search and Firefly Algorithm: Theory and
Applications, 516, 315--330.
\item Heidari AA, Mirjalili S, Faris H, Aljarah I, Mafarja M, Chen H (2019).
"Harris hawks optimization: Algorithm and applications." Future Generation
Computer Systems, 97, 849--872. doi: 10.1016/j.future.2019.02.028,
\url{https://doi.org/10.1016/j.future.2019.02.028}.
\item Karaboga D, Basturk B (2007). "A powerful and efficient algorithm for
numerical function optimization: artificial bee colony (ABC) algorithm."
Journal of Global Optimization, 39(3), 459--471.
doi: 10.1007/s10898-007-9149-x, \url{https://doi.org/10.1007/s10898-007-9149-x}.
\item Li J, Dong N (2017). "Gravitational Search Algorithm with a New
Technique." In 2017 13th International Conference on Computational
Intelligence and Security (CIS), 516--519. doi: 10.1109/CIS.2017.00120,
\url{https://doi.org/10.1109/CIS.2017.00120}.
\item Mason GA, Jacobson RD (2007). "Fuzzy Geographically Weighted
Clustering." In Proceedings of the 9th International Conference on
Geocomputation, 1--7.
\item Nasution BI, Kurniawan R, Siagian TH, Fudholi A (2020). "Revisiting
social vulnerability analysis in Indonesia: An optimised spatial fuzzy
clustering approach." International Journal of Disaster Risk Reduction, 51,
101801. doi: 10.1016/j.ijdrr.2020.101801,
\url{https://doi.org/10.1016/j.ijdrr.2020.101801}.
\item Pamungkas IH, Pramana S (2019). "Improvement Method of Fuzzy
Geographically Weighted Clustering using Gravitational Search Algorithm."
Journal of Computer Science and Information, 11(1).
\item Putra FH, Kurniawan R (2017). "Clustering for Disaster Areas Endemic
Dengue Hemorrhagic Fever Based on Factors had Caused in East Java Using Fuzzy
Geographically Weighted Clustering - Particle Swarm Optimization."
Jurnal Aplikasi Statistika & Komputasi Statistik, 8(01), 27. ISSN 2615-1367.
\item Rao RV, Patel V (2012). "An elitist teaching-learning-based
optimization algorithm for solving complex constrained optimization
problems." International Journal of Industrial Engineering Computations,
3(4), 535--560. ISSN 19232926, doi: 10.5267/j.ijiec.2012.03.007,
\url{https://doi.org/10.5267/j.ijiec.2012.03.007}.
\item Rao RV, Savsani VJ, Balic J (2012). "Teaching- learning-based
optimization algorithm for unconstrained and constrained real-parameter
optimization problems." Engineering Optimization, 44(12), 1447--1462.
doi: 10.1080/0305215x.2011.652103,
\url{https://doi.org/10.1080/0305215x.2011.652103}.
\item Rashedi E, Nezamabadi-pour H, Saryazdi S (2009). "GSA: A Gravitational
Search Algorithm." Information Sciences, 179(13).
\item Runkler TA, Katz C (2006). "Fuzzy Clustering by Particle Swarm
Optimization." In 2006 IEEE International Conference on Fuzzy Systems.
doi: 10.1109/fuzzy.2006.1681773,\url{https://doi.org/10.1109/fuzzy.2006.1681773}.
\item Wijayanto AW, Purwarianti A (2014). "Improvement design of fuzzy
geo-demographic clustering using Artificial Bee Colony optimization." I
n 2014 International Conference on Cyber and IT Service Management (CITSM),
69--74. ISBN 978-1-4799-7975-2.
\item Wijayanto AW, Purwarianti A (2014). "Improvement of fuzzy
geographically weighted clustering using particle swarm optimization."
In 2014 International Conference on Information Technology Systems and
Innovation (ICITSI), 7--12. ISBN 978-1-4799-6527-4.
\item Wijayanto AW, Purwarianti A, Son LH (2016). "Fuzzy geographically
weighted clustering using artificial bee colony: An efficient geo-demographic
analysis algorithm and applications to the analysis of crime behavior in
population." Applied Intelligence, 44(2), 377--398. ISSN 0924-669X.
\item Yang X (2014). Nature-Inspired Optimization Algorithms, Elsevier
insights. Elsevier Science. ISBN 9780124167452.
\item Yang X (2012). "Flower Pollination Algorithm for Global Optimization."
In Unconventional Computation and Natural Computation, 240--249. Springer
Berlin Heidelberg. doi: 10.1007/978-3-642-32894-7_27,
\url{https://doi.org/10.1007/978-3-642-32894-7_27}.
\item Yang X (2009). "Firefly Algorithms for Multimodal Optimization." In
Stochastic Algorithms: Foundations and Applications, 169--178. Springer
Berlin Heidelberg. doi: 10.1007/978-3-642-04944-6_14,
\url{https://doi.org/10.1007/978-3-642-04944-6_14}.
}
}
\seealso{
\code{\link[naspaclust]{fgwcuv}}, \code{\link[naspaclust]{abcfgwc}},
\code{\link[naspaclust]{fpafgwc}}, \code{\link[naspaclust]{gsafgwc}},
\code{\link[naspaclust]{hhofgwc}}, \code{\link[naspaclust]{ifafgwc}},
\code{\link[naspaclust]{psofgwc}}, \code{\link[naspaclust]{tlbofgwc}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
