% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_geary.R
\name{gearyGlobalC}
\alias{gearyGlobalC}
\title{Calculate Geary's C Global Statistic}
\usage{
gearyGlobalC(x, listw, n = NULL, n1 = NULL, S0 = NULL, zero.policy = NULL)
}
\arguments{
\item{x}{A numeric vector of observations. Must be the same length as the
neighbours list in listw.}

\item{listw}{A listw object created, for example, by
\code{link[spdep]{nb2listw}} or \code{link[spdep]{nb2listwdist}}.}

\item{n}{Number of zones. If not provided, it is calculated as
length(listw$neighbours).}

\item{n1}{\code{n - 1}. If not provided, it is calculated as
\code{length(listw$neighbours) - 1}.}

\item{S0}{Global sum of weights. If not provided, it is calculated as
\code{spdep::Szero(listw)}.}

\item{zero.policy}{Default \code{attr(listw, "zero.policy")} as set when
\code{listw} was created.
If TRUE, assign zero to the lagged value of zones without neighbours.
If FALSE, assign NA. If attribute not set, use the global option value.}
}
\value{
Geary's C statistic. A list with C: Geary's C and K: sample kurtosis
of x
}
\description{
This function is a wrapper of the \code{\link[spdep]{geary}} function from
the \code{spdep} package developed by Roger Bivand. It computes Geary's C
statistic for spatial autocorrelation using the spdep package.
}
\details{
If n, n1, and S0 are not provided, they are calculated based on the
\code{listw} object.
}
\examples{
\dontrun{
# Export an SFE object from an MSFE object
sfe <- getSFE(msfe, "JBO019")
# A vector of gene expression of one gene over all locations
gene_exp <- assay(sfe, "logcounts")[1,]
# A listw object containing distance-based spatial weights for neighbors
listw <- listw
# Calculate Geary's C for this gene
geary <- gearyGlobalC(x = gene_exp, listw = listw, zero.policy = TRUE)
}

}
\seealso{
\code{\link[spdep]{geary}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{autocorrelation}
\keyword{geary}
\keyword{listw}
\keyword{spatial}
