% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_geary.R
\name{gearyGlobalCPerm}
\alias{gearyGlobalCPerm}
\title{Calculate Geary's C Global Statistic with permutation testing}
\usage{
gearyGlobalCPerm(
  x,
  listw,
  nsim,
  zero.policy = attr(listw, "zero.policy"),
  alternative = "greater",
  spChk = NULL,
  adjust.n = TRUE,
  return_boot = FALSE
)
}
\arguments{
\item{x}{A numeric vector of observations. Must be the same length as the
neighbours list in listw.}

\item{listw}{A listw object created, for example, by
\code{link[spdep]{nb2listw}} or \code{link[spdep]{nb2listwdist}}.}

\item{nsim}{Number of permutations.}

\item{zero.policy}{Default \code{attr(listw, "zero.policy")} as set when
\code{listw} was created.
If TRUE, assign zero to the lagged value of zones without neighbours.
If FALSE, assign NA. If attribute not set, use the global option value.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "greater" (default), or "less"; this reversal corresponds to
that on geary.test described in the section on the output statistic value,
based on Cliff and Ord 1973, p. 21.}

\item{spChk}{should the data vector names be checked against the spatial
objects for identity integrity, TRUE, or FALSE, default NULL to use
\code{get.spChkOption()}.}

\item{adjust.n}{Default TRUE, if FALSE the number of observations is not
adjusted for no-neighbour observations, if TRUE, the number of observations
is adjusted.}

\item{return_boot}{Return an object of class boot from the equivalent
permutation bootstrap rather than an object of class htest}
}
\value{
A list with class \code{htest} and \code{mc.sim} containing the
following components:
\itemize{
\item{statistic}{the value of the observed Geary's C.}
\item{parameter}{the rank of the observed Geary's C.}
\item{p.value}{the pseudo p-value of the test.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string giving the method used.}
\item{data.name}{a character string giving the name(s) of the data, and the
number of simulations.}
\item{res}{nsim simulated values of statistic, final value is observed
statistic}
}
}
\description{
This function is a wrapper of the \code{\link[spdep]{geary.mc}} function
from the \code{spdep} package developed by Roger Bivand. It computes Geary's C
statistic with permutation testing for spatial autocorrelation using the
spdep package. A permutation test for Geary's C statistic calculated by
using \code{nsim} random permutations of \code{x} for the given spatial weighting
scheme, to establish the rank of the observed statistic in relation to the
\code{nsim} simulated values.
}
\details{
If n, n1, and S0 are not provided, they are calculated based on the
\code{listw} object.
}
\examples{
\dontrun{
# Export an SFE object from an MSFE object
sfe <- getSFE(msfe, "JBO019")
# A vector of gene expression of one gene over all locations
gene_exp <- assay(sfe, "logcounts")[1,]
# A listw object containing distance-based spatial weights for neighbours
listw <- listw
# Calculate Geary's C and statistical significance for this gene
geary <- gearyGlobalCPerm(x = gene_exp, listw = listw, nsim = 999,
zero.policy = TRUE)
}

}
\seealso{
\code{\link[spdep]{geary.test}} \code{\link[spdep]{geary}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{autocorrelation}
\keyword{geary}
\keyword{listw}
\keyword{spatial}
