% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_geary.R
\name{gearyGlobalCTest}
\alias{gearyGlobalCTest}
\title{Geary's C Test for Spatial Autocorrelation}
\usage{
gearyGlobalCTest(
  x,
  listw,
  randomisation = TRUE,
  zero.policy = attr(listw, "zero.policy"),
  alternative = "greater",
  spChk = NULL,
  adjust.n = TRUE
)
}
\arguments{
\item{x}{A numeric vector the same length as the neighbours list in listw.}

\item{listw}{A \code{listw} object created, for example, by \code{nb2listw}.}

\item{randomisation}{Variance of I calculated under the assumption of
randomisation. If FALSE, normality.}

\item{zero.policy}{Default \code{attr(listw, "zero.policy")} as set when
\code{listw} was created. If attribute not set, use the global option value.
If TRUE, assign zero to the lagged value of zones without neighbours, if
FALSE, assign NA.}

\item{alternative}{A character string specifying the alternative hypothesis.
Must be one of "greater" (default), "less" or "two.sided".}

\item{spChk}{Should the data vector names be checked against the spatial
objects for identity integrity? TRUE or FALSE, default NULL to use
\code{get.spChkOption()}.}

\item{adjust.n}{Default TRUE. If FALSE, the number of observations is not
adjusted for no-neighbour observations. If TRUE, the number of observations
is adjusted.}
}
\value{
A list with class \code{htest} containing the following components:
\item{statistic}{the value of the standard deviate of Geary's C.}
\item{p.value}{the p-value of the test.}
\item{estimate}{the value of the observed Geary's C, its expectation, and
variance under the method assumption.}
\item{alternative}{a character string describing the alternative
hypothesis.}
\item{method}{a character string giving the assumption used for
calculating the standard deviate.}
\item{data.name}{a character string giving the name(s) of the data.}
}
\description{
This function is a wrapper of the \code{\link[spdep]{geary.test}} function
from the \code{spdep} package developed by Roger Bivand. It performs Geary's test
for spatial autocorrelation using a spatial weights matrix in weights list
form. The assumptions underlying the test are sensitive to the form of the
graph of neighbour relationships and other factors, and results may be
checked against those of \code{\link[spdep]{geary.mc}} permutations.
}
\details{
The derivation of the test (Cliff and Ord, 1981, p. 18) assumes
that the weights matrix is symmetric. For inherently non-symmetric matrices,
such as k-nearest neighbour matrices, \code{listw2U()} can be used to make
the matrix symmetric. In non-symmetric weights matrix cases, the variance of
the test statistic may be negative. Geary's C is affected by non-symmetric
weights under normality much more than Moran's I. From 0.4-35, the sign of
the standard deviate of C is changed to match Cliff and Ord (1973, p. 21).
}
\examples{
\dontrun{
# Export an SFE object from an MSFE object
sfe <- getSFE(msfe, "JBO019")
# A vector of gene expression of one gene over all locations
gene_exp <- assay(sfe, "logcounts")[1,]
# A listw object containing distance-based spatial weights for neighbours
listw <- listw
# Calculate Geary's C and statistical significance for this gene
geary <- gearyGlobalCTest(x = gene_exp, listw = listw, zero.policy = TRUE)
}

}
\references{
Cliff, A. D., Ord, J. K. 1981 Spatial processes, Pion, p. 21,
Cliff, A. D., Ord, J. K. 1973 Spatial Autocorrelation, Pion, pp. 15-16, 21;
Bivand RS, Wong DWS 2018 Comparing implementations of global and local
indicators of spatial association. TEST, 27(3), 716--748 \doi{10.1007/s11749
-018-0599-x}
}
\seealso{
\code{\link{geary}}, \code{\link{geary.mc}}, \code{\link{listw2U}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{autocorrelation}
\keyword{geary}
\keyword{listw}
\keyword{spatial}
