% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_utils.R
\name{get.QC.CoefficientOfVar}
\alias{get.QC.CoefficientOfVar}
\title{Calculate gene Coefficient of Variation}
\usage{
get.QC.CoefficientOfVar(sfe, assay, .sample_id)
}
\arguments{
\item{sfe}{A SpatialFeaturesExperiment object.}

\item{assay}{The name of the assay to use.}

\item{.sample_id}{The sample ID to calculate the CV for.}
}
\value{
The modified SpatialFeaturesExperiment object with added Coefficient of
Variance statistic calculated.
}
\description{
IMPORTANT: This function is not exported to be used alone.
A function to calculate the coefficient of variation (CV) for each gene.
CV for each gene is calculated as the ratio of the standard deviation to the
mean and is expressed as a percentage. A lower CV would therefore mean higher
stability.
}
\details{
This function calculates the CV for each gene in the given
\code{SpatialFeaturesExperiment} object and assigns the CV values to the rowData
of the object. The CV is calculated separately for the sample (\code{s_CV}) and
population (\code{p_CV}) data based on the provided \code{assay} and \code{.sample_id}.
}
\seealso{
\code{\link{SpatialFeatureExperiment}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
