% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.QC.ExprStats.R
\name{get.QC.ExprStats}
\alias{get.QC.ExprStats}
\title{Calculate gene expression statistics}
\usage{
get.QC.ExprStats(sfe, assay, .sample_id)
}
\arguments{
\item{sfe}{A SpatialFeaturesExperiment object.}

\item{assay}{The name of the assay to use.}

\item{.sample_id}{The sample ID to run the calculations for.}
}
\value{
The modified SpatialFeaturesExperiment object with added expression
statistics.
}
\description{
A function to calculate additional expression statistics.
}
\details{
This function calculates various expression statistics for a given sample in
a SpatialFeaturesExperiment object. The statistics include the number of
locations a gene is present, the mean and median UMI counts over the
locations, the minimum and maximum UMI counts, and the standard deviation of
UMI counts.
}
\seealso{
\code{\link{get.QC.CoefficientOfVar}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
