% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_utils.R
\name{get.QC.FindZeroExpr}
\alias{get.QC.FindZeroExpr}
\title{Find non-expressed genes}
\usage{
get.QC.FindZeroExpr(sfe, assay, .sample_id)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object.}

\item{assay}{The name of the assay to use.}

\item{.sample_id}{The sample ID to run the calculations for.}
}
\value{
The modified SpatialFeaturesExperiment object with the total UMIs
stored in the rowData.
}
\description{
A function to find zero expression values and calculate total UMIs.
}
\details{
This function takes a SpatialFeaturesExperiment object and
calculates the total number of UMIs for each gene in the specified sample.
It extracts the data for the given sample from the specified assay, finds
the total UMIs for each gene, and stores the results in the rowData of the
SpatialFeaturesExperiment object.
}
\seealso{
\code{\link{SpatialFeatureExperiment}}, \code{\link{rowData}},
\code{\link{assay}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
