% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_utils.R
\name{get.QC.Sparsity}
\alias{get.QC.Sparsity}
\title{Calculate sparsity}
\usage{
get.QC.Sparsity(sfe, assay, MARGIN, sampleNo = NULL, .sample_id = NULL)
}
\arguments{
\item{sfe}{A SpatialFeaturesExperiment object.}

\item{assay}{The name of the assay to use.}

\item{MARGIN}{Specifies the aspect for which the sparsity will be calculated.
Use 1 for features (genes) or 2 for locations.}

\item{sampleNo}{The sample number used to call the sparsity over all datasets
when multiple samples exist. This parameter is used only when the function
is called by addPerGeneQC.}

\item{.sample_id}{The sample ID to run the calculations for. This parameter
is used only when the function is called by addPerGeneQC.}
}
\description{
A function to calculate the sparsity of the dataset and get the
sparsity of each location or each gene.
}
\details{
The sparsity is calculated as the proportion of zeros in the dataset.
It provides an indication of the density of expression values. A higher
sparsity indicates a higher number of zero values, indicating low expression
levels or absence of expression.
}
\seealso{
addPerGeneQC
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
