% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getColours}
\alias{getColours}
\title{Get a vector of colours for plotting}
\usage{
getColours(number)
}
\arguments{
\item{number}{The desired number of colours.}
}
\value{
A vector of colours.
}
\description{
This function returns a vector of colours from pre-defined palettes based on
the desired number of colours.
}
\details{
This function selects colours from various pre-defined palettes to
accommodate the desired number of colours. It supports up to 167 colours. If
a number greater than 167 is provided, an error is thrown.
}
\examples{
# Get a set of 10 colours
colours <- getColours(10)
colours

}
\seealso{
\code{\link[=c4a]{c4a()}} from the \code{\link{cols4all}} package.
}
\author{
Eleftherios (Lefteris) Zormpas
}
