% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_Discrepancy.R
\name{getDiscrepancyGeneData}
\alias{getDiscrepancyGeneData}
\title{Fetch gene expression for discrepancy location}
\usage{
getDiscrepancyGeneData(
  m_sfe,
  assay,
  vars = NULL,
  focus,
  dMetric,
  sample_id,
  diam,
  mean.diff = 1,
  show.vars = c("top", "all"),
  exportExpression = FALSE
)
}
\arguments{
\item{m_sfe}{a SpatialFeatureExperiment or a MetaSpatialFeatureExperiment
data object.}

\item{assay}{Assay type for the spatial expression data (counts, logcounts,
etc.).}

\item{vars}{Variables of interest (genes) to be evaluated. Default is NULL,
which includes all variables.}

\item{focus}{Tissue locations of interest (barcodes) for which discrepancy
heatmaps will be generated.}

\item{dMetric}{Distance metric used for generating the distance matrix which
will be used to identify the focus location's neighbors.}

\item{sample_id}{Sample ID for which discrepancy data is retrieved.}

\item{mean.diff}{Threshold for selecting genes based on the difference from
the mean discrepancy score. Default is 1.}

\item{show.vars}{Display option for variables (genes) in the heatmap. Options
are "top" (genes with higher discrepancies) or "all" (all genes).}

\item{exportExpression}{Defaults to FALSE. If TRUE, it prints out a data
frame of the gene data that were used in the discrepancy heatmap for the
specific discrepancy location. The data contain a column named 'gene_name'
which includes the gene IDs of the genes. ENSGene IDs are used as row names.}

\item{bw}{Bandwidth parameter for selecting neighbors for the heatmap.}
}
\value{
If exportExpression is FALSE, the function returns a list containing the
following:
\itemize{
\item vars: Variables (genes) of interest.
\item nbrlist: List of neighbours within the specified bandwidth.
\item data.focus: Data frame containing the gene data for the discrep. heatmap.
\item annot: Data frame containing the annotation data for the genes.
If exportExpression is TRUE, the function returns a data frame containing the
gene data that were used in the discrepancy heatmap.
}
}
\description{
A function to get the required data for the genes that make the spot in
question have a high discrepancy score. This function is based on the
gw.pcplot() function from the \code{GWmodel} package.
}
\examples{
# Load data
data(sfe)

# Set parameters
assay <- "logcounts"
focus <- colnames(sfe)[1] # outlier Barcode
sample_id <- "JBO019"
dMetric <- "euclidean"
bw <- 450
show.vars <- "top"

# Get Gene expression data
discData <- getDiscrepancyGeneData(sfe, assay = assay, focus = focus,
dMetric = dMetric, sample_id = sample_id, bw = bw, show.vars = show.vars)
class(discData)
names(discData)

}
\author{
Eleftherios (Lefteris) Zormpas
}
