% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_Discrepancy.R
\name{getDiscrepancyLocData}
\alias{getDiscrepancyLocData}
\title{Get Discrepancy Data}
\usage{
getDiscrepancyLocData(m_sfe, gwpca, sample_id)
}
\arguments{
\item{m_sfe}{a SpatialFeatureExperiment or a MetaSpatialFeatureExperiment
data object.}

\item{gwpca}{GWPCA results object.}

\item{sample_id}{Sample ID for which discrepancy data is retrieved.}
}
\value{
A data frame containing the discrepancy data, including barcodes,
coordinates, discrepancy scores, and geometries.
}
\description{
Retrieves discrepancy data based on sample ID from spatial expression data
and GWPCA results.
}
\details{
This function extracts the discrepancy data from spatial expression
data and GWPCA results based on a specified sample ID. It selects the sample
locations matching the given sample ID, identifies the discrepancies using
the is_outlier information from GWPCA, and retrieves the corresponding
barcodes, coordinates, discrepancy scores, and geometries. The discrepancy
data is returned as a data frame.
}
\examples{
# Load data
data(sfe)
data(gwpca)

# Set sample name
sample_id <- "JBO019"

# Get Location data
discData <- getDiscrepancyLocData(sfe, gwpca, sample_id)
head(discData)

}
\author{
Eleftherios (Lefteris) Zormpas
}
