% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDistMat}
\alias{getDistMat}
\title{Get Distance Matrix}
\usage{
getDistMat(msfe, dMetric, sample_id = TRUE)
}
\arguments{
\item{msfe}{\code{MetaSpatialFeatureExperiment} (Meta-SFE) object.}

\item{dMetric}{Character string specifying the distance metric.}

\item{sample_id}{character string, or TRUE specifying sample/image
identifier(s); here, TRUE is equivalent to all samples/images.}
}
\value{
Distance matrix corresponding to the specified metric.
}
\description{
Retrieve the distance matrix based on the specified metric from the metadata
of a SpatialFeatureExperiment object.
}
\examples{
# Load sfe object
data(sfe)

# Add distance matrix to sfe
sfe <- addDistMat(sfe, p = 2)

# Fetch distance matrix from sfe
dMat <- getDistMat(sfe, dMetric = "euclidean")

# Check distance matrix
dMat[1:4, 1:4]

}
\author{
Eleftherios (Lefteris) Zormpas
}
