% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGradients}
\alias{getGradients}
\title{Generate a Gradient of Colours}
\usage{
getGradients(colours, steps)
}
\arguments{
\item{colours}{A character vector representing the colours to use for the gradient.}

\item{steps}{An integer specifying the number of steps in the gradient.}
}
\value{
A list of vectors, each containing the colour values representing the gradient for a single input colour.
}
\description{
This function creates a gradient of colours from a given vector of colours to white,
with a specified number of steps.
}
\examples{
getGradients(c("red", "green", "blue"), steps = 3)
getGradients(c("orange", "purple"), steps = 5)

}
