% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_getisG.R
\name{getisGlobalGTest}
\alias{getisGlobalGTest}
\title{Global G test for spatial autocorrelation}
\usage{
getisGlobalGTest(
  m_sfe,
  sample_id = NULL,
  genes = TRUE,
  zero.policy = NULL,
  alternative = "greater",
  spChk = NULL,
  adjust.n = TRUE,
  B1correct = TRUE,
  adjust.x = TRUE,
  Arc_all_x = FALSE,
  mc.cores = getOption("mc.cores", 2L)
)
}
\arguments{
\item{m_sfe}{An object of class SpatialFeatureExperiment or
MetaSpatialFeatureExperiment.}

\item{sample_id}{A character vector specifying the sample IDs to include
(only relevant if a MetaSpatialFeatureExperiment has been provided in the
\code{m_sfe} argument).}

\item{genes}{TRUE or a named character vector with gene names for which the
SA statistic needs to be calculated. If left to TRUE then the SA statistic
is calculated for every gene.}

\item{zero.policy}{Default is NULL. If not changed then internally, it is
set to \code{attr(listw, "zero.policy")} as set when \code{listw} was
created. If TRUE, assign zero to the lagged value of zones without
neighbours. If FALSE, assign NA. If attribute not set, use the global option
value.}

\item{alternative}{A character string specifying the
alternative hypothesis, must be one of "greater"
(default), "less" or "two.sided".}

\item{spChk}{Should the data vector names be checked
against the spatial objects for identity integrity,
TRUE, or FALSE, default NULL to use
\code{get.spChkOption()}.}

\item{adjust.n}{Default TRUE, if FALSE the number of
observations is not adjusted for no-neighbour
observations, if TRUE, the number of observations is
adjusted.}

\item{B1correct}{Default TRUE, if TRUE, the erratum
referenced below: "On page 195, the coefficient of W2
in B1, (just below centre of the page) should be 6,
not 3." is applied; if FALSE, 3 is used (as in CrimeStat
IV).}

\item{adjust.x}{Default TRUE, if TRUE, x values of
observations with no neighbours are omitted in the
denominator of G.}

\item{Arc_all_x}{Default FALSE, if Arc_all_x=TRUE and
adjust.x=TRUE, use the full x vector in part of the
denominator term for G.}

\item{mc.cores}{Argument from \code{link[parallel]{mclapply}}. The number of
cores to use, i.e., at most how many child processes will be run
simultaneously. The option is initialized from environment variable MC_CORES
if set. Must be at least one, and parallelisation requires at least two
cores.}
}
\value{
An SFE object with the below two columns added in the \code{rowData}:
\item{getisG_test}{the value of the standard deviate
of Getis-Ord G.}
\item{getisG_PvalTest}{the p-value of the test.}
}
\description{
This function is a wrapper of the \code{\link[spdep]{moran}} function from
the \code{spdep} package developed by Roger Bivand. The global G statistic for
spatial autocorrelation, complementing the local Gi LISA measures:
\code{\link{localG}}.
}
\examples{
\dontrun{
# Example usage:

# Export an SFE object from an MSFE object
sfe <- getSFE(msfe, "JBO019")

# A vector of gene expression of one gene over all locations
gene_exp <- assay(sfe, "logcounts")[1,]

# A listw object containing distance-based spatial weights for neighbours
listw <- listw

# Calculate Moran's I for this gene
getis <- getisGlobalGTest(x = gene_exp, listw = listw,zero.policy = TRUE)
}

}
\references{
Getis. A, Ord, J. K. 1992 The analysis of spatial association by
use of distance statistics, \emph{Geographical Analysis}, 24, p. 195; see
also Getis. A, Ord, J. K. 1993 Erratum, \emph{Geographical Analysis}, 25,
p. 276; Bivand RS, Wong DWS 2018 Comparing implementations of global and
local indicators of spatial association. TEST, 27(3), 716--748 \doi{10.1007/
s11749-018-0599-x}
}
\seealso{
\code{\link{localG}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
