% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_getisG.R
\name{getisLocalG}
\alias{getisLocalG}
\alias{getisLocalGPerm}
\title{G and Gstar local spatial statistics}
\usage{
getisLocalG(
  x,
  listw,
  zero.policy = NULL,
  spChk = NULL,
  GeoDa = FALSE,
  alternative = "two.sided"
)

getisLocalGPerm(
  x,
  listw,
  nsim = 499,
  zero.policy = NULL,
  spChk = NULL,
  alternative = "two.sided",
  iseed = NULL,
  fix_i_in_Gstar_permutations = TRUE,
  no_repeat_in_row = FALSE
)
}
\arguments{
\item{x}{Numeric vector the same length as the neighbors
list in listw}

\item{listw}{A \code{listw} object created, for example,
by \code{nb2listw}}

\item{zero.policy}{Default NULL, use global option value;
if TRUE assign zero to the lagged value of zones without
neighbors, if FALSE assign NA}

\item{spChk}{Should the data vector names be checked
against the spatial objects for identity integrity,
TRUE, or FALSE, default NULL to use \code{get.spChkOption()}}

\item{GeoDa}{Default FALSE, if TRUE, drop x values for
no-neighbour and self-neighbour only observations from all
summations}

\item{alternative}{A character string specifying the
alternative hypothesis, must be one of \code{"two.sided"}
(default), \code{"greater"} or \code{"less"}.}

\item{nsim}{Default 499, number of conditional permutation
simulations}

\item{iseed}{Default NULL, used to set the seed for
possible parallel RNGs}

\item{fix_i_in_Gstar_permutations}{Default \code{TRUE}
(fix x at self in permutations for local G-star), set
\code{FALSE} to use pre-1.2-8 behaviour}

\item{no_repeat_in_row}{Default \code{FALSE}, if
\code{TRUE}, sample conditionally in each row without
replacements to avoid duplicate values,
\url{https://github.com/r-spatial/spdep/issues/124}}
}
\value{
A vector of G or Gstar standard deviate values,
with attributes "gstari" set to TRUE or FALSE, "call"
set to the function call, and class "localG". For
conditional permutation, the returned value is the same
as for \code{localG()}, and the simulated standard deviate
is returned as column \code{"StdDev.Gi"} in
\code{attr(., "internals")}.
}
\description{
The local spatial statistic G is calculated for each
zone based on the spatial weights object used. The value
returned is a Z-value, and may be used as a diagnostic
tool. High positive values indicate the posibility of a
local cluster of high values of the variable being
analysed, very low relative values a similar cluster of
low values. For inference, a Bonferroni-type test is
suggested in the references, where tables of critical
values may be found (see also details below).
}
\details{
If the neighbours member of listw has a
"self.included" attribute set to TRUE, the Gstar variant,
including the self-weight \eqn{w_{ii} > 0}, is calculated
and returned. The returned vector will have a "gstari"
attribute set to TRUE. Self-weights must be included by
using the \code{include.self} function before converting
the neighbor list to a spatial weights list with
\code{nb2listw} as shown below in the example.

The critical values of the statistic under assumptions
given in the references for the 95th percentile are for
n=1: 1.645, n=50: 3.083, n=100: 3.289, n=1000: 3.886.
}
\note{
Conditional permutations added for comparative
purposes; permutations are over the whole data vector
omitting the observation itself, and from 1.2-8 fixing the
observation itself as its own neighbour for local G-star.
}
\examples{
\dontrun{
# Example usage:

# Export an SFE object from an MSFE object
sfe <- getSFE(msfe, "JBO019")

# A vector of gene expression of one gene over all locations
gene_exp <- assay(sfe, "logcounts")[1,]

# A listw object containing distance-based spatial weights for neighbours
listw <- listw

# Calculate Getis and Ord's G for this gene
getis <- getisLocalG(x = gene_exp, listw = listw,zero.policy = TRUE)
}

\dontrun{
# Calculate Getis and Ord's G for this gene
getis <- getisLocalGPerm(x = gene_exp, listw = listw, nsim = 499,
zero.policy = TRUE)
}

}
\references{
Ord, J. K. and Getis, A. 1995 Local spatial
autocorrelation statistics: distributional issues and an
application. \emph{Geographical Analysis}, 27, 286--306;
Getis, A. and Ord, J. K. 1996 Local spatial statistics: an
overview. In P. Longley and M. Batty (eds) \emph{Spatial
analysis: modelling in a GIS environment} (Cambridge:
Geoinformation International), 261--277; Bivand RS, Wong
DWS 2018 Comparing implementations of global and local
indicators of spatial association.
TEST, 27(3), 716--748 \doi{10.1007/s11749-018-0599-x}
}
\seealso{
\code{\link{localG}}

\code{\link{localG}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
