% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_FunctionalClust.R
\name{getMSigDBData}
\alias{getMSigDBData}
\title{Get MSigDB Data}
\usage{
getMSigDBData(species, ...)
}
\arguments{
\item{species}{A character string specifying the target species
(e.g., "Homo sapiens").}

\item{...}{Other arguments passed onto the \code{msigdbr} function.}
}
\value{
A data table with the MSigDB data.
}
\description{
This function retrieves data from MSigDB for a specified species.
}
\examples{
# Download data from MSigDB
msigdb <- getMSigDBData("Homo sapiens")

}
