% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metaSFE.R
\name{getMultiSFE}
\alias{getMultiSFE}
\title{Method to retrieve multiple SFE objects as a single SFE object from within
an MSFE object}
\usage{
getMultiSFE(x, sample_ids)
}
\arguments{
\item{x}{An instance of the MetaSpatialFeatureExperiment class.}

\item{sample_ids}{Character string. The sample IDs from the SFE objects to
be combined and retrieved as one SFE object}
}
\value{
An SFE object with multiple samples inside.
}
\description{
Method to retrieve multiple SFE objects as a single SFE object from within
an MSFE object
}
