% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_utils.R
\name{getSAGlobalGenes}
\alias{getSAGlobalGenes}
\title{Get Spatial Autocorrelation Global Genes}
\usage{
getSAGlobalGenes(
  m_sfe = sfe,
  statistic = c("moran", "geary", "getis"),
  test = c("z-score", "permutation"),
  pVal = 0.05,
  stat_thresh = 0.5
)
}
\arguments{
\item{m_sfe}{An object of class 'SpatialFeatureExperiment' or
'MultiSpatialFeatureExperiment' containing spatial data.}

\item{statistic}{A character string specifying the type of spatial
autocorrelation results ("moran", "geary", "getis") to look
for in the m_sfe object for the specified sample.}

\item{test}{A character string specifying the statistical test method
("z-score" or "permutation") to look for in the m_sfe object
for the specified sample.}

\item{pVal}{A numeric value representing the significance threshold for
plotting. Ignored for Geary's C, where permutation is suggested.
See details for more information.}

\item{stat_thresh}{Numeric value specifying the threshold for the spatial
autocorrelation statistic. Genes with an absolute value
of the statistic greater than this threshold will be
included. Default is 0.5.}
}
\value{
A named vector of gene names that pass the specified thresholds.
}
\description{
This function calculates spatial autocorrelation statistics and returns a
vector of genes that pass specified thresholds.
}
\examples{
getSAGlobalGenes()

}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{autocorrelation}
\keyword{spatial,}
