% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metaSFE.R
\name{getSampleIDs}
\alias{getSampleIDs}
\title{Method to retrieve sample IDs from MetaSpatialFeatureExperiment}
\usage{
getSampleIDs(x)
}
\arguments{
\item{x}{An instance of the MetaSpatialFeatureExperiment class.}
}
\value{
A character vector containing the sample IDs.
}
\description{
This method retrieves the sample IDs from a MetaSpatialFeatureExperiment object.
The sample IDs are stored in the @sample_ids slot of the MetaSpatialFeatureExperiment.
}
\examples{
# Create a MetaSpatialFeatureExperiment object
msfe <- MetaSpatialFeatureExperiment()

# Create or load a list of SpatialFeatureExperiment objects
# Replace sfe1, sfe2, sfe3 with your actual SFE objects
# sfe_list <- list(sfe1, sfe2, sfe3)

# Add multiple SFE objects to the MetaSpatialFeatureExperiment
# msfe <- addMultipleSFEs(msfe, sfe_list)

# Retrieve sample IDs
sample_ids <- getSampleIDs(msfe)

}
\seealso{
\code{\link{MetaSpatialFeatureExperiment}}, \code{\link{addMultipleSFEs}}
}
\keyword{internal}
