% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metaSFE.R
\name{getSFE}
\alias{getSFE}
\title{Method to access a SpatialFeatureExperiment by name}
\usage{
getSFE(x, name)
}
\arguments{
\item{x}{An instance of the MetaSpatialFeatureExperiment class.}

\item{name}{Name of the experiment/sample for the SFE to access.}
}
\value{
The SFE object associated with the provided name.
}
\description{
Method to access a SpatialFeatureExperiment by name
}
