% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_getSubset.R
\name{getSubset}
\alias{getSubset}
\title{Get Subset of Features from SpatialFeatureExperiment or
MetaSpatialFeatureExperiment}
\usage{
getSubset(
  m_sfe,
  sample_id,
  subset,
  set = c("colData", "rowData"),
  col_name = NULL
)
}
\arguments{
\item{m_sfe}{An object of class SpatialFeatureExperiment or
MetaSpatialFeatureExperiment.}

\item{sample_id}{A character vector specifying the sample IDs to include
(TRUE for all IDs in m_sfe).}

\item{subset}{A regular expression string to be used for subsetting.}

\item{set}{Either "rowData" or "colData" specifying whether to operate on
row or column data.}

\item{col_name}{The name of the column to be used for subsetting if set to
"rowData" or "colData".}
}
\value{
If m_sfe is a SpatialFeatureExperiment (SFE) object, it returns
a logical vector indicating the subset of features based on the criteria.
If m_sfe is a MetaSpatialFeatureExperiment (MetaSFE) object, it returns
a list where each element is a logical vector indicating the subset of
samples for each SFE object within the MetaSFE.
}
\description{
This function retrieves a subset of features based on the specified criteria
from either a SpatialFeatureExperiment (SFE) or a
MetaSpatialFeatureExperiment (MetaSFE) object.
}
\examples{
\dontrun{
# Example usage:
# For SFE object
sfe <- SpatialFeatureExperiment()
getSubset(sfe, sample_id = TRUE, subset = "(^MT-)|(^mt-)", set = "rowData", col_name = "gene_name")

# For MetaSFE object
metaSFE <- MetaSpatialFeatureExperiment(list(sfe1, sfe2))
getSubset(metaSFE, sample_id = TRUE, subset = "(^MT-)|(^mt-)", set = "colData", col_name = "sample_id")
}

}
