% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_FunctionalClust.R
\name{getTerm2Gene}
\alias{getTerm2Gene}
\title{Get Term-to-Gene Data}
\usage{
getTerm2Gene(user_data = NULL, msig_data = NULL, cat = NULL, subcat = NULL)
}
\arguments{
\item{user_data}{A user-supplied term-to-gene data frame (optional).}

\item{msig_data}{A term-to-gene data frame as generated using the
\code{getMSigDBData} function.}

\item{cat}{Filter for "gs_cat" (e.g., "C2", "C3").}

\item{subcat}{Filter for "gs_subcat" (e.g., "GO:BP", "CGP").}
}
\value{
A term-to-gene data frame with columns "term" and "gene."
}
\description{
This function retrieves and prepares a term-to-gene data frame based on user
choices.
}
\examples{
# Get MSigDB data
data <- getMSigDBData("Homo sapiens")

# View categories and subcategories
viewCollections()

# Get term-to-gene data.table
t2g <- getTerm2Gene(msig_data = data, cat = C1, subcat = "")

# Note in the above that if a category has no subcategory then we put "" in
# the 'subcat' argument.

#' @seealso
\code{\link[clusterProfiler]{GSEA}}

}
