% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopHighVarGenes.R
\name{getTopHighVarGenes}
\alias{getTopHighVarGenes}
\title{Get Top High Variable Genes from a Spatial Transcriptomics Experiment}
\usage{
getTopHighVarGenes(stats, sample_id = TRUE, ...)
}
\arguments{
\item{stats}{A \code{DataFrame} of variance modelling statistics with one
row per gene. Alternatively, a SpatialFeatureExperiment object, in which
case it is supplied to modelGeneVar to generate the required DataFrame.}

\item{sample_id}{TRUE selects all samples. Alternatively, provide a
character vector with the sample IDs for filtering.}

\item{...}{Additional arguments to be passed to \code{scran::getTopHVGs}.}
}
\value{
A list of top high variance genes for each sample, as determined by
the provided DE analysis results.
}
\description{
This function retrieves the top high variance genes from a spatial
transcriptomics experiment based on a provided DataFrame of variance
modelling statistics.
}
\details{
The function uses a DE analysis result list (\code{stats}) to extract
the top high variance genes for each sample. It allows for optional
filtering based on sample IDs. Additional arguments can be passed to
the underlying \code{scran::getTopHVGs} function.
}
\examples{
\dontrun{
# Get top high variance genes from a spatial transcriptomics experiment
# 'stats' a DataFrame of variance modelling statistics
top_hvg <- getTopHighVarGenes(stats, sample_id = TRUE)
}

}
\seealso{
\code{\link[=getTopHVGs]{getTopHVGs()}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{experiment}
\keyword{genes,}
\keyword{high}
\keyword{spatial}
\keyword{transcriptomics}
\keyword{variance}
