% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gTruth}
\alias{gTruth}
\title{10X Visium Steatotic Human Liver - annotation}
\format{
\subsection{\code{matrix}}{

A 3-column matrix with annotation data as annotated by the Liver Cell Atlas
project. Includes the below columns:
\itemize{
\item Barcode: Spot barcode.
\item sample_id: Sample ID.
\item annotation: The manual annotation of each spot.
}
Each column is \code{character}.
}
}
\source{
\url{https://www.livercellatlas.org/download.php}
}
\usage{
data(gTruth)
}
\description{
Annotation data from the JBO019 Human Liver dataset.
}
\keyword{datasets}
