% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwpca}
\alias{gwpca}
\title{Geographically Weighted Principal Components Analysis - GWPCA}
\format{
\subsection{\code{gwpca}}{

A list object of class \code{gwpca}. Includes all the data generated from running
the \code{\link[=gwpcaSTE]{gwpcaSTE()}} function. Specifically includes:
\itemize{
\item pca: global PCA results.
\item loadings: loading scores of each genes in each Principal Component
(PC) in each location.
\item SDF a \code{SpatialPointsDataFrame} with Percentage of Variation
explained by each PC in each location.
\item gwpca.scores: PCA scores for each component in each location.
\item var:
\item local.PV: Percent of Variation in each location.
\item GW.arguments: The arguments used to run \code{\link[=gwpcaSTE]{gwpcaSTE()}}.
\item CV: Cross Validation results.
\item geometry: Hexagonal geometries for each location. Used for plotting.
}

The arguments used for this dataset are the below:
\itemize{
\item vars: 526 genes vector as generated using the \code{\link[scran:getTopHVGs]{scran::getTopHVGs()}}
function after running the \code{\link[scran:modelGeneVar]{scran::modelGeneVar()}} function on the
preprocessed SFE object from the JBO019 sample. For the SFE object and
the parameters used in the QC see also the \code{?STExplorerDev::sfe} help
page. The parameters used in \code{\link[scran:getTopHVGs]{scran::getTopHVGs()}} are the below:
\itemize{
\item var.field: "bio"
\item prop: 0.5
\item fdr.threshold: 0.05
}
\item k: 20
\item kernel: "gaussian"
\item adaptive: FALSE
\item p: 2
\item cv: TRUE
\item scores: FALSE
\item robust: FALSE
\item sfe: sfe object preprocessed and QC'ed.
\item assay: "logcounts"
\item bw: 882.09951902194
}
}
}
\usage{
data(gwpca)
}
\description{
Results from running the \code{\link[=gwpcaSTE]{gwpcaSTE()}} function on the processed and QC'ed
10X Visium Steatotic Human Liver sample JBO019 from the Liver Cell Atlas
\url{https://www.livercellatlas.org/download.php}.
}
\keyword{datasets}
