% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_FunctionalClust.R
\name{gwpca_FunctionalClustering}
\alias{gwpca_FunctionalClustering}
\title{Run Gene Set Enrichment Analysis (GSEA) per location using GWPCA}
\usage{
gwpca_FunctionalClustering(
  gwpca,
  pc,
  genes_no = 1,
  NES = 1.5,
  minGSSize = 5,
  pvalueCutoff = 0.25,
  TERM2GENE,
  pAdjustMethod = "fdr",
  scoreType = "std",
  nPermSimple = 10000,
  mc.cores = 4,
  regex = "",
  ...
)
}
\arguments{
\item{gwpca}{A GWPCA object containing the results of the spatial
transcriptomics analysis.}

\item{pc}{The principal component (PC) index to be used for GSEA.}

\item{genes_no}{The minimum number of genes in a gene set to be considered
for enrichment.}

\item{NES}{The minimum Normalized Enrichment Score (NES) for considering a
gene set as enriched.}

\item{minGSSize}{The minimum gene set size to be considered in GSEA.}

\item{pvalueCutoff}{The p-value cutoff for identifying significant gene sets
in GSEA.}

\item{TERM2GENE}{The term-to-gene mapping for gene sets.}

\item{pAdjustMethod}{The method for multiple testing correction in GSEA
(default is "fdr").}

\item{scoreType}{The GSEA scoring type (default is "std").}

\item{nPermSimple}{The number of permutations for simple GSEA.}

\item{mc.cores}{The number of cores to use for parallel processing.}

\item{regex}{A regular expression pattern for cluster identification.}

\item{...}{Additional arguments to be passed to internal functions.}
}
\value{
A tibble containing the GSEA results for each spatial location.
}
\description{
This function performs GSEA on multiple cores to analyze the enrichment of
gene sets in different spatial locations based on Gene Set Enrichment
Analysis. It takes the output of a GWPCA analysis, extracts the necessary
data for each spatial location, and applies GSEA. The results are combined
into a final dataframe, and additional filtering and processing are applied.
}
\details{
Detailed explanation of the function, including any relevant details about the
algorithm used, data processing steps, and result interpretation.
}
\examples{
## Example usage:
# gwpca_result <- gwpcaSTE()  # run the gwpcaSTE function

## Functional clustering on Principal Component 1 (PC1)
# gsea_result <- gwpca_FunctionalClustering(gwpca_result, pc = 1)

}
\seealso{
\code{\link{gwpcaSTE}}, \code{\link[clusterProfiler]{GSEA}}

Other Spatial Transcriptomics Analysis: 
\code{\link{.int_runGSEA}()},
\code{\link{plotGWPCA_FuncCLust}()}
}
\concept{Spatial Transcriptomics Analysis}
\keyword{GSEA}
\keyword{GWPCA}
\keyword{analysis}
\keyword{enrichment}
\keyword{gene}
\keyword{set}
\keyword{spatial}
\keyword{transcriptomics}
