% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_LeadingGene.R
\name{gwpca_LeadingGene}
\alias{gwpca_LeadingGene}
\title{Find Leading genes}
\usage{
gwpca_LeadingGene(
  gwpca,
  m_sfe,
  sample_id,
  pc_nos,
  genes_n = 5,
  type = c("single", "multi"),
  method = c("membership", "order"),
  names = c("id", "gene_names")
)
}
\arguments{
\item{gwpca}{A list of class \code{gwpca} containing the geographically
weighted principal component analysis results.}

\item{m_sfe}{A \code{SpatialFeatureExperiment} or a
\code{MetaSpatialFeatureExperiment} object with the spatial
feature experiment data.}

\item{sample_id}{The sample ID (or the SFE object name inside a Meta-SFE
object) for which the GWPCA was run.}

\item{pc_nos}{A vector containing the principal component numbers for which
you want to find the leading genes.}

\item{genes_n}{An integer indicating how many genes you want to include in
the top leading genes. This argument is used only when the type parameter is
set to 'multi'. The default value is 5.}

\item{type}{A character indicating the type of leading genes to be
identified. Possible values are 'single' for a single leading gene per
location or 'multi' for multiple leading genes per location.}

\item{method}{A character indicating the method to be used for grouping the
spots together. Possible values are "membership" or "order".}

\item{names}{A character indicating the output format for gene identifiers.
Possible values are "id" for ENSG IDs or "gene_names" for gene names.}
}
\value{
A modified \code{gwpca} object with additional information on the
leading genes. If the type parameter is set to 'single', the modified object
contains a new slot leadingGeneSingle with the leading genes for each
principal component. If the type parameter is set to 'multi', the modified
object contains a new slot leadingGeneMulti with the top leading genes
(specified by genes_n) for each principal component. The leading genes are
represented as a data frame with columns corresponding to principal component
numbers and rows corresponding to locations. The data frame also includes the
geometry information of the locations.
}
\description{
This function identifies the leading genes in a location for one or multiple
principal components.
}
\details{
The function calculates the leading genes in a location based on the
specified principal component numbers. It returns the results as a list of
leading genes for each principal component. If the type parameter is set to
'single', it returns a single leading gene for each location and principal
component. If the type parameter is set to 'multi', it returns multiple
leading genes (specified by genes_n) for each location and principal
component.
}
\examples{
# Load Geographically Weighted principal component analysis data
data(gwpca)
data(sfe)

#Identify single leading gene per location for PC1 and PC2
gwpca <- gwpca_LeadingGene(gwpca = gwpca, sfe = sfe, pc_nos = c(1, 2),
type = "single", names = "id")

# Identify top 3 leading genes per location for PC1 and PC2
gwpca <- gwpca_LeadingGene(gwpca = gwpca, sfe = sfe, pc_nos = c(1, 2),
type = "multi", genes_n = 3, method = "membership", names = "gene_names")

}
\author{
Eleftherios (Lefteris) Zormpas
}
