% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_PropVar.R
\name{gwpca_propVar}
\alias{gwpca_propVar}
\alias{gwpca_PropVar}
\title{Calculate Proportional Variance Explained by Principal Components}
\usage{
gwpca_PropVar(gwpca, n_comp, sfe, sample_id = NULL)
}
\arguments{
\item{gwpca}{A \code{gwpca} class object containing the result of the
Geographically Weighted Principal Component Analysis. As generated by
\code{gwpca.ste}}

\item{n_comp}{A numeric vector specifying the number of principal components
for which to calculate the proportional variance.}

\item{sample_id}{In the case of a Meta-SFE object, sample_id argument needs
to be defined with the ID of the sample to be used. Defaults to NULL.}

\item{m_sfe}{A \code{SpatialFeatureExperiment} (SFE) or a
\code{MetaSpatialFeatureExperiment} (Meta-SFE) object used to extract the
spot geometries.}
}
\value{
A data frame with columns for the proportional variance explained by
each principal component and the corresponding spot geometries from SFE's
\code{spotHex}.
}
\description{
Calculates the proportional variance explained by the specified number of
principal components in a gwpca object's variance matrix. The function
returns a data frame with the proportional variance for each component on
each location and the corresponding spot geometries.
}
\examples{
# Load data
data(sfe)
data(gwpca)

# Calculate percentage of Variation
gwpca <- gwpca_PropVar(gwpca, n_comp = 2:10, sfe = sfe)

}
\author{
Eleftherios (Lefteris) Zormpas
}
