% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_STE.R
\name{gwr_stats}
\alias{gwr_stats}
\title{GWR summary table}
\usage{
gwr_stats(gwr)
}
\arguments{
\item{gwr}{A GWR object as created by \code{gwrSTE}.}
}
\value{
A summary table containing descriptive statistics for the GWR object.
}
\description{
This function generates a summary table for a Geographically Weighted
Regression (GWR) object.
}
\author{
Eleftherios Zormpas
}
