% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_STE.R
\name{gwr_toSF}
\alias{gwr_toSF}
\title{Extract SDF from gwr object}
\usage{
gwr_toSF(gwr)
}
\arguments{
\item{gwr}{A GWR object as created by \code{gwrSTE}.}
}
\value{
A data frame of class sf.
}
\description{
This function extracts the SDF data frame from within a gwr class object and
converts it to class sf.
}
\author{
Eleftherios (Lefteris) Zormpas
}
