% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_STE.R
\name{int.gwpca.FLoop}
\alias{int.gwpca.FLoop}
\title{GWPCA calculating for loop}
\usage{
int.gwpca.FLoop(
  i,
  ..pcafun,
  ..x,
  ..dMat,
  ..k,
  ..bw,
  ..adaptive,
  ..kernel,
  ..scores
)
}
\arguments{
\item{i}{The index of the current location.}

\item{..pcafun}{The PCA function to use.}

\item{..x}{The data matrix.}

\item{..dMat}{The distance matrix.}

\item{..k}{The number of components to retain.}

\item{..bw}{The bandwidth parameter.}

\item{..adaptive}{Logical indicating if adaptive bandwidth should be used.}

\item{..kernel}{The kernel function to use for weighting.}

\item{..scores}{Logical indicating if local scores should be calculated.}
}
\value{
The GWPCA outcome as a list.
}
\description{
IMPORTANT: this function is not exported to be used alone.
Internal function used in the Geographically Weighted Principal Component
Analysis (GWPCA).
}
\details{
This internal function calculates the GWPCA for a single location (index i)
using the specified PCA function. It selects the weights for the location
based on the distance matrix and bandwidth parameter. The function skips the
calculation if the bandwidth is too small (less than or equal to 5). If the
calculation proceeds, it runs the PCA function on the subset of data with
non-zero weights, and stores the resulting weights and scores (if specified).
}
\seealso{
\code{\link{gwpca}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
