% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_STE.R
\name{int.gwpca}
\alias{int.gwpca}
\title{GWPCA core}
\usage{
int.gwpca(
  .sfe,
  .assay,
  .elocat,
  .vars,
  .p,
  .k,
  .bw,
  .type,
  .kernel,
  .adaptive,
  .scores,
  .robust,
  .cv,
  .verbose
)
}
\arguments{
\item{.sfe}{The spatial feature experiment object.}

\item{.assay}{The name of the assay to use from the object.}

\item{.elocat}{The evaluation locations.}

\item{.vars}{The variables to include in the analysis.}

\item{.p}{The power parameter for distance calculation.}

\item{.k}{The number of components to retain.}

\item{.bw}{The bandwidth parameter.}

\item{.type}{hex or spot geometries.}

\item{.kernel}{The kernel function to use for weighting.}

\item{.adaptive}{Logical indicating if adaptive bandwidth should be used.}

\item{.scores}{Logical indicating if local scores should be calculated.}

\item{.robust}{Logical indicating if robust WPCA should be used.}

\item{.cv}{Logical indicating if cross-validation should be performed.}

\item{.verbose}{Logical indicating if verbose output should be displayed.}
}
\value{
A list of class "gwpca".
}
\description{
IMPORTANT: this function is not exported to be used alone.
Internal function used in the Geographically Weighted Principal Component
Analysis (GWPCA).
}
\details{
This internal function performs the GWPCA analysis using the specified
parameters. It takes a spatial feature experiment object (.sfe) and extracts
the necessary data. The evaluation locations (.elocat) are checked, and if
not provided or in an invalid format, the default locations from the spatial
feature experiment object are used. The function also performs checks on the
variables, bandwidth, and other input parameters. It generates the distance
matrix and selects the subset of data for analysis based on the variables.
The global PCA is run using the selected data. The memory is pre-allocated
for storing the results of the loop. The type of WPCA (robust or non-robust)
is determined based on the input parameter. The GWPCA is then performed
either in parallel or in a verbose mode. The results are stored in the
appropriate arrays and lists. Cross-validation is performed if requested, and
the results are added to the output. The influence of weights or variances in
the data is adjusted if not using robust WPCA. The final output includes the
PCA results, loadings, spatial points data frame (SDF), GWPCA scores
(if calculated), variance, local percent of variation, input arguments,
cross-validation results, and geometry information.
}
\seealso{
\code{\link{gwpcaSTE}}, and the original function from GWmodel package;
\code{\link[=gwpca]{gwpca()}}.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
