% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_LeadingGene.R
\name{int_LeadingGene_multi}
\alias{int_LeadingGene_multi}
\title{Find the top-k leading genes}
\usage{
int_LeadingGene_multi(gwpca, pc_no, genes_n, sfe, method)
}
\arguments{
\item{gwpca}{a list of class \code{gwpca}.}

\item{pc_no}{a numeric value for the principal component numbers for which
you want to find the leading genes.}

\item{genes_n}{an integer indicating how many genes you want to be included.}

\item{sfe}{a \code{SpatialFeatureExperiment} object.}

\item{method}{takes values either "membership" or "order". Is the method to
be used for grouping the spots together.}
}
\description{
A function to find the top k leading genes per principal components in a
location.
}
\keyword{internal}
