% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_utils.R
\name{int_spotHex_gen}
\alias{int_spotHex_gen}
\alias{.int_spotHex_gen}
\title{Internal: generate perimeter spots}
\usage{
.int_spotHex_gen(subsetList, sp_diam, flipped)
}
\arguments{
\item{subsetList}{A list containing subsets of spot data frames.}

\item{.sp_diam}{The diameter of the spots.}
}
\value{
A data frame containing the modified spot data frames for the
perimeter spots.
}
\description{
IMPORTANT: this function is not exported to be used alone.
This function generates the perimeter spots only where they are needed based
on the provided subset list and spot diameter.
}
\details{
This function is used internally to generate the perimeter spots only for the
locations where they are needed. It takes a subset list containing subsets of
spot data frames and modifies them to create the perimeter spots. The
modification includes adding a ".perim" suffix to the Barcode column, setting
the Section value to 0, and adjusting the Image_X and Image_Y coordinates
based on the subset. The modified spot data frames are then combined into a
single data frame and returned.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
