% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_visium_msfe}
\alias{load_visium_msfe}
\title{Load and Process Visium SpatialFeatureExperiment (SFE) Data}
\usage{
load_visium_msfe(dataset)
}
\arguments{
\item{dataset}{A character string specifying the dataset to load. Must be
either \code{"lung"} or \code{"prostate"}. Based on this input, the function loads
the corresponding SFE object and processes the associated data.}
}
\value{
Returns the updated SFE object after processing, with images properly
reloaded and associated with the corresponding samples.
}
\description{
This function loads and processes SpatialFeatureExperiment (SFE) data from
within the package's \code{extdata} folder. It addresses the issue of missing
image data when using the \code{usethis::use_data} function to store SFE objects
in \code{.rda} format. The function ensures that the data is loaded correctly and
that the images are reloaded from the external data located in the package's
\verb{inst/extdata/visium/} directory.
}
\details{
Depending on the dataset specified, the function loads relevant data objects
(\code{msfe_2} for lung data or \code{msfe} for prostate data), processes each sample
by adding the corresponding low-resolution spatial tissue images, and applies
the appropriate scale factors. It also mirrors the images as part of the
processing.

This function is essential for working with spatial data, as it reconstructs
the image pointers (which are lost when saving SFE objects via
\code{usethis::use_data}) and ensures the correct association between images and
sample data.

For the \code{"lung"} dataset, the function loads the \code{msfe_2} object and
processes samples located in the \verb{inst/extdata/visium/lung/Fibrotic/} and
\verb{inst/extdata/visium/lung/Healthy/} directories. For the \code{"prostate"}
dataset, it loads the \code{msfe} object and processes the sample located in
\verb{inst/extdata/visium/prostate/H2_5/}.

The function retrieves scale factors from the \code{scalefactors_json.json} file
within each sample's directory and adds the corresponding low-resolution
tissue image to the SFE object. It also mirrors the image using the
\code{SpatialFeatureExperiment::mirrorImg} function.
}
\author{
Eleftherios (Lefteris) Zormpas
}
