% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{makeClusterGWPCA}
\alias{makeClusterGWPCA}
\title{Create a Parallel Cluster for GWPCA Computation}
\usage{
makeClusterGWPCA(spec, type = "FORK", ...)
}
\arguments{
\item{spec}{The number of clusters to be created. If not provided and type
is "FORK", it defaults to availableCores() - 1.}

\item{type}{The type of cluster to be created. Default is "FORK". Other
options may include "PSOCK" or others supported by parallel::makeCluster.}

\item{...}{Additional arguments to be passed to parallel::makeCluster.}
}
\value{
A parallel cluster for GWPCA computation.
}
\description{
This function creates a parallel cluster for running the GWPCA
(Geographically Weighted Principal Component Analysis) computation in
parallel on multiple cores.
}
\details{
The function sets up a parallel cluster for running the GWPCA
computation. The number of clusters can be specified, and the default is to
use availableCores() - 1 if none is provided.
}
\examples{
# Create a parallel cluster for GWPCA computation
cluster <- makeClusterGWPCA(spec = 4, type = "PSOCK")

}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{GWPCA}
\keyword{cluster,}
\keyword{computing,}
\keyword{parallel}
