% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers}
\alias{markers}
\title{Marker genes for IPF lung cell types}
\format{
\subsection{\code{markers}}{

A dataframe with cell-type markers containing the below columns
\describe{
\item{gene.name}{List of 1: contains the SFE object named "Fibrotic"}
\item{ensg.ID}{List of 1: contains the SFE object's name}
\item{Type}{High level classes: Immune, Epithelieal, Mesenchymal, and
Endothelial}
\item{Subtype}{Low level classification of cells (i.e., HAS1-high
Fibroblasts)}
\item{TypeClass}{Mid level classification of cells (i.e., HAS1-high
Fibroblasts and Fibroblasts both belong to Mid-level class Fibroblasts)}
\item{MarkerInfo}{markers are divided in two categories: positive and
negative}
\item{ExpressionLevelInfo}{Some markers are marked as high or low
expression levels}
\item{description}{The gene's full name}
}
No preprocessing has been performed.
}
}
\source{
\url{https://doi.org/10.1126/sciadv.aba1972}
}
\usage{
markers
}
\description{
The information comes from the Habermann et. al., 2020, Sci. Adv.
publication \url{https://doi.org/10.1126/sciadv.aba1972}. The information can be
found in Supplementary Table 3 (Table S3).
}
\keyword{datasets}
