% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metaSFE.R
\docType{class}
\name{MetaSpatialFeatureExperiment-class}
\alias{MetaSpatialFeatureExperiment-class}
\title{Define the MetaSpatialFeatureExperiment class using S4 classes and methods}
\description{
This class represents a collection of SpatialFeatureExperiment (SFE) objects
for spatial transcriptomics data. Each SFE object is stored in a slot within
the MetaSpatialFeatureExperiment object, and the slot names correspond to
the experiment/sample names.
}
