% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metaSFE.R
\name{MetaSpatialFeatureExperiment}
\alias{MetaSpatialFeatureExperiment}
\title{Constructor function for MetaSpatialFeatureExperiment class}
\usage{
MetaSpatialFeatureExperiment(sfe_data = list(), sample_ids = list())
}
\arguments{
\item{sfe_data}{A list of SpatialFeatureExperiment objects.}

\item{sample_ids}{A list of sample IDs matching the sample IDs of the stored
SpatialFeatureExperiment (SFE) objects.}
}
\value{
An instance of the MetaSpatialFeatureExperiment class.
}
\description{
Constructor function for MetaSpatialFeatureExperiment class
}
\examples{
## Create an empty MetaSpatialFeatureExperiment object
msfe <- MetaSpatialFeatureExperiment()

## Add an SFE object to the MetaSpatialFeatureExperiment
## create or load your SpatialFeatureExperiment object

# sfe <- ...

## a SpatialFeatureExperiment object encompassing a single sample named, in
##  this instance, 'sample_1'.

# msfe <- addSFE(msfe, sfe)

## Access an SFE object by name
# sfe_sample_1 <- getSFE(msfe, "sample_1")

}
