% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelGeneVariance.R
\name{modelGeneVariance}
\alias{modelGeneVariance}
\title{Model Gene Variance}
\usage{
modelGeneVariance(
  msfe,
  sample_id,
  method = c("Var", "VarPoisson", "VarSpikes", "CV2", "CV2Spikes"),
  ...
)
}
\arguments{
\item{msfe}{A SpatialFeatureExperiment object.}

\item{sample_id}{Either a logical vector indicating the samples to include
(if TRUE, all samples are included), or a character vector specifying the
sample IDs to include. It is suggested to use a character vector to specify
a specific sample.}

\item{method}{A character string specifying the method for modeling gene
variance.
Possible values: "Var" (default), "VarPoisson", "VarSpikes", "CV2",
"CV2Spikes".
\itemize{
\item - "Var": Model gene variance using scran::modelGeneVar.
\item - "VarPoisson": Model gene variance using
scran::modelGeneVarByPoisson.
\item - "VarSpikes": Model gene variance using
scran::modelGeneVarWithSpikes.
\item - "CV2": Model gene coefficient of variation squared using
scran::modelGeneCV2.
\item - "CV2Spikes": Model gene coefficient of variation squared using
scran::modelGeneCV2WithSpikes.
}}

\item{...}{Additional arguments to be passed to the chosen method.}
}
\value{
A list of modeled gene variances for each gene.
}
\description{
This function models gene variance for each gene in a
SpatialFeatureExperiment.
}
\examples{
\dontrun{
# Model gene variance using the default method ("Var").
modelGeneVariance(msfe)

# Model gene variance using the "VarPoisson" method.
modelGeneVariance(msfe, method = "VarPoisson")
}

}
\references{
McCarthy DJ, Campbell KR, Lun ATL, Wills QF (2017).
Scater: pre-processing, quality control, normalization and visualization of
single-cell RNA-seq data in R. Bioinformatics, 33(8), 1179–1186.
}
\seealso{
\code{\link{scran::modelGeneVar}},
\code{\link{scran::modelGeneVarByPoisson}},
\code{\link{scran::modelGeneVarWithSpikes}},
\code{\link{scran::modelGeneCV2}},
\code{\link{scran::modelGeneCV2WithSpikes}}
}
