% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_moran.R
\name{moranGlobalIPerm}
\alias{moranGlobalIPerm}
\title{Permutation test for Moran's I statistic}
\usage{
moranGlobalIPerm(
  x,
  listw,
  nsim,
  zero.policy = NULL,
  alternative = "greater",
  na.action = na.fail,
  spChk = NULL,
  return_boot = FALSE,
  adjust.n = TRUE
)
}
\arguments{
\item{x}{Numeric vector, same length as neighbours list in listw}

\item{listw}{\code{listw} object, e.g., from \code{nb2listw}}

\item{nsim}{Number of permutations}

\item{zero.policy}{Default \code{attr(listw, "zero.policy")} as set in
\code{listw}}

\item{alternative}{Character string specifying the alternative hypothesis.
Options: "greater" (default), "two.sided", or "less".}

\item{na.action}{Function (default \code{na.fail}). Options: \code{na.omit},
\code{na.exclude}. The weights list is subsetted to remove NAs in the data.
\code{na.pass} is not permitted because it is meaningless in a permutation
test.}

\item{spChk}{Check data vector names against spatial objects for identity
integrity. TRUE or FALSE, default NULL to use \code{get.spChkOption()}}

\item{return_boot}{Return an object of class \code{boot} from the equivalent
permutation bootstrap.}

\item{adjust.n}{Default TRUE. If FALSE, the number of observations is not
adjusted for no-neighbour observations. If TRUE, the number of observations
is adjusted.}
}
\value{
A list with class \code{htest} and \code{mc.sim} containing the
following components:
\item{statistic}{the value of the observed Moran's I.}
\item{parameter}{the rank of the observed Moran's I.}
\item{p.value}{the pseudo p-value of the test.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string giving the method used.}
\item{data.name}{a character string giving the name(s) of the data, and the
number of simulations.}
\item{res}{nsim simulated values of the statistic, final value is observed
statistic}
}
\description{
This function is a wrapper of the \code{\link[spdep]{moran.mc}} function from
the \code{spdep} package developed by Roger Bivand. A permutation test for
Moran's I statistic calculated by using nsim random
permutations of x for the given spatial weighting scheme, establishing the
rank of the observed statistic in relation to the nsim simulated values.
}
\examples{
\dontrun{
# Example usage:

# Export an SFE object from an MSFE object
sfe <- getSFE(msfe, "JBO019")

# A vector of gene expression of one gene over all locations
gene_exp <- assay(sfe, "logcounts")[1,]

# A listw object containing distance-based spatial weights for neighbours
listw <- listw

# Calculate Moran's I for this gene
moran <- moranGlobalIPerm(x = gene_exp, listw = listw, nsim = 499,
zero.policy = TRUE)
}

}
\references{
Cliff, A. D., Ord, J. K. 1981 Spatial processes, Pion, p. 63-5.
}
\seealso{
\code{\link{moran}}, \code{\link{moran.test}}

\code{\link{moran.test}}, \code{\link{moran}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
