% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_moran.R
\name{moranGlobalITest}
\alias{moranGlobalITest}
\title{Moran's I test for spatial autocorrelation}
\usage{
moranGlobalITest(
  x,
  listw,
  randomisation = TRUE,
  zero.policy = NULL,
  alternative = "greater",
  rank = FALSE,
  na.action = na.fail,
  spChk = NULL,
  adjust.n = TRUE,
  drop.EI2 = FALSE
)
}
\arguments{
\item{x}{Numeric vector, same length as neighbors list in listw}

\item{listw}{\code{listw} object, e.g., from \code{nb2listw}}

\item{randomisation}{Variance of I calculated under the assumption of
randomization. If FALSE, normality.}

\item{zero.policy}{Default \code{attr(listw, "zero.policy")} as set in
\code{listw}}

\item{alternative}{Character string specifying the alternative hypothesis.
Options: greater (default), less, or two.sided.}

\item{rank}{Logical value - default FALSE for continuous variables.
If TRUE, uses the adaptation of Moran's I for ranks suggested by Cliff and
Ord (1981, p. 46).}

\item{na.action}{Function (default \code{na.fail}). Options: \code{na.omit},
\code{na.exclude}. The weights list is subsetted to remove NAs in the data.
\code{na.pass} is used, zero is substituted for NA values in calculating the
spatial lag.}

\item{spChk}{Check data vector names against spatial objects for identity
integrity.
TRUE or FALSE, default NULL to use \code{get.spChkOption()}}

\item{adjust.n}{Default TRUE. If FALSE, the number of observations is not
adjusted for no-neighbor observations. If TRUE, the number of observations
is adjusted.}

\item{drop.EI2}{Default FALSE. If TRUE, emulate CrimeStat <= 4.02.}
}
\value{
A list with class \code{htest} containing the following components:
\item{statistic}{the value of the standard deviate of Moran's I.}
\item{p.value}{the p-value of the test.}
\item{estimate}{the value of the observed Moran's I, its expectation and
variance under the method assumption.}
\item{alternative}{a character string describing the alternative hypothesis.}
\item{method}{a character string giving the assumption used for calculating
the standard deviate.}
\item{data.name}{a character string giving the name(s) of the data.}
}
\description{
This function is a wrapper of the \code{\link[spdep]{moran.mc}} function from
the \code{spdep} package developed by Roger Bivand. Moran's test for spatial
autocorrelation using a spatial weights matrix in weights list form. The
assumptions underlying the test are sensitive to the form of the graph of
neighbour relationships and other factors, and results may be checked
against those of \code{moran.mc} permutations.
}
\note{
Var(I) is taken from Cliff and Ord (1969, p. 28),
and Goodchild's CATMOG 47 (1986), see also Upton & Fingleton (1985) p. 171;
it agrees with SpaceStat, see Tutorial workbook Chapter 22; VI is the second
crude moment minus the square of the first crude moment. The derivation of the
test (Cliff and Ord, 1981, p. 18) assumes that the weights matrix is symmetric.
For inherently non-symmetric matrices, such as k-nearest neighbor matrices,
\code{listw2U()} can be used to make the matrix symmetric.
}
\examples{
\dontrun{
# Example usage:

# Export an SFE object from an MSFE object
sfe <- getSFE(msfe, "JBO019")

# A vector of gene expression of one gene over all locations
gene_exp <- assay(sfe, "logcounts")[1,]

# A listw object containing distance-based spatial weights for neighbors
listw <- listw

# Calculate Moran's I for this gene
moran <- moranGlobalITest(x = gene_exp, listw = listw, zero.policy = TRUE)
}

}
\references{
Cliff, A. D., Ord, J. K. 1981 Spatial processes, Pion, p. 21;
Bivand RS, Wong DWS 2018 Comparing implementations of global and local
indicators of spatial association. TEST, 27(3), 716--748 \doi{10.1007/s11749-
018-0599-x}
}
\seealso{
\code{\link{moran}}, \code{\link{moran.mc}}, \code{\link{listw2U}}

\code{\link{moran}}, \code{\link{moran.mc}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
