% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msfe}
\alias{msfe}
\title{Fibrotic Lung data}
\format{
\subsection{\code{msfe}}{

A MetaSpatialFeatureExperiment class object containing with 2 slots, one for
the SpatialFeatureExperiment (SFE) objects and one with the sample names for
each SFE object, containing one SFE object:
\describe{
\item{sfe_data}{List of 1: contains the SFE object named "Fibrotic"}
\item{sample_ids}{List of 1: contains the SFE object's name}
}
No preprocessing has been performed.
}
}
\source{
\url{https://doi.org/10.1038/s41588-024-01819-2}
}
\usage{
msfe
}
\description{
The fibrotic sample (V19S23-092-D1) from Franzen et. al., 2024, Nat. Genet.
publication \url{https://doi.org/10.1038/s41588-024-01819-2}.
}
\keyword{datasets}
