% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseCounts.R
\name{normaliseCounts}
\alias{normaliseCounts}
\title{Normalise Counts in a Spatial Transcriptomics Experiment}
\usage{
normaliseCounts(msfe, method = c("log"), ...)
}
\arguments{
\item{msfe}{An object containing spatial transcriptomics experiment data
(MetaSpatialFeatureExperiment).}

\item{method}{The normalisation method to be applied. Currently, only "log"
transformation is supported.}

\item{...}{Additional arguments to be passed to the normalisation function.}
}
\value{
A normalised SpatialFeatureExperiment object.
}
\description{
This function normalises counts in a spatial transcriptomics experiment
using specified methods.
}
\details{
The function normalises counts in a spatial transcriptomics
experiment using specified methods. Currently, only the "log" transformation
method is supported. Additional arguments can be passed to the underlying
normalisation function.
}
\examples{
\dontrun{
# Normalise counts using log transformation
# msfe <- # MetaSpatialFeatureExperiment object
msfe <- normaliseCounts(msfe, method = "log", base = 2)
}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{SpatialFeatureExperiment}
\keyword{log}
\keyword{normalisation}
\keyword{transformation}
