% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{outlierCutoff}
\alias{outlierCutoff}
\title{Get outlier cut off value}
\usage{
outlierCutoff(dt, coef = 1.5)
}
\arguments{
\item{dt}{A numeric vector of data.}

\item{coef}{The coefficient used to determine the cutoff for outliers.
The default value is 1.5, which is commonly used in boxplots to define
outliers as values that are more than 1.5 times the IQR away from the upper
or lower quartiles.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{out_up}: The upper cutoff value for outliers.
\item \code{out_down}: The lower cutoff value for outliers.
\item \code{outliers}: A logical vector indicating whether each data point
is an outlier or not. \code{TRUE} indicates an outlier, and \code{FALSE}
indicates a non-outlier.
}
}
\description{
This function identifies outliers in a vector of data based on a coefficient
and the interquartile range (IQR).
}
\examples{
# Generate a vector of random data
data <- rnorm(100)

# Detect outliers using the default coefficient (1.5)
outlierCutoff(data)

# Detect outliers using a custom coefficient (2.0)
outlierCutoff(data, coef = 2.0)

}
