% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_genes.R
\name{perGeneLogMean}
\alias{perGeneLogMean}
\title{Calculate Per-Gene Logarithmic Means in a Spatial Transcriptomics Experiment}
\usage{
perGeneLogMean(msfe, ...)
}
\arguments{
\item{msfe}{A MetaSpatialFeatureExperiment object containing spatial
transcriptomics experiment data.}

\item{...}{Additional arguments. Not used currently}
}
\value{
An updated MetaSpatialFeatureExperiment with a column in rowData
for per-gene logarithmic means for each unique sample.
}
\description{
This function calculates per-gene logarithmic means in a spatial
transcriptomics experiment for each unique sample.
}
\details{
The function iterates through each unique sample in the
MetaSpatialFeatureExperiment and calculates per-gene logarithmic means.
Additional arguments can be passed to the underlying calculation function.
}
\examples{
\dontrun{
# Calculate per-gene logarithmic means in a MetaSpatialFeatureExperiment
# msfe <- # A MetaSpatialFeatureExperiment object
msfe <- perGeneLogMean(msfe)
}

}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{MetaSpatialFeatureExperiment}
\keyword{experiment,}
\keyword{log}
\keyword{mean,}
\keyword{per-gene}
\keyword{spatial}
\keyword{transcriptomics}
