% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{plotFGWC_factorSelection}
\alias{plotFGWC_factorSelection}
\title{Plot Decrease in Reconstruction Error versus Random Noise}
\usage{
plotFGWC_factorSelection(kErrDiff, font.size = 12)
}
\arguments{
\item{kErrDiff}{Reduction in reconstruction error above noise output by
\code{fgwc_nmfFactorNumber}.}

\item{font.size}{Font size to use for plotting (default is 12).}
}
\value{
A ggplot object showing the decrease in reconstruction error versus
random noise.
}
\description{
This function plots the decrease in reconstruction error versus random noise.
The point where the change in reconstruction error matches the change in
reconstruction error for the randomized matrix is the optimal number of
factors to use.
}
\details{
The function generates a plot that visualizes the reduction in
reconstruction error above noise for different numbers of factors. The
optimal number of factors is where the error reduction matches the change
in reconstruction error for the randomized matrix.
}
\examples{
\dontrun{
# Example data for reduction in reconstruction error
k_err_diff <- fgwc_
plotFGWC_factorSelection(k_err_diff, font.size = 10)
}

}
\seealso{
\code{\link[ggplot2]{ggplot}} for additional details on plotting.

Other clustering functions: 
\code{\link{fgwc_nmf}()},
\code{\link{fgwc_nmfFactorNumber}()},
\code{\link{fgwc_params}()}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\concept{clustering functions}
\keyword{clustering}
\keyword{nmf}
\keyword{spatial-expression}
