% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{plotFGWC_markersHeatmap}
\alias{plotFGWC_markersHeatmap}
\title{Plot Heatmap for Fuzzy Geographically Weighted Clustering (FGWC)}
\usage{
plotFGWC_markersHeatmap(
  fgwc,
  m_sfe,
  sample_id,
  markers,
  cluster_no,
  cutree_cols = NA,
  annot_colours = NULL,
  ph.color = NULL,
  ...
)
}
\arguments{
\item{fgwc}{The result of FGWC clustering.}

\item{m_sfe}{The spatial feature experiment (SFE) or metaSFE object.}

\item{sample_id}{The sample ID.}

\item{markers}{A data frame with 4 columns: gene.name, ensg.ID, Type,
and Subtype.}

\item{cluster_no}{The cluster number for which the heatmap is generated.}

\item{cutree_cols}{Optional, cutree result for columns.}

\item{annot_colours}{Annotation colours passed to the \code{annotation_colors}
argument of \code{pheatmap}. Default is NULL and colours are automatically
selected. If you wish to provide your own colours make sure is in a named
list like this one: list(Type = c(colours here), Subtype = c(colours here)).}

\item{ph.color}{A colour vector used in heatmap. Passed to the \code{color}
argument in \code{pheatmap}. More info in \code{?pheatmap}.}

\item{...}{Arguments passed to \code{pheatmap}.}
}
\value{
A pheatmap object.
}
\description{
This function plots a heatmap for the gene expression of marker genes for a
specific cluster.
}
\details{
Additional details about the function or its behaviour can be added here.
}
\examples{
\dontrun{
# Example usage:
fgwc_result <- fgwc_nmf(m_sfe, sample_id, top_hvgs, ncomponents = 2,
ntop = 600)

markers_data <- read.csv("markers_data.csv") # Provide your markers data
cluster_no <- 1 # Provide the cluster number for which heatmap is desired

plotFGWC_markersHeatmap(fgwc_result, m_sfe, sample_id, markers_data, cluster_no)
}

}
\seealso{
\code{\link{fgwc_nmf}}, \code{\link{plotFGWC_single}},
\code{\link{plotFGWC_multi}}, \code{\link{pheatmap}}

Other plotting functions: 
\code{\link{plotFGWC_multiMap}()},
\code{\link{plotFGWC_multiViolin}()},
\code{\link{plotFGWC_nmfFactorsMap}()},
\code{\link{plotFGWC_singleMap}()}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\concept{plotting functions}
\keyword{clustering}
\keyword{fgwc}
\keyword{fuzzy-logic}
\keyword{heatmap}
\keyword{spatial}
\keyword{visualization}
