% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{plotFGWC_multi}
\alias{plotFGWC_multi}
\title{Plot Multiple Cluster Results for Fuzzy Geographically Weighted Clustering
(FGWC)}
\usage{
plotFGWC_multi(fgwc, m_sfe, sample_id, palette = "YlGnBu")
}
\arguments{
\item{fgwc}{The result of FGWC clustering.}

\item{m_sfe}{The spatial feature experiment (SFE) or metaSFE object.}

\item{sample_id}{The sample ID.}

\item{palette}{The color palette for the cluster memberships.
Default is "YlGnBu".}
}
\value{
A ggplot object.
}
\description{
This function plots the results of Fuzzy Geographically Weighted Clustering
(FGWC) for multiple clusters. It plots the membership percentage of each
cluster in each location.
}
\details{
The function produces a panel of maps. One map per cluster. The colours in
all maps are scaled to be the same. Meaning that the colour for 50\% in map 1
is going to be the same with colour for 50\% in map 3.
}
\examples{
\dontrun{
# Example usage:
fgwc_result <- fgwc_nmf(m_sfe, sample_id, top_hvgs, ncomponents = 2,
ntop = 600)
plotFGWC_multi(fgwc_result, m_sfe, sample_id)
}

}
\seealso{
\code{\link{fgwc_nmf}}, \code{\link{plotFGWC_single}},
\code{\link{plotFGWC_heatmap}}

Other plotting functions: 
\code{\link{plotFGWC_heatmap}()},
\code{\link{plotFGWC_single}()}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\concept{plotting functions}
\keyword{clustering}
\keyword{fgwc}
\keyword{fuzzy-logic}
\keyword{ggplot2}
\keyword{spatial}
\keyword{visualization}
