% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{plotFGWC_nmfFactorsMap}
\alias{plotFGWC_nmfFactorsMap}
\title{Plot NMF Factors as a Map}
\usage{
plotFGWC_nmfFactorsMap(nmf, m_sfe, sample_id = NULL)
}
\arguments{
\item{nmf}{The NMF result object. As generated by STExplorer's \code{fgwc_nmf}
function.}

\item{m_sfe}{The SpatialFeatureExperiment (SFE) or metaSFE object.}

\item{sample_id}{The sample ID.}
}
\value{
A ggplot object.
}
\description{
This function plots the Non-Negative Matrix Factorization (NMF) factors as a map.
}
\examples{
\dontrun{
# Example usage:
plotFGWC_nmfFactorsMap(nmf_result, m_sfe, sample_id)
}

}
\seealso{
\code{\link{nmf}}, \code{\link{plotFGWC_single}},
\code{\link{plotFGWC_multi}}

Other plotting functions: 
\code{\link{plotFGWC_markersHeatmap}()},
\code{\link{plotFGWC_multiMap}()},
\code{\link{plotFGWC_multiViolin}()},
\code{\link{plotFGWC_singleMap}()}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\concept{plotting functions}
\keyword{factorization}
\keyword{ggplot2}
\keyword{nmf}
\keyword{spatial}
\keyword{visualization}
