% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{plotFGWC_nmfMetagenesHeatmap}
\alias{plotFGWC_nmfMetagenesHeatmap}
\title{Plot FGWC NMF Metagene Heatmap}
\usage{
plotFGWC_nmfMetagenesHeatmap(fgwc, ...)
}
\arguments{
\item{fgwc}{An object of class \code{fgwc} containing the results of FGWC
analysis.}

\item{...}{Arguments passed to \code{pheatmap}.}
}
\value{
A heatmap plot of factor scores with optional annotations.
}
\description{
This function generates a clustered heatmap of metagene signatures.
Essentially is visualising the gene scores obtained from NMF for each
signature.
}
\examples{
\dontrun{
# Assuming `fgwc_object` is a valid FGWC object
plotFGWC_nmfMetagenesHeatmap(fgwc_object)
}

}
