% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{plotFGWC_pie}
\alias{plotFGWC_pie}
\title{Plot Fuzzy Geographically Weighted Clustering (FGWC) - Pie and Donut Plot}
\usage{
plotFGWC_pie(
  fgwc,
  m_sfe,
  sample_id = NULL,
  mapping = aes(pie = cluster, donut = annotation),
  start = 0,
  addPieLabel = TRUE,
  addDonutLabel = TRUE,
  showRatioDonut = TRUE,
  showRatioPie = TRUE,
  ratioByGroup = TRUE,
  showRatioThreshold = 0.02,
  labelposition = 2,
  labelpositionThreshold = 0.1,
  r0 = 0.3,
  r1 = 1,
  r2 = 1.2,
  explode = NULL,
  selected = NULL,
  explodePos = 0.1,
  color = "white",
  pieColours = NULL,
  pieAlpha = 0.8,
  donutAlpha = 1,
  maxx = NULL,
  showPieName = TRUE,
  showDonutName = FALSE,
  title = NULL,
  pieLabelSize = 4,
  donutLabelSize = 3,
  titlesize = 5,
  explodePie = TRUE,
  explodeDonut = FALSE,
  use.label = TRUE,
  use.labels = TRUE,
  family = ""
)
}
\arguments{
\item{fgwc}{Object of class \linkS4class{fgwc}. The Fuzzy Geographically
Weighted Clustering object, typically created by
\code{\link[STExplorer]{fgwcSTE}}.}

\item{m_sfe}{The spatial feature experiment (SFE) or metaSFE object.}

\item{sample_id}{A character vector. The sample ID. Defaults to NULL and
selects the fisrt sample if a metaSFE object is provided to the \code{m_sfe}
argument. If an SFE object is provided then this argument is omitted.}

\item{mapping}{A aesthetic mapping created by \code{\link[ggplot2]{aes}}. It
specifies how the data is mapped to the plot. The expected mappings are
"pie" and "donut" which determine the segmentation of the pie and donut
plots, respectively. The variables that can be supplied to those mappings
are currently one of \code{annotation}, \code{clustering}, and \code{factors}. The use of
the \code{annotation} assumes the existence of an \code{annotation} column in the
\code{colData}.}

\item{start}{Angle in radians at which the pie/donut plot starts.}

\item{addPieLabel}{A logical indicating whether to add labels to the pie plot
segments.}

\item{addDonutLabel}{Logical indicating whether to add labels to the donut
plot segments.}

\item{showRatioDonut}{Logical indicating whether to show the ratio of each
donut segment to the whole.}

\item{showRatioPie}{Logical indicating whether to show the ratio of each pie
segment to the whole.}

\item{ratioByGroup}{Logical indicating whether the ratio displayed on the
donut plot should be calculated by group or individually.}

\item{showRatioThreshold}{The minimum ratio required to display a segment
label in the pie or donut plot.}

\item{labelposition}{Position of the labels in the plot. It can take values
0, 1, 2, where 0 places the labels outside the plot, 1 places them on the
segments, and 2 places them mostly outside with some overlapping.}

\item{labelpositionThreshold}{Threshold for the label position 2. Labels
with ratios below this threshold will be positioned outside the plot, while
others may overlap.}

\item{r0, }{@param r1, @param r2 Radii of the pie plot, inner donut plot, and
outer donut plot, respectively.}

\item{explode}{A numeric vector or NULL. If not NULL, the specified segments
in the pie plot will be exploded.}

\item{selected}{A numeric vector or NULL. If not NULL, the specified
segments in the donut plot will be exploded.}

\item{explodePos}{Position to which the exploded segments in the pie
or donut plots will be moved.}

\item{color}{The border colour of the pie and donut plots.}

\item{pieColours}{A character vector of colours to be used in the pie plot
segments. If not provided, colours are automatically generated.}

\item{pieAlpha}{Transparency of the pie plot segments.}

\item{donutAlpha}{Transparency of the donut plot segments.}

\item{maxx}{Maximum radius of the pie/donut plots. If not provided, it is
determined automatically.}

\item{showPieName}{Logical indicating whether to display the name of the
pie plot.}

\item{showDonutName}{Logical indicating whether to display the name of
the donut plot.}

\item{title}{A character string to add as title above the donut plot.
If NULL, no title will be displayed.}

\item{pieLabelSize}{Size of the labels in the pie plot.}

\item{donutLabelSize}{Size of the labels in the donut plot.}

\item{titlesize}{Size of the title or plot names.}

\item{explodePie}{Logical indicating whether to explode the pie plot
segments at the positions specified by \code{explode}.}

\item{explodeDonut}{Logical indicating whether to explode the donut plot
segments at the positions specified by \code{selected}.}

\item{use.label}{Logical indicating whether to use the original labels from
the data for the pie and donut plots.}

\item{use.labels}{Logical indicating whether to use the original labels from
the data for the segment labels.}

\item{family}{The font family to use in the plot.}
}
\value{
A ggplot2 plot object.
}
\description{
This function plots the FGWC using pie and donut plots. It visualises the
distribution of clusters, annotations or NMF factors for each location in
the spatial omics dataset. The function accepts several parameters to
customise the plot appearance and behaviour.
}
\examples{
\dontrun{
# Plot FGWC using default settings
plotFGWC_pie(fgwC_obj, m_sfe_obj)
}

}
