% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{plotFGWC_singleMap}
\alias{plotFGWC_singleMap}
\alias{plotFGWC_single}
\title{Plot Single Cluster Result for Fuzzy Geographically Weighted Clustering
(FGWC)}
\usage{
plotFGWC_singleMap(fgwc, m_sfe, sample_id, colours = NULL)
}
\arguments{
\item{fgwc}{The result of FGWC clustering.}

\item{m_sfe}{The spatial feature experiment (SFE) or metaSFE object.}

\item{sample_id}{The sample ID.}

\item{colours}{A vector of colours for the clusters. If NULL, default
colours will be used.}
}
\value{
A ggplot object.
}
\description{
This function plots the result of Fuzzy Geographically Weighted Clustering
(FGWC) for a single cluster. The selected cluster is the cluster with the
highest membership percentage for each location.
}
\examples{
\dontrun{
# Example usage:
nmf <- fgwc_nmf(m_sfe, sample_id, top_hvgs, ncomponents = 2,
ntop = 600)
fgwc_result <- fgwc(m_sfe, sample_id, data = nmf)
plotFGWC_single(fgwc_result, m_sfe, sample_id)
}

}
\seealso{
\code{\link{fgwc_nmf}}, \code{\link{plotFGWC_multi}},
\code{\link{plotFGWC_heatmap}}

Other plotting functions: 
\code{\link{plotFGWC_markersHeatmap}()},
\code{\link{plotFGWC_multiMap}()},
\code{\link{plotFGWC_multiViolin}()},
\code{\link{plotFGWC_nmfFactorsMap}()}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\concept{plotting functions}
\keyword{clustering}
\keyword{fgwc}
\keyword{fuzzy-logic}
\keyword{ggplot2}
\keyword{spatial}
\keyword{visualization}
