% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwcSTE.R
\name{plotFGWC_subClust}
\alias{plotFGWC_subClust}
\title{Plot FGWC Sub-Clusters in Map}
\usage{
plotFGWC_subClust(
  heatmap,
  k,
  clust,
  m_sfe,
  sample_id,
  annot_cols = NULL,
  subClust_cols = NULL
)
}
\arguments{
\item{heatmap}{The FGWC heatmap object.}

\item{k}{The number of sub-clusters.}

\item{clust}{The main cluster number.}

\item{m_sfe}{Either a SpatialfeatureExperiment or a
SpatialfeatureExperiment object containing spatial transcriptomics data.}

\item{sample_id}{The sample.}

\item{annot_cols}{Vector of colours for ground truth annotations. If NULL,
colours are assigned automatically.}

\item{subClust_cols}{Vector of colours for sub-clusters. If NULL, colours are
generated automatically.}
}
\value{
A ggplot object visualizing the sub-clusters in the spatial context.
}
\description{
This function generates a map of sub-clusters based on the fuzzy
geographically weighted clustering (FGWC) results.
}
\details{
This function takes the results of the FGWC clustering and produces a spatial
map of sub-clusters for a specific FGWC cluster. It overlays the identified
sub-clusters on top of the original spatial features, allowing users to
visualise the distribution of sub-clusters within the main cluster areas.
}
\examples{
\dontrun{
# Plot sub-clusters with default colors.
plotFGWC_subClust(heatmap, k = 5, clust = 3, m_sfe, sample_id = "JBO019")

# Plot sub-clusters with custom colors.
plotFGWC_subClust(heatmap, k = 5, clust = 3, m_sfe, sample_id = "JBO019",
                  annot_cols = c("red", "blue"),
                  subClust_cols = c("orange","green","yellow","red","blue"))
}

}
\author{
Eleftherios (Lefteris) Zormpas
}
