% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{plotFGWC_subHeatmap}
\alias{plotFGWC_subHeatmap}
\title{Plot an FGWC Sub-Cluster Heatmap of gene expression}
\usage{
plotFGWC_subHeatmap(
  heatmap,
  k,
  markers,
  m_sfe,
  sample_id,
  cluster_no,
  cutree_cols = NA,
  annot_colours = NULL,
  ph.color = NULL,
  ...
)
}
\arguments{
\item{heatmap}{The FGWC heatmap object.}

\item{k}{The number of sub-clusters used in FGWC.}

\item{markers}{A data frame with gene markers information, including columns
"gene.name", "ensg.ID"  "Type", and "Subtype".}

\item{m_sfe}{Either a a SpatialfeatureExperiment or a
SpatialfeatureExperiment object containing spatial transcriptomics data.}

\item{sample_id}{The sample ID used for FGWC.}

\item{cluster_no}{The specific sub-cluster number for which the heatmap is
generated.}

\item{cutree_cols}{Number of columns for cutree to cut the sample tree of the
heatmap. If NA, the tree will not be cut.}

\item{annot_colours}{Annotation colours passed to the \code{annotation_colors}
argument of \code{pheatmap}. Default is NULL and colours are automatically
selected. If you wish to provide your own colours make sure is in a named
list like this one: list(Type = c(colours here), Subtype = c(colours here)).}

\item{ph.color}{A colour vector used in heatmap. Passed to the \code{color}
argument in \code{pheatmap}. More info in \code{?pheatmap}.}

\item{...}{Arguments passed to \code{pheatmap}.}
}
\value{
A pheatmap object representing the gene expression heatmap for the
specified sub-cluster.
}
\description{
This function generates a heatmap for the gene expression profiles within a
specific sub-cluster identified by fuzzy geographically weighted clustering
(FGWC).
}
\details{
This function takes the results of the FGWC clustering, extracts the gene
expression profiles within the specified sub-cluster, and generates a
heatmap. The heatmap provides insights into the expression patterns of
selected markers within the identified sub-cluster.
}
\examples{
\dontrun{
# Plot a heatmap for sub-cluster 3 using default colors.
plotFGWC_subHeatmap(heatmap, k = 5, markers, m_sfe, sample_id = "JBO019",
cluster_no = 3)

# Plot a heatmap for sub-cluster 2 with custom cutree colors.
plotFGWC_subHeatmap(heatmap, k = 5, markers, m_sfe, sample_id = "JBO019",
cluster_no = 2, cutree_cols = 4)
}

}
