% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_plot.R
\name{plotGeneExpression}
\alias{plotGeneExpression}
\title{Map Gene Expression}
\usage{
plotGeneExpression(
  m_sfe,
  genes,
  sample_id = NULL,
  assay = c("counts", "logcounts", "unNormLogCounts"),
  minmax = c(0, Inf),
  type = c("spot", "hex", "cntd"),
  res = c("lowres", "hires", "fullres", "none"),
  fill_args = list(),
  alpha = 0.3,
  title_type = "gene_name",
  ...
)
}
\arguments{
\item{m_sfe}{An object of class 'SpatialFeatureExperiment' or
'MetaSpatialFeatureExperiment' containing spatial
data.}

\item{genes}{A vector of gene names or ENSG_IDs to be plotted.}

\item{sample_id}{A character string specifying the sample ID for
analysis. If NULL the first sample is selected.}

\item{assay}{Character vector specifying the type of assay data
to plot. Options are "counts" for raw counts,
"logcounts" for log2-normalized counts, and "unNormLogCounts"
for un-normalised log-2 transformed only counts. Default is
"counts".}

\item{minmax}{Numeric vector of length 2 specifying the minimum
and maximum expression values to include in the
plot. Default is c(0, Inf), meaning all values
above 0 will be included.}

\item{type}{Character vector specifying the type of spatial data
to use for plotting. Options are "spot" for spot
geometry and "hex" for hexagon geometry. Default is
"spot".}

\item{res}{Character vector specifying the resolution of the image
to fetch for plotting. Options are "lowres" for
low-resolution, "hires" for high-resolution,
"fullres" for full-resolution, and "none" to skip
fetching image data. Default is "none".}

\item{fill_args}{Additional arguments for customizing the fill
colour palette.}

\item{alpha}{Numeric value specifying the transparency level of
the plotted polygons. Default is 0.3.}

\item{title_type}{A character string indicating the label for the y-axis.
It can be "gene_name" to use gene names or any other value
to use Ensemble Gene IDs. Default is "gene_name".}

\item{...}{Additional arguments to be passed to the plotting
functions.}
}
\value{
A plot of gene expression data on spatial feature
experiment objects.
}
\description{
A function to plot gene expression data on spatial feature
experiment (SFE) objects as a map.
}
\examples{
\dontrun{
# Example usage:
plotGeneExpression(m_sfe = sfe_object,
                   genes = c("Gene1", "Gene2"),
                   sample_id = "sample123",
                   assay = "logcounts",
                   minmax = c(0, 100),
                   type = "hex",
                   res = "hires",
                   fill_args = list(option = "magma",
                                    na.value = "grey"),
                   alpha = 0.5)
}

}
\seealso{
\code{\link{addSpatialData}}, \code{\link{plotSpatialData}}
}
\author{
Eleftherios Zormpas
}
