% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneVariance.R
\name{plotGeneVariance}
\alias{plotGeneVariance}
\title{Plot Gene Variance}
\usage{
plotGeneVariance(dec, hvgs, sample_id = TRUE, ...)
}
\arguments{
\item{dec}{A list of modelled gene variances for each gene as generated by
the \code{modelGenevariance} function.}

\item{hvgs}{Character vector specifying the high variable genes (HVGs) to
plot as generated by the \code{getTopHighVarGenes}
function.}

\item{sample_id}{Logical or character vector specifying the sample IDs to
include in the plot. If TRUE (default), all samples are included.}

\item{...}{Additional arguments to pass to the ggplot2 plotting functions.}
}
\value{
Returns a ggplot2 plot object representing the mean-variance
relationship of the specified genes.
}
\description{
Visualizes the relationship between the mean and variance of log-expression
for specified genes in a SpatialFeatureExperiment.
}
\details{
This function calculates the mean and variance of log-expression
for the specified high variance genes (HVGs) in a SpatialFeatureExperiment.
It then visualises this relationship using a scatter plot, with a trend line
fitted to the data.
}
\examples{
\dontrun{
# hvgs : high variable genes as generated by the getTopHighVarGenes
plotGeneVariance(dec = modelled_var, hvgs = hvgs, sample_id = "JBO019")
}

}
\seealso{
\code{\link[SpatialFeatureExperiment]{SpatialFeatureExperiment}},
\code{\link[ggplot2]{ggplot}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{ggplot2}
\keyword{spatial}
\keyword{visualisation}
