% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_plot.R
\name{plotGWPCA_discr}
\alias{plotGWPCA_discr}
\title{Visualise outliers in box-plot or map}
\usage{
plotGWPCA_discr(
  gwpca,
  type = c("box", "map"),
  box_args = list(),
  jitt_args = list(),
  hline_args = list(),
  lab_args = list(),
  ...
)
}
\arguments{
\item{gwpca}{A list of class \code{gwpca} containing the necessary data.}

\item{type}{Specifies the type of plot to generate. Can be either "box" for
a boxplot or "map" for a map. Default is "box".}

\item{box_args}{Additional arguments to customize the appearance of the
boxplot.}

\item{jitt_args}{Additional arguments to customize the appearance of the
jitter points in the boxplot.}

\item{hline_args}{Additional arguments to customize the appearance of the
horizontal line in the boxplot.}

\item{lab_args}{Additional arguments to customize the labels of the plot.}

\item{...}{Additional arguments to be passed to the \code{geom_boxplot()} or
\code{geom_sf()} functions.}
}
\value{
The function returns the created plot.
}
\description{
This function plots discrepancies using either a boxplot or a map. It
visualizes the discrepancies in the local Principal Components (PCs) for each
location based on the coefficient of variation (CV).
}
\section{Dependencies}{
The function depends on the \code{ggplot2} and \code{sf} packages.

The function first checks the validity of the input arguments. It then
fetches the necessary data from the \code{gwpca} object and calculates the
outliers using the \code{outlier_cutoff()} function. Depending on the \code{type}
specified, the data is prepared accordingly for either a boxplot or a map.

For the boxplot, the discrepancies (\code{cvs}) are reshaped into a data frame and
plotted using \code{geom_boxplot()}. Jitter points can also be added using
\code{geom_jitter()}. A horizontal line is drawn at the upper outlier threshold.
Customization of the plot appearance is possible through the provided
arguments.

For the map, the discrepancies, along with the geometry information, are
combined into a data frame. The map is then created using \code{geom_sf()}, with
the discrepancies represented by fill color. Outliers are also marked with a
transparent fill color and a black border. Customization options for the map
appearance are available through the arguments.
}

\examples{
# Load data
data(gwpca)

# Plot discrepancies as a boxplot
plotGWPCA_discr(gwpca, type = "box")

# Plot discrepancies as a map
plotGWPCA_discr(gwpca, type = "map")

}
\author{
Eleftherios (Lefteris) Zormpas
}
