% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGWPCA_.R
\name{plotGWPCA_discrHeatmap}
\alias{plotGWPCA_discrHeatmap}
\title{Plot GWPCA Discrepancy Heatmap}
\usage{
plotGWPCA_discrHeatmap(
  m_sfe,
  assay,
  vars = NULL,
  focus,
  dMetric,
  sample_id,
  diam,
  mean.diff = 1,
  show.vars = c("top", "all"),
  scale = "row",
  gene.names = FALSE,
  ...
)
}
\arguments{
\item{m_sfe}{a SpatialFeatureExperiment or a MetaSpatialFeatureExperiment
object.}

\item{assay}{Assay type for the spatial expression data (counts, logcounts
etc.).}

\item{vars}{Variables of interest (genes) to be evaluated. Default is NULL,
which includes all variables.}

\item{focus}{Tissue locations of interest (barcodes) for which discrepancy
heatmaps will be generated.}

\item{dMetric}{Distance metric used for generating the distance matrix which
will be used to identify the focus location's neighbours.}

\item{sample_id}{Sample ID for which discrepancy data is retrieved.}

\item{diam}{Bandwidth parameter for selecting neighbours for the heatmap.}

\item{mean.diff}{Threshold for selecting genes based on the difference from
the mean discrepancy score.}

\item{show.vars}{Display option for variables (genes) in the heatmap. Options
are "top" (genes with higher discrepancies) or "all" (all genes).}

\item{scale}{Scaling method for the heatmap. Options are "row" (row scaling)
or "none" (no scaling).}

\item{gene.names}{Logical indicating whether to display gene names instead of
ENSGene IDs.}

\item{...}{Additional arguments to be passed to the pheatmap function.}
}
\value{
plots a heatmap.
}
\description{
Generates a discrepancy heatmap based on GWPCA results for specific tissue
locations.
}
\details{
This function generates discrepancy heatmaps based on GWPCA results
for specific tissue locations (focus). It retrieves the discrepancy data
using the \code{gwpca_Discrepancy} function, annotates the columns with distance
information, generates annotation colors based on distance levels, and plots
the heatmap using the \code{pheatmap} function.
}
\examples{
# Load data
data(sfe)

# Set parameters
assay <- "logcounts"
focus <- colnames(sfe)[1] # outlier Barcode
sample_id <- "JBO019"
dMetric <- "euclidean"
bw <- 450
show.vars <- "top"

# Plot
plotGWPCA_discrHeatmap(sfe, assay = assay,
focus = focus,
dMetric = dMetric, sample_id = sample_id,
bw = bw, mean.diff = 1, show.vars = show.vars, scale = "row",
gene.names = TRUE)

}
\author{
Eleftherios (Lefteris) Zormpas
}
