% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_plot.R
\name{plotGWPCA_FuncCLust}
\alias{plotGWPCA_FuncCLust}
\title{Plot Functional Clustering Results from GWPCA}
\usage{
plotGWPCA_FuncCLust(gsea_map, count = 5, legend, legend.title)
}
\arguments{
\item{gsea_map}{A spatial dataset containing functional clustering results,
typically obtained from the output of the
\code{\link{gwpca_FunctionalClustering}} function.}

\item{count}{An integer specifying the minimum count required for a cluster
to be displayed in the legend.}

\item{legend}{A character string specifying the legend position. The
arguments are the \code{ggplot2}'s \code{legend.position} arguments: "top",
"bottom", "left", "right".}

\item{legend.title}{A character string specifying the legend title.}
}
\value{
A grid of spatial plots representing different aspects of functional
clustering results.
}
\description{
This function generates a series of spatial plots representing various
aspects of functional clustering results obtained from Gene Set Enrichment
Analysis (GSEA) combined with Geometrically Weighted Principal Component
Analysis (GWPCA).
}
\examples{
# Example usage:
# plotGWPCA_FuncCLust(gsea_map, count = 5, legend = "right")

}
\seealso{
Other Spatial Transcriptomics Analysis: 
\code{\link{.int_generateGeneList}()},
\code{\link{.int_runGSEA}()},
\code{\link{gwpca_FunctionalClustering}()}
}
\concept{Spatial Transcriptomics Analysis}
\keyword{clustering}
\keyword{functional}
\keyword{gsea}
\keyword{gwpca}
\keyword{plot}
\keyword{visualisation}
