% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGWPCA_.R
\name{plotGWPCA_global}
\alias{plotGWPCA_global}
\title{Plot global PCA results}
\usage{
plotGWPCA_global(
  gwpca,
  comps = 1:10,
  type = c("scree", "scatter"),
  point_args = list(),
  line_args = list(),
  ...
)
}
\arguments{
\item{gwpca}{a list of class \code{gwpca}, generated by \code{gwpca.ste}.}

\item{comps}{the components to be plotted. If type == scree, it is a numeric
vector. If type == scatter, it is a numeric vector of length 2
(i.e., comps 1:2 will plot the two first components).}

\item{type}{either 'scree' or 'scatter'.}

\item{point_args}{arguments to pass to \code{geom_point()} when you don't
want them to be shared with \code{geom_line()}.}

\item{line_args}{arguments to pass to \code{geom_line()} when you don't
want them to be shared with \code{geom_point()}.}

\item{...}{arguments to be passed to both \code{geom_point()} and
\code{geom_line()}.}
}
\value{
A ggplot object.
}
\description{
A function to plot the global PCA results as generated by \code{gwpca.ste}.
Has to options, 'scree' and 'scatter'. The first will plot a scree plot of
the percentage of total variation of a selection of principal components. The
latter will produce a scatter plot of the two selected components.
}
\details{
This function plots the results of a global principal component analysis. It
can generate two types of plots: a scree plot and a scatter plot.
\itemize{
\item The scree plot (type = "scree") shows the percentage of variation explained
by each component. It helps in determining the optimal number of components
to retain.
\item The scatter plot (type = "scatter") shows the scores of two components on
a two-dimensional plot. It requires running the gwpcaSTE function with the
scores argument set to TRUE to calculate the scores.
}

The comps argument specifies the components to include in the plot. By
default, it includes the first 10 components. The point_args and line_args
arguments allow for customization of the appearance of the points and lines
in the plot, respectively.
}
\examples{
# Load gwpca data
data(gwpca)

# Plot scree plot of PCA results
plotGWPCA_global(gwpca, type = "scree")

\dontrun{
# To plot a scatter plot it requires you to run the [gwpcaSTE()] function
# with `scores = FALSE`.
# Plot scatter plot of PCA results for components 1 and 2
plotGWPCA_global(gwpca, type = "scatter", comps = c(1, 2))

# Customize the appearance of points in the scatter plot
plotGWPCA_global(gwpca, type = "scatter",
point_args = list(size = 3, color = "red"))
}
}
\author{
Eleftherios (Lefteris) Zormpas
}
