% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGWPCA_.R
\name{plotGWPCA_leadingG}
\alias{plotGWPCA_leadingG}
\title{Plot Leading Genes or Leading Gene Groups in GWPCA}
\usage{
plotGWPCA_leadingG(
  gwpca,
  comps,
  type = c("single", "multi"),
  arrange = FALSE,
  legend.position = "right",
  cutoff = 12,
  size = 8,
  ...
)
}
\arguments{
\item{gwpca}{a list of class \code{gwpca}, generated by \code{gwpca.ste} and
updated with leading genes by \code{gwpca.LeadingGene}.}

\item{comps}{the components to be plotted.}

\item{type}{either 'single' for single leading gene per location or 'multi'
for k leading genes per location.}

\item{arrange}{to indicate whether to arrange the plots in a panel or not.
The default is set to FALSE.}

\item{...}{arguments to be passed to both \code{geom_sf()} outside of the
aesthetic.}
}
\value{
Returns a list of plots; one per Principal Component. If arrange ==
TRUE then a 2x2 panel is plotted too. If more than 4 PCs are provided in the
\code{comps} argument then only a list of plots is returned.
}
\description{
A function to map the leading genes or leading gene groups per location.
}
\details{
This function plots the leading genes or gene groups in a global weighted
principal component analysis (GWPCA). It can generate two types of plots:
individual plots for leading genes (type = "single") or a single plot for
gene groups (type = "multi").

For type = "single", the function plots the leading genes individually. The
comps argument specifies the components for which to plot the leading genes.
The leading genes are represented as spatial polygons on a plot, where each
gene is coloured based on its score for the corresponding component. The ...
argument allows for customizing the appearance of the plots using additional
arguments passed to the geom_sf function.

For type = "multi", the function plots the gene groups. The comps argument
specifies the components for which to plot the gene groups. The gene groups
are represented as spatial polygons on a single plot, where each group is
coloured based on the scores of the leading genes in that group. The ...
argument allows for customizing the appearance of the plot using additional
arguments passed to the geom_sf function.

The arrange argument, when set to TRUE, arranges the individual plots in a
grid layout when the number of components specified is less than or equal to
4. By default, the individual plots are returned as a list. If arrange is
TRUE and the number of components is greater than 4, a message is displayed,
and the individual plots are returned as a list.
}
\examples{
# Load gwpca data
data(gwpca)

#Identify single leading gene per location for PC1 and PC2
gwpca <- gwpca_LeadingGene(gwpca = gwpca, sfe = sfe, pc_nos = c(1, 2),
type = "single", names = "id")

# Identify top 3 leading genes per location for PC1 and PC2
gwpca <- gwpca_LeadingGene(gwpca = gwpca, sfe = sfe, pc_nos = c(1, 2),
type = "multi", genes_n = 3, method = "membership", names = "gene_names")

# Plot individual leading genes for components 1 and 2
plotGWPCA_leadingG(gwpca, comps = c(1, 2), type = "single")

# Plot gene groups for components 1 and 2
plotGWPCA_leadingG(gwpca, comps = c(1, 2), type = "multi")

# Arrange individual plots in a grid layout
plotGWPCA_leadingG(gwpca, comps = c(1, 2),
type = "single", arrange = TRUE)

}
\author{
Eleftherios (Lefteris) Zormpas
}
