% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_plot.R
\name{plotGWPCA_leadingScores}
\alias{plotGWPCA_leadingScores}
\title{Plot Leading Scores from GWPCA}
\usage{
plotGWPCA_leadingScores(
  gwpca,
  comps,
  genes,
  type = c("spot", "hex"),
  colours = c("viridis", "custom"),
  col_args = list(),
  gene_names = TRUE
)
}
\arguments{
\item{gwpca}{a list of class \code{gwpca}, generated by \code{gwpca.ste} and
updated with leading genes by \code{gwpca.LeadingGene}.}

\item{comps}{the components to be plotted. Can be either in a \code{c(1,3,5)} or
in a \code{1:3} form.}

\item{genes}{A vector of ENSG_IDs to be plotted. At the moment, gene names
are not supported.}

\item{type}{Character vector specifying the type of spatial data
to use for plotting. Options are "spot" for spot
geometry and "hex" for hexagon geometry. Default is
"spot".}

\item{colours}{One of either "viridis" or "custom". Select between Viridis'
package colour schemes and custom colours provided by you.
IMPORTANT: Look at Details for more information.}

\item{col_args}{List of arguments to customise the map fill colour.
Default is an empty list. If you select "viridis" at the
\code{colours} argument, here you can select the colour scheme
you prefer. I.e., \code{list(option = "magma")}.}

\item{gene_names}{A character vector with the corresponding gene names for
the ENSG IDs you provided in the \code{genes} argument.
IMPORTANT: the order of names must match that of the ENSG
IDs.}
}
\description{
This function plots leading scores for specific genes from a GWPCA object.
}
\details{
About using custom colours:
At the moment, the function is implementing \code{scale_fill_gradientn} which
expects a gradient like the ones generated by \code{colorRampPalette}. For
example: \code{colors <- colorRampPalette(c("blue", "red"))(100)}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{gwpca}
\keyword{leading-scores}
\keyword{map}
\keyword{plot}
\keyword{spatial}
