% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_plot.R
\name{plotGWPCA_ptv}
\alias{plotGWPCA_ptv}
\title{Plot violins and map for Percent of Total Variation.}
\usage{
plotGWPCA_ptv(
  gwpca,
  comps,
  type = c("violin", "map"),
  violin_args = list(),
  box_args = list(),
  fill_args = list(),
  lab_args = list(),
  ...
)
}
\arguments{
\item{gwpca}{a list of class \code{gwpca}, generated by \code{gwpca.ste} and
updated with leading genes by \code{gwpca.LeadingGene}.}

\item{comps}{the components to be plotted. The maximum is based on the max
number of components used in the \code{gwpca.propVar} function. If one
component is provided here, then only one violin plot or map will be plotted.
If it is a numeric vector then the corresponding plots will be plotted.}

\item{type}{either 'single' for single leading gene per location or 'multi'
for k leading genes per location.}

\item{violin_args}{a list of arguments to be passed to \code{geom_violin()}
for customizing the violin plots. Defaults to an empty list.}

\item{box_args}{a list of arguments to be passed to \code{geom_boxplot()} for
customizing the boxplots. Defaults to an empty list.}

\item{fill_args}{a list of arguments to be passed to
\code{scale_fill_manual()} for customizing the fill colors of the plots.
Defaults to an empty list.}

\item{lab_args}{a list of arguments to be passed to \code{labs()} for
customizing the plot labels. Defaults to an empty list.}

\item{...}{arguments to be passed to both \code{geom_sf()} outside of the
aesthetic.}
}
\value{
Returns a ggplot object.
}
\description{
A function to plot and map the Percent of Total Variation (PTV) of the local
Principal Components for each location.
}
\details{
The comps argument is passed down to select the aggregates of components to
be plotted. Remember that PTVs are cumulative. This means that Comps_01
inside \code{gwpca$ptvs} shows the PTV for PC 1 per location. Next, Comps_02
shows the cumulative PTV for PCs 1 and 2 per location etc. As a result, when
you give \code{comps} = c(1, 5) then the violin plots for the cumulative PTV of
the first and the first 5 PCs per location will be plotted or mapped. If you
provide a single number then a single violin plot or map will be plotted. If
you provide a vector of numbers then a plot with multiple violins (arrange =
TRUE), or multiple single violins, or a panel of maps will be plotted, or
multiple single maps.
}
\examples{
# Load required packages
library(ggplot2)

# Load data
data(gwpca)
data(sfe)

# Add PTV data
gwpca <- gwpca_PropVar(gwpca = gwpca, n_comp = 2:10, sfe = sfe)

# Plot PTV as violin plot
plotGWPCA_ptv(gwpca, comps = 1:5, type = "violin")

# Plot PTV as map
plotGWPCA_ptv(gwpca, comps = 1:5, type = "map")

}
\author{
Eleftherios (Lefteris) Zormpas
}
