% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_plot.R
\name{plotGWR}
\alias{plotGWR}
\title{Plot Geographically Weighted Regression (GWR) Results}
\usage{
plotGWR(gwr, predictorVar = NULL, globalModel = FALSE, title = NULL)
}
\arguments{
\item{gwr}{A GWR object returned by the GWR fitting function \code{gwrSTE}.}

\item{predictorVar}{A character string specifying the name of the predictor
variable to plot. If \code{NULL}, the function will automatically select the
first predictor variable from the model formula. If the model contains
multiple variables as predictors, only the first one will be used.}

\item{globalModel}{TRUE or FALSE. Indicates whether a table with the global
model's summary statistics has to be printed.}

\item{title}{A character string for the plot's title. Defaults to NULL where
a title made from the formula is used.}
}
\value{
A \code{ggplot} object containing the plot of GWR coefficients and a
summary table.
}
\description{
This function generates a plot that visualises the results of a
Geographically Weighted Regression (GWR) model. The function plots
the coefficients of the GWR model, highlighting significant areas
based on t-values, and includes a summary table of the model's coefficients.
}
\details{
The generated plot shows the spatial distribution of the GWR coefficients
and their significance. Areas where the coefficients are significant are
highlighted. The plot also includes a summary table of the global model's
intercept and covariate coefficients, t-values, and p-values to allow the
user to compare results with GWR (the mapped coefficients).
}
\examples{
\dontrun{
pltGWR(gwr_object)
}


}
\author{
Eleftherios (Lefteris) Zormpas
}
