% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_plot.R
\name{plotGWR_R2}
\alias{plotGWR_R2}
\title{Plot Local R-squared from GWR Model}
\usage{
plotGWR_R2(gwr)
}
\arguments{
\item{gwr}{A GWR object returned by the GWR fitting function \code{gwrSTE}.}
}
\value{
A \code{ggplot} object representing the spatial distribution of the
local R-squared values.
}
\description{
This function generates a plot of the local R-squared values from a
Geographically Weighted Regression (GWR) model. The plot provides a
spatial visualization of the goodness of fit across the study area.
}
\examples{
\dontrun{
plotGWR_R2(gwr_object)
}

}
\author{
Eleftherios (Lefteris) Zormpas
}
