% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_plot.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot a Heatmap of Gene Expression Data}
\usage{
plotHeatmap(
  m_sfe,
  sample_id = NULL,
  assay = "logcounts",
  subset_col = NULL,
  subset_row = NULL,
  loc_annot = list(),
  order_col = NULL,
  scale = "row",
  ...
)
}
\arguments{
\item{m_sfe}{The \code{SpatialFeatureExperiment} or
\code{MetaSpatialFeatureExperiment} object containing spatial expression
data.}

\item{sample_id}{A character string indicating the name of the sample. It is
required when the \code{m_sfe} argument is provided with a MetaSFE object.
Defaults to \code{NULL} which selects the first sample from the \code{m_sfe} object.}

\item{subset_col}{subset cols (locations) using a boolean vector of length
equal to the number of locations.}

\item{subset_row}{subset rows (genes) using a boolean vector of length
equal to the number of genes.}

\item{loc_annot}{a named list of annotations for the columns (locations). If
left empty, the function looks if there is an \code{annotation} column
in the \code{colData}. If there is not, then, no annotations are added.}

\item{order_col}{order the columns (locations) based on the alphabetical
order of the values provided in the \code{loc_annot} argument. The value needs to
be a character string of the name of the annotation to be used for ordering.
This name MUST match the names in the named list provided in the \code{loc_annot}
argument. Defaults to \code{NULL} which leads to columns being clustered
by \code{pheatmap}.}

\item{scale}{character indicating if the values should be centred and scaled
in either the row direction or the column direction, or none. Corresponding
values are "row", "column" and "none". (this is a \code{pheatmap} parameter)}

\item{...}{Additional arguments passed to \code{pheatmap}.}
}
\value{
Plots a heatmap using \code{pheatmap}. It can be saved as a \code{pheatmap}
object.
}
\description{
This function plots a heatmap using the \code{pheatmap} package, providing
options to subset and annotate the data.
}
\details{
The function assumes that the annotations provided in the \code{loc_annot}
argument are in the same order as the locations within the \code{SFE} object.
Please make sure this is the case.

The named list provided in the \code{loc_annot} argument MUST have the values as
vectors (i.e., \code{list(annot = vector_of_annotations)}). The
\code{vector_of_annotations} in the above example MUST have the same length as
the number of locations in the \code{SFE} object.
}
\examples{
\dontrun{
# Assuming `sfe` is a valid SFE object
# Assuming there is also cluster information from running `fgwc_STE`

# Preparing the annotations
loc_annots <- list(cluster = fgwc$cluster,
                   annotation = colData(sfe)$annotation)
# The values provided in the list MUST be vectors.

# Preparing to order the columns based on cluster number
order_cols <- "cluster"

plotHeatmap(sfe,
            loc_annot = loc_annots,
            order_col = order_cols)
}

}
\author{
Eleftherios (Lefteris) Zormpas
}
