% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NeighbourGraph.R
\name{plotNeighbourGraph}
\alias{plotNeighbourGraph}
\title{Visualise Neighbour Graphs and Associated Images}
\usage{
plotNeighbourGraph(
  msfe,
  sample_id = TRUE,
  res = c("lowres", "hires", "fullres"),
  plotImage = TRUE,
  ...
)
}
\arguments{
\item{msfe}{The spatial transcriptomics data as a
MetaSpatialFeatureExperiment object.}

\item{sample_id}{A character string, \code{TRUE}, or \code{NULL} specifying
sample/image identifier(s). If \code{TRUE}, all samples/images are
considered. If \code{NULL}, the first available entry is considered.}

\item{res}{Character vector specifying the resolution of the images to be
used ('lowres', 'hires', 'fullres').}

\item{plotImage}{Logical. Whether to overlay image data on the neighbour
graphs.}

\item{...}{Additional parameters passed to internal functions.}
}
\value{
A grid of neighbour graphs, with or without overlaid image data.
}
\description{
This function generates a grid of neighbour graphs for spatial
transcriptomics data, optionally overlaying image data.
}
\details{
The function creates a grid of neighbour graphs for spatial
transcriptomics data. Each graph represents the spatial relationships
between spots. The optional overlay of image data provides context to the
spatial distribution.
}
\examples{
\dontrun{
# Example usage:
data(msfe_example)
plotNeighbourGraph(msfe_example, sample_id = "JBO019",
                   res = 'lowres', plotImage = TRUE)
}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{graph,}
\keyword{grid,}
\keyword{image}
\keyword{neighbour}
\keyword{overlay}
\keyword{spatial}
\keyword{transcriptomics,}
