% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC_.R
\name{plotQC_filtered}
\alias{plotQC_filtered}
\title{Plot Quality Control Filtered Locations Map in a SpatialFeatureExperiment}
\usage{
plotQC_filtered(
  sfe,
  metric = c("libsize", "mito", "NAs", "detected", "cellCount", "discard"),
  sample_id = TRUE
)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object.}

\item{metric}{Character vector specifying the metric, either "libsize",
"mito", "NAs", "detected", "cellCount", or "discard".}

\item{sample_id}{Character string, TRUE, or NULL specifying sample/image
identifier(s). TRUE is equivalent to all samples/images, and NULL specifies
the first available entry.}
}
\description{
This function plots the quality control filtered locations map in a
SpatialFeatureExperiment, providing options for different metrics and
sample IDs.
}
\details{
This function plots the quality control filtered locations map in a
SpatialFeatureExperiment. Users can specify the metric (libsize, mito, NAs,
detected, cellCount, or discard) and sample ID.
}
\examples{
\dontrun{
# Example usage:
data(sfe)
plotQC_filtered(sfe, metric = "libsize")
}

}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{filtering}
\keyword{ggplot2}
\keyword{mapping}
\keyword{quality-control}
\keyword{spatial}
