% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC_.R
\name{plotQC_hist}
\alias{plotQC_hist}
\title{Plot Quality Control Histogram in a SpatialFeatureExperiment}
\usage{
plotQC_hist(
  sfe,
  limits,
  metric = c("libsize", "mito", "detected", "cellCount"),
  sample_id = TRUE,
  hist_args = list(),
  dens_args = list(),
  vline_args = list(),
  ...
)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object.}

\item{limits}{Numeric vector specifying the limits of the x-axis.}

\item{metric}{Character vector specifying the metric, either "libsize",
"mito", "detected", "sizeFact", or "cellCount".}

\item{sample_id}{Character string, TRUE, or NULL specifying sample/image
identifier(s). TRUE is equivalent to all samples/images, and NULL specifies
the first available entry.}

\item{hist_args}{List of arguments to customize the histogram. Default is an
empty list.}

\item{dens_args}{List of arguments to customize the density plot. Default is
an empty list.}

\item{vline_args}{List of arguments to customize the vertical line. Default
is an empty list.}

\item{...}{Additional arguments to be passed to the underlying plotting
functions.}
}
\value{
A ggplot object representing the quality control histogram.
}
\description{
This function plots the quality control histogram in a
SpatialFeatureExperiment, providing options for different metrics,
sample IDs, and customisation.
}
\details{
This function plots the quality control histogram in a
SpatialFeatureExperiment. Users can specify the metric (libsize, mito,
detected, sizeFact, or cellCount), sample ID, and customize the appearance
of the histogram, density plot, and vertical line.
}
\examples{
\dontrun{
# Example usage:
data(sfe)
plotQC_hist(sfe, limits = c(0, 5000), metric = "libsize")
}

}
\seealso{
Explore the related internal functions: \code{\link{.int_getSmplIDs}}.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{ggplot2}
\keyword{histogram}
\keyword{plotting}
\keyword{quality-control}
\keyword{spatial}
