% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{plotQC_scat}
\alias{plotQC_scat}
\title{Plot Quality Control Scatter Plot in a SpatialFeatureExperiment}
\usage{
plotQC_scat(
  m_sfe,
  limits,
  metric = c("libsize", "mito", "detected", "cellCount"),
  sample_id = TRUE,
  point_args = list(),
  hline_args = list(),
  ...
)
}
\arguments{
\item{m_sfe}{A SpatialFeatureExperiment or MetaSFE object.}

\item{limits}{Numeric vector specifying the limits of the x and y axes.}

\item{metric}{Character vector specifying the metric, either "libsize",
"mito", or "detected".}

\item{sample_id}{Character string, TRUE, or NULL specifying sample/image
identifier(s). TRUE is equivalent to all samples/images, and NULL specifies
the first available entry.}

\item{point_args}{List of arguments to customize the scatter points.
Default is an empty list.}

\item{hline_args}{List of arguments to customize the horizontal line.
Default is an empty list.}

\item{...}{Additional arguments to be passed to the underlying plotting
functions.}
}
\value{
A ggplot object representing the quality control scatter plot.
}
\description{
This function plots the quality control scatter plot in a
SpatialFeatureExperiment, providing options for different metrics,
sample IDs, and customisation.
}
\details{
This function plots the quality control scatter plot in a
SpatialFeatureExperiment. Users can specify the metric (libsize, mito, or
detected), sample ID, and customize the appearance of the scatter points
and horizontal line.
}
\examples{
\dontrun{
# Example usage:
data(sfe)
plotQC_scat(sfe, limits = c(0, 5000), metric = "libsize")
}

}
\seealso{
Explore the related internal functions: \code{\link{.int_getSmplIDs}}.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{ggplot2}
\keyword{plotting}
\keyword{quality-control}
\keyword{scatter-plot}
\keyword{spatial}
