% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{plotQC_sizeFactors}
\alias{plotQC_sizeFactors}
\title{Plot Size Factors Histogram and Density in a SpatialFeatureExperiment}
\usage{
plotQC_sizeFactors(
  m_sfe,
  sample_id = TRUE,
  hist_args = list(),
  dens_args = list(),
  ...
)
}
\arguments{
\item{m_sfe}{A SpatialFeatureExperiment or MetaSFE object.}

\item{sample_id}{Character string, TRUE, or NULL specifying sample/image
identifier(s). TRUE is equivalent to all samples/images, and NULL specifies
the first available entry.}

\item{hist_args}{List of arguments to customize the histogram. Default is
an empty list.}

\item{dens_args}{List of arguments to customize the density plot. Default is
an empty list.}

\item{...}{Additional arguments to be passed to the underlying plotting
functions.}
}
\value{
A ggplot object representing the size factors histogram and density.
}
\description{
This function plots the size factors histogram and density in a
SpatialFeatureExperiment, providing options for sample IDs and customisation.
}
\details{
This function plots the size factors histogram and density in a
SpatialFeatureExperiment. Users can specify the sample ID and customise the
appearance of the histogram and density plot.
}
\examples{
\dontrun{
# Example usage:
data(sfe)
plotQC_sizeFactors(sfe)
}

}
\seealso{
Explore the related internal functions: \code{\link{.int_getSmplIDs}}.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{density}
\keyword{ggplot2}
\keyword{histogram}
\keyword{plotting}
\keyword{size-factors}
\keyword{spatial}
