% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{plotQC_spots}
\alias{plotQC_spots}
\title{Plot Quality Control of Spots in a SpatialFeatureExperiment}
\usage{
plotQC_spots(
  m_sfe,
  type = c("spot", "hex", "cntd"),
  sample_id = NULL,
  in_tissue = TRUE,
  colours = NULL,
  y_reverse = FALSE
)
}
\arguments{
\item{m_sfe}{A SpatialFeatureExperiment or MetaSFE object.}

\item{type}{Character vector specifying the spot type, either "spot", "cntd"
or "hex".}

\item{sample_id}{Character string, TRUE, or NULL specifying sample/image
identifier(s). TRUE  equivalent to all samples/images, and NULL specifies
the first available entry.}

\item{in_tissue}{Logical, indicating whether to consider only spots in
tissue. Default is TRUE.}

\item{colours}{Vector of colours for spots. Default is NULL, which uses the
default colours.}

\item{y_reverse}{Logical, indicating whether to reverse the y-axis. Default
is FALSE.}
}
\value{
A ggplot object representing the quality control plot.
}
\description{
This function plots the quality control of spots in a
SpatialFeatureExperiment, providing options for different spot types and
customization.
}
\details{
This function plots the quality control of spots in a
SpatialFeatureExperiment. It provides options for different spot types
(spot or hexagon or centroid) and customisation. The in_tissue argument
controls whether to consider only spots in tissue. The colours argument
allows users to customise spot colours, and y_reverse can be used to reverse
the y-axis.
}
\examples{
# Example usage:
data(sfe)
plotQC_spots(sfe, type = "spot", sample_id = TRUE, in_tissue = TRUE,
             colours = NULL, y_reverse = FALSE)

}
\seealso{
Explore the related internal functions: \code{\link{.int_getSmplIDs}},
\code{\link{.int_plotSpot}}, and \code{\link{.int_plotHex}}.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{ggplot2}
\keyword{plotting}
\keyword{quality-control}
\keyword{spatial}
