% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC_.R
\name{plotQC_tissueImg}
\alias{plotQC_tissueImg}
\title{Plot Tissue Image in a SpatialFeatureExperiment}
\usage{
plotQC_tissueImg(
  sfe,
  res = c("lowres", "hires", "fullres"),
  type = c("spot", "hex", "none"),
  sample_id = NULL,
  fill_args = list(),
  annotate = FALSE,
  colours = NULL,
  alpha = 0.3,
  ...
)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object.}

\item{res}{Character vector specifying the image resolution, either
"lowres", "hires", or "fullres".}

\item{type}{Character vector specifying the spot type, either "spot",
"hex", or "none".}

\item{sample_id}{Character string, TRUE, or NULL specifying sample/image
identifier(s). TRUE is equivalent to all samples/images, and NULL specifies
the first available entry.}

\item{fill_args}{List of arguments to customize the fill scale. Default is
an empty list.}

\item{annotate}{Logical, indicating whether to annotate the plot. Default
is FALSE.}

\item{colours}{Vector of colours for annotation. Default is NULL, which uses
default colours.}

\item{alpha}{Numeric, indicating the alpha (transparency) value for the
spots and the annotation (if selected). Default is 0.3.}

\item{...}{Additional arguments to be passed to the underlying plotting functions.}
}
\description{
This function plots the tissue image in a SpatialFeatureExperiment,
providing options for different resolutions and spot types.
}
\details{
This function plots the tissue image in a SpatialFeatureExperiment. Users
can specify the image resolution, spot type, and customize the fill scale
for annotation.
The annotate argument controls whether to annotate the plot, and colours
can be used to customise annotation colours. The alpha argument adjusts the
transparency of the spots overlaying the image.
}
\examples{
\dontrun{
# Example usage:
data(sfe)
# With annotations
plotQC_tissueImg(sfe, res = "lowres", type = "spot", annotate = TRUE)

# Without annotations
plotQC_tissueImg(sfe, res = "lowres", type = "spot", annotate = FALSE)
}

}
\seealso{
Explore the related internal functions: \code{\link{.int_getImgDt}},
\code{\link{.int_getImgLims}}, and \code{\link{.int_tissueImgPlot}}.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{annotation}
\keyword{ggplot2}
\keyword{plotting}
\keyword{quality-control}
\keyword{spatial}
