% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_plot.R
\name{plotSA_globalSum}
\alias{plotSA_globalSum}
\title{Title: Visualise Global Spatial Autocorrelation result summary}
\usage{
plotSA_globalSum(
  m_sfe,
  sample_id = NULL,
  genes,
  statistic = c("moran", "getis", "geary"),
  ylab = "gene_name",
  moran_thresh = c(-0.5, 0.5),
  geary_thresh = c(0.5, 1.5),
  getis_thresh = 2,
  ...
)
}
\arguments{
\item{m_sfe}{An object of class 'SpatialFeatureExperiment' or
'MultiSpatialFeatureExperiment' containing spatial data.}

\item{sample_id}{A character string specifying the sample ID for analysis.}

\item{genes}{A character vector containing a selection of genes. It must be
either EnsgIDs OR gene names. A mix of both will throw an error.}

\item{statistic}{A character vector specifying which spatial autocorrelation
statistics to plot. Valid options are "moran", "getis", and
"geary". Default is c("moran", "getis", "geary").}

\item{ylab}{A character string indicating the label for the y-axis. It can
be "gene_name" to use gene names or any other value to use
Ensemble Gene IDs. Default is "gene_name".}

\item{moran_thresh}{A numeric vector of length 2 specifying the threshold
range for Moran's I values to be considered significant.
Default is c(-0.5, 0.5).}

\item{geary_thresh}{A numeric vector of length 2 specifying the threshold
range for Geary's C values to be considered significant.
Default is c(0.5, 1.5).}

\item{getis_thresh}{A numeric value specifying the threshold for Getis-Ord's
G statistic to be considered significant. Default is 2.}

\item{...}{Additional arguments passed on to \code{geom_point()} for customising
the point aesthetics in the ggplot.}
}
\value{
A ggplot object representing the dot plot of global spatial
autocorrelation statistics for the specified genes across the given
statistics.
}
\description{
Description: This function plots a dot plot of global SA values for selected
genes.
}
\details{
The function creates a faceted dot plot where each facet corresponds to one
of the specified spatial autocorrelation statistics. Points represent genes,
and their position and colour indicate the statistic value. Point size shows
the significance of the statistic, while the shape indicates whether the
statistic suggests positive or negative spatial autocorrelation.
}
\examples{
plotSA_globalSum(m_sfe, "sample1", c("Gene1", "Gene2"), c("moran", "getis"))

}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{autocorrelation}
\keyword{global}
\keyword{spatial}
\keyword{visualisation}
