% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_plot.R
\name{plotSA_local}
\alias{plotSA_local}
\title{Title: Visualise Local Spatial Autocorrelation}
\usage{
plotSA_local(
  m_sfe,
  sample_id = NULL,
  feature,
  statistic = c("moran", "geary", "getis"),
  test = c("z-score", "permutation"),
  pVal = 0.05,
  type = c("spot", "hex"),
  viridis_col = "inferno",
  signif_col = "#E0E0E0",
  title = c("ENSGID", "name"),
  locations = "all"
)
}
\arguments{
\item{m_sfe}{An object of class 'SpatialFeatureExperiment' or
'MultiSpatialFeatureExperiment' containing spatial data.}

\item{sample_id}{A character string specifying the sample ID for analysis.}

\item{feature}{A character string indicating the feature of interest.
It can be either ENSGene IDs or Gene Names.}

\item{statistic}{A character string specifying the type of spatial
autocorrelation results ("moran", "geary", "getis") to look
for in the m_sfe object for the specified sample.}

\item{test}{A character string specifying the statistical test method
("z-score" or "permutation") to look for in the m_sfe object
for the specified sample.}

\item{pVal}{A numeric value representing the significance threshold for
plotting. Ignored for Geary's C, where permutation is suggested.
See details for more information.}

\item{type}{A character string specifying the type of geometries ("spot" or
"hex") to use in the plot.}

\item{viridis_col}{A character string specifying the colour palette for
the autocorrelation values. Currently supporting only the
viridis palettes.}

\item{signif_col}{A character string specifying the colour for the outline
of significant locations in the plot. Used only when the
\code{locations} argument is set to \code{"all"}. It helps to
identify which SA values are also statistically
significant.}

\item{title}{A character string ("ENSGID" or "name") specifying whether the
title should show the ENSGene ID or the gene name. Default is
"name".}

\item{locations}{A character string ("all" or "significant") specifying
which locations to include in the plot. If \code{"all"} is
selected, then all spots are included and the locations
where the local SA is significant has a differently
coloured outline as set by the \code{signif_col} argument. If
\verb{"significant} is selected, then only the locations with
statistically significant SA values are plotted while the
rest of the locations are greyed out.}
}
\value{
A ggplot object visualizing local spatial autocorrelation.
}
\description{
Description: Generates a map visualizing local spatial autocorrelation
values for a given feature in the input spatial feature experiment.
}
\details{
Creates an informative plot of local spatial autocorrelation values.

Internally this function checks the compatibility between selected statistic
and statistical significance method. Geary's C in particular returns no
Z-score statistic. As a result, when someone is trying to plot Geary's C the
\code{pVal} argument is not used. A message will be printed to
inform for this and also to suggest using \code{permutation} in the \code{test}
argument after running the \code{gearyLocalCPerm} function first.
}
\examples{
plotSA_local(m_sfe, "sample1", "gene_expression", "moran", "z-score",
0.05, "spot")

}
\seealso{
See also the \code{plotSA_localClust} function for clustering analysis.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{autocorrelation,}
\keyword{spatial,}
\keyword{visualisation}
