% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_plot.R
\name{plotSA_localClust}
\alias{plotSA_localClust}
\title{#' Title: Visualise Clusters in Local Spatial Autocorrelation}
\usage{
plotSA_localClust(
  m_sfe,
  sample_id = NULL,
  feature,
  statistic = c("moran", "geary", "getis"),
  test = c("z-score", "permutation"),
  pVal = 0.05,
  type = c("spot", "hex"),
  title = c("ENSGID", "name"),
  clust_col = NULL,
  signif_col = "#E0E0E0",
  locations = "all"
)
}
\arguments{
\item{m_sfe}{An object of class 'SpatialFeatureExperiment' or
'MultiSpatialFeatureExperiment' containing spatial data.}

\item{sample_id}{A character string specifying the sample ID for analysis.}

\item{feature}{A character string indicating the feature of interest.
Currently supports only ENSGene IDs.}

\item{statistic}{A character string specifying the type of spatial
autocorrelation results ("moran", "geary", "getis") to look
for in the m_sfe object for the specified sample.}

\item{test}{A character string specifying the statistical test method
("z-score" or "permutation") to look for in the m_sfe object
for the specified sample.}

\item{pVal}{A numeric value representing the significance threshold for
plotting. Ignored for Geary's C, where permutation is suggested.}

\item{type}{A character string specifying the type of geometries ("spot" or
"hex") to use in the plot.}

\item{title}{A character string ("ENSGID" or "name") specifying whether the
title should show the ENSGene ID or the gene name.}

\item{clust_col}{A named character vector of colours to be used for
plotting. The names must match the cluster names. Look at
details below for more information about what these names
should be. Default is \code{NULL} and the default colours are
applied.}

\item{signif_col}{A character string specifying the colour for the outline
of significant locations in the plot. Used only when the
\code{locations} argument is set to \code{"all"}. It helps to
identify which SA values are also statistically
significant.}

\item{locations}{A character string ("all" or "significant") specifying
which locations to include in the plot. If \code{"all"} is
selected, then all spots are included and the locations
where the local SA is significant has a differently
coloured outline as set by the \code{signif_col} argument. If
\verb{"significant} is selected, then only the locations with
statistically significant SA values are plotted while the
rest of the locations are greyed out.}
}
\value{
A ggplot object visualising clusters in local spatial
autocorrelation.
}
\description{
Description: Generates a map visualising clusters in local spatial
autocorrelation values for a given feature in the input spatial feature
experiment.
}
\details{
Creates an informative plot of local spatial autocorrelation
values.
The named vector with colours provided in the \code{clust_col} argument
should include the below names:
"Low", "High", "low-Low", "High-Low", "Low-High", "High-High",
"Not Signif."
If you don't provide a vector with the exact names then this
function will produce an error.
}
\examples{
plotSA_localClust(m_sfe, "sample1", "gene_expression", "moran", "z-score",
0.05, "spot")

}
\seealso{
See also the 'plotSA_local' function for local autocorrelation
visualization.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{autocorrelation,}
\keyword{clustering,}
\keyword{spatial,}
\keyword{visualisation}
