% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_STE.R
\name{print.gwpca}
\alias{print.gwpca}
\title{Print GWPCA output}
\usage{
\method{print}{gwpca}(x, ...)
}
\arguments{
\item{x}{An object of class 'gwpca'.}

\item{...}{Not used currently. Left for future additions.}
}
\description{
A function to print GWPCA results in the prompt
}
\author{
Eleftherios (Lefteris) Zormpas
}
