% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpcaSTE.R
\name{progrFUN}
\alias{progrFUN}
\title{Progress function for foreach loop.}
\usage{
progrFUN(foreach, FUN, ForEachArgs = list(), ...)
}
\arguments{
\item{foreach}{Numerical vector to iterate over.}

\item{FUN}{Function to apply to each iteration.}

\item{ForEachArgs}{Additional arguments to pass to the foreach loop.}

\item{...}{Additional arguments passed to the FUN function.}
}
\value{
Returns the result of applying the FUN function to each iteration.
}
\description{
IMPORTANT: this function is not exported to be used alone.
}
\details{
This function is used as a progress function for the foreach loop.
It displays the progress of the loop iteration using the progressor function.
}
\seealso{
\code{\link{progressor}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
