% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_PropVar.R
\name{prop.var}
\alias{prop.var}
\title{Internal: calculate proportion of variance}
\usage{
prop.var(n_components, gwpca_var)
}
\arguments{
\item{n_components}{Numeric vector specifying the number of principal
components.}

\item{gwpca_var}{Matrix containing the variance values from a gwpca object.}
}
\value{
Numeric vector representing the proportion of variance explained by
the specified number of principal components.
}
\description{
IMPORTANT: this function is not exported to be used alone.
The \code{prop.var} function calculates the proportion of variance explained by
the specified number of principal components. It takes a numeric vector
\code{n_components} indicating the number of principal components for which the
variance is calculated. The variance values are obtained from the \code{gwpca_var}
matrix, which is typically the \verb{$var} slot of a gwpca object. The resulting
proportion of variance explained is expressed as a percentage.
}
\seealso{
Other functions related to principal component analysis: \code{gwpca}, \code{pca},
\code{prcomp}, \code{princomp}
}
\author{
Eleftherios (Lefteris) Zormpas
}
