% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readCurioSeeker}
\alias{readCurioSeeker}
\title{Read data from a Curio Seeker (Slide-seq) experiment}
\usage{
readCurioSeeker(samples, sample_id, fileType = c("AnnData", "Seurat", "CSV"))
}
\arguments{
\item{samples}{a character vector specifying one or more directories, each
corresponding to a 10x Genomics Visium sample (see Details); if provided,
names will be used as sample identifiers}

\item{sample_id}{character string specifying unique sample identifiers, one
for each directory specified via samples; ignored if !is.null(names(samples))}

\item{fileType}{character string specifying the type of format to read count
data from. One of (AnnData, Seurat, CSV).}
}
\value{
A \code{SpatialFeatureExperiment} object containing the spatial
transcriptomics data with counts, spatial coordinates, and feature
data.
}
\description{
Reads spatial transcriptomics data from specified directory paths and
formats it into a \code{SpatialFeatureExperiment} object. This function supports
data in AnnData, Seurat, or CSV formats.
}
\details{
The function processes directories specified in the \code{samples} parameter, each
expected to represent a sample in the spatial transcriptomics experiment.
Based on the \code{fileType} parameter, the function looks for specific data
files:
\itemize{
\item \strong{AnnData}: Looks for \code{.h5ad} files containing the count
matrix, feature data, and spatial coordinates.
\item \strong{Seurat}: Expects \code{.rds} files which are R serialised objects
of Seurat objects containing similar data as above.
\item \strong{CSV}: Reads raw CSV files for counts, and TSV
files for barcodes and features, along with a CSV for spatial
coordinates. The counts csv MUST include \code{counts.csv} in its name.
}

For \code{AnnData} format, the function  uses the \code{anndata} package to load the
data.

For \code{Seurat} format, it expects a Seurat object with counts in
\code{obj@assays$RNA@counts}, and spatial coordinates in
\code{obj@images$slice1@coordinates}.

For \code{CSV} format, it expects files that contain a \code{counts.csv} in their name
for the counts matrix, \code{barcodes.tsv} for barcode names, \code{genes.tsv} for
feature names, and \code{MatchedBeadLocation.csv} for spatial coordinates.
}
\examples{
\dontrun{
  sfe <- readCurioSeeker(samples = c("sample1/", "sample2/"),
                         sample_id = c("sample1", "sample2"),
                         fileType = "AnnData")
}

}
