% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_STE.R
\name{robustSvd.ste}
\alias{robustSvd.ste}
\title{A robust SVD function.}
\usage{
robustSvd.ste(x, alpha = 0.75)
}
\arguments{
\item{x}{Data matrix.}

\item{alpha}{Tuning parameter for the MCD estimator (default = 0.75).}
}
\value{
Returns a list with the loadings (v) and variances (d) obtained from
the SVD.
}
\description{
IMPORTANT: this function is not exported to be used alone.
Performs a robust Singular Value Decomposition (SVD) using the MCD estimator.
}
\details{
This function performs a robust Singular Value Decomposition (SVD) using the
Minimum Covariance Determinant (MCD) estimator. The MCD estimator is used to
estimate the covariance matrix of the data, and then the SVD is applied to
the estimated covariance matrix. The loadings (v) and variances (d) obtained
from the SVD are returned. The true variances (d) are returned, unlike in the
basic GWPCA case, i.e., d=(pc$sdev)^2.
}
\seealso{
\code{\link{princomp}}, \code{\link{covMcd}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
