% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpcaSTE.R
\name{rwpca.ste}
\alias{rwpca.ste}
\title{A robust WPCA.}
\usage{
rwpca.ste(x, wt, nu = 0, nv = 2)
}
\arguments{
\item{x}{Data matrix.}

\item{wt}{Weight vector.}

\item{nu}{Number of observations to use as initial estimates (default = 0).}

\item{nv}{Number of components to retain (default = 2).}
}
\value{
Returns a list with the loadings (v) and variances (d) obtained from
the WPCA.
}
\description{
IMPORTANT: this function is not exported to be used alone.
Performs a robust Weighted Principal Component Analysis (WPCA) with distance
weighting.
}
\details{
This function performs a robust Weighted Principal Component Analysis (WPCA)
with distance weighting. It calculates the medians of each column of the data
matrix and subtracts them from the data to centre it. The centred matrix is
then multiplied by the weight vector to apply the distance weighting. The
resulting weighted matrix is subjected to a robust Singular Value
Decomposition (SVD) using the \code{robustSvd.ste} function. The loadings (v) and
variances (d) obtained from the SVD are returned, with the number of
components retained determined by the \code{nv} parameter.
}
\seealso{
\code{\link{robustSvd.ste}}, \code{\link{wt.median.ste}},
\code{\link{princomp}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
