% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{set.parallel}
\alias{set.parallel}
\title{Chekc and set backend}
\usage{
set.parallel(.strategy, .workers)
}
\arguments{
\item{.strategy}{The strategy to be used for parallel processing. If not
specified or set to "sequential", the function will run sequentially.}

\item{.workers}{The number of workers to be used for parallel processing.
Ignored if the strategy is set to "sequential".}
}
\description{
IMPORTANT: this function is not exported to be used alone.
It is used internally to check and set a backend for the "gwcSTE" and
"gwpcaSTE" functions.
}
\details{
This function sets the parallel processing backend based on the specified
strategy. If the strategy is set to "sequential" or if no strategy is
specified, the function will run sequentially using the "registerDoSEQ"
backend. If a different strategy is specified, the function will set the
future plan and register the appropriate backend using the "registerDoFuture"
function from the "doFuture" package. The number of workers can also be
specified, which determines the level of parallelism.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
