% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.parallelPlan.R
\name{set.parallelPlan}
\alias{set.parallelPlan}
\title{Set parallel plan}
\usage{
set.parallelPlan(..strategy, ..workers, ...)
}
\arguments{
\item{..strategy}{The strategy for parallel processing. It can be one of the
following: "multicore", "multisession", "cluster", "sequential",
or "clustAuto".}

\item{..workers}{The number of workers to be used for parallel processing. If
the strategy is set to "cluster", this should be a cluster object. Otherwise,
it should be a numeric value specifying the number of workers. For the
"clustAuto" strategy, this parameter is ignored.}

\item{...}{Further arguments passed to \code{\link{plan}} function.}
}
\description{
IMPORTANT: this function is not exported to be used alone.
It is used internally to check the parallelisation parameters and set up the
environment with the future package.
}
\details{
This function checks the specified parallelisation parameters and sets up the
environment with the future package. It verifies the availability of cores,
checks the strategy and workers, and sets the future plan accordingly. If the
strategy is set to "cluster", a valid cluster object should be provided as
the workers argument. For other strategies, a numeric value specifying the
number of workers should be provided. The function provides informative
messages about the chosen strategy and the number of cores to be used.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
