% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setQCthresh.R
\name{setQCthresh_CellCount}
\alias{setQCthresh_CellCount}
\title{Set QC Threshold for Cell Counts}
\usage{
setQCthresh_CellCount(sfe, sample_id = TRUE, min_t, max_t)
}
\arguments{
\item{sfe}{A SpatialFeatureExoeriment object containing cell count data.}

\item{sample_id}{A character string, \code{TRUE}, or \code{NULL} specifying
sample/image identifier(s). If \code{TRUE}, all samples/images are
considered. If \code{NULL}, the first available entry is considered.}

\item{min_t}{Numeric, the minimum threshold for cell counts. Locations with
counts below this threshold will be flagged. Use NA for no minimum threshold.}

\item{max_t}{Numeric, the maximum threshold for cell counts. Locations with
counts above this threshold will be flagged. Use NA for no maximum threshold.}
}
\value{
A modified SpatialFeatureExoeriment object with an added column in
colData indicating whether each location passes the QC thresholds for cell
counts.
}
\description{
This function sets quality control (QC) thresholds for cell counts based on
the values in the cellCount column.
}
\details{
The function sets QC thresholds for cell counts based on the
specified minimum and maximum values. Locations failing to meet the
thresholds are flagged, and a summary table of filtered locations is printed.
}
\examples{
# 'sfe' is a SpatialFeatureExoeriment object

# Set QC thresholds with sample IDs for cell counts above 15
sfe <- setQCthresh_CellCount(sfe, sample_id = TRUE, max_t = 15)

}
\seealso{
\code{\link{.int_getSmplIDs}}, \code{\link[SpatialFeatureExperiment]{colData}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{QC,}
\keyword{SpatialFeatureExoeriment}
\keyword{cell}
\keyword{counts,}
