% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setQCthresh.R
\name{setQCthresh_custom}
\alias{setQCthresh_custom}
\title{Set QC Thresholds using Custom Metrics}
\usage{
setQCthresh_custom(msfe, sample_id = TRUE, MARGIN, qcMetric)
}
\arguments{
\item{msfe}{A SpatialFeatureExperiment or a MetaSpatialFeatureExperiment
object.}

\item{sample_id}{A character string, \code{TRUE}, or \code{NULL} specifying
sample/image identifier(s). If \code{TRUE}, all samples/images are
considered. If \code{NULL}, the first available entry is considered.}

\item{MARGIN}{Numeric, indicating whether to apply the custom metric along
rows (1) or columns (2). If rows is selected then the filters are applied to
genes and the flagged genes are stored in a column in rowData. If columns is
selected then the filters are applied to locations and the flagged locations
are stored in a column in colData.}

\item{qcMetric}{A vector of TRUE and FALSE for genes or locations that are
going to be discarded. You need to have prepared this vector in advance. The
vector needs to be of length equal to nrow(rowData(sfe)) if MARGIN == 1, or
equal to nrow(colData(sfe)) if MARGIN == 2.}
}
\value{
A SpatialFeatureExperiment or a MetaSpatialFeatureExperiment
object with added columns indicating whether each gene or location passes
the specified QC thresholds using custom metrics.
}
\description{
This function sets quality control (QC) thresholds using custom metrics for
gene expression data or for location data.
}
\details{
The function sets QC thresholds for gene expression data using
custom metrics provided in the \code{qcMetric} vector. The user can specify
whether to apply the custom metric along rows (genes) or columns (samples)
using the \code{MARGIN} argument. Genes or locations failing to meet the
thresholds are flagged.
}
\examples{
# 'msfe' is a SpatialFeatureExoeriment or a MetaSpatialFeatureExoeriment,
# object

# Set QC thresholds using custom metrics for specific samples
qcMetric <- c(TRUE, FALSE, TRUE, FALSE, TRUE, FALSE, TRUE, FALSE, TRUE)
msfe <- setQCthresh_custom(msfe,
                           sample_id = c("JBO019"),
                           MARGIN = 1,
                           qcMetric)

}
\seealso{
\code{\link{.int_custom_1}}, \code{\link{.int_custom_2}},
\code{\link[SpatialFeatureExperiment]{colData}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{QC,}
\keyword{SummarizedExperiment}
\keyword{expression,}
\keyword{gene}
