% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_setQCthresh.R
\name{setQCthresh_GenesExpr}
\alias{setQCthresh_GenesExpr}
\title{Set QC Threshold for Genes Expressed in each location}
\usage{
setQCthresh_GenesExpr(sfe, sample_id = TRUE, min_t, max_t)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object containing gene expression data.}

\item{sample_id}{A character string, \code{TRUE}, or \code{NULL} specifying
sample/image identifier(s). If \code{TRUE}, all samples/images are
considered. If \code{NULL}, the first available entry is considered.}

\item{min_t}{Numeric, the minimum threshold for detection. Locations with
number of expressed genes below this threshold will be flagged. Use NA for
no minimum threshold.}

\item{max_t}{Numeric, the maximum threshold for detection. Locations with
number of expressed genes above this threshold will be flagged. Use NA for
no maximum threshold.}
}
\value{
A modified SpatialFeatureExperiment object with an added column
indicating whether each location passes the QC thresholds.
}
\description{
This function sets quality control (QC) thresholds for the number of
expressed genes in each location.
}
\details{
The function sets QC thresholds for the number of expressed genes
in each location based on the specified minimum and maximum detection values.
Locations failing to meet the thresholds are flagged, and a summary table
of filtered locations is printed.
}
\examples{
# 'sfe' is a SpatialFeatureExoeriment object

# Set QC thresholds for number of expressed genes in each location
sfe <- setQCthresh_GenesExpr(sfe, sample_id = "JBO019", min_t = 100)

}
\seealso{
\code{\link{.int_getSmplIDs}}, \code{\link[SpatialFeatureExperiment]{colData}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{QC,}
\keyword{SpatialFeatureExoeriment}
\keyword{expressed,}
\keyword{genes}
