% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setQCthresh.R
\name{setQCthresh_LibSize}
\alias{setQCthresh_LibSize}
\title{Set Quality Control Threshold for Library Size}
\usage{
setQCthresh_LibSize(sfe, sample_id = TRUE, min_t, max_t)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object containing count data.}

\item{sample_id}{A character string, \code{TRUE}, or \code{NULL} specifying
sample/image identifier(s). If \code{TRUE}, all samples/images are
considered. If \code{NULL}, the first available entry is considered.}

\item{min_t}{Numeric, the minimum library size threshold. If NA, no lower
threshold is applied. Default is NA.}

\item{max_t}{Numeric, the maximum library size threshold. If NA, no upper
threshold is applied. Default is NA.}
}
\value{
A SpatialFeatureExperiment object with a new column 'qc_lib_size'
indicating whether the library size passes the thresholds.
}
\description{
This function sets quality control thresholds for library size in a
SpatialFeatureExperiment object.
}
\details{
This function filters out locations based on the specified library
size thresholds. It also prints out a summary table showing the number of filtered
locations per sample.
}
\examples{
# 'sfe' is a SpatialFeatureExoeriment object

# Both filters
sfe <- setQCthresh_LibSize(sfe, sample_id = "JBO019", 500, 35000)

# Upper only
sfe <- setQCthresh_LibSize(sfe, sample_id = "JBO019", NA, 35000)

# Lower only
sfe <- setQCthresh_LibSize(sfe, sample_id = "JBO019", 500, NA)

}
\seealso{
\code{\link{.int_getSmplIDs}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{data}
\keyword{manipulation}
