% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setQCthresh.R
\name{setQCthresh_LowLogMean}
\alias{setQCthresh_LowLogMean}
\title{Set QC Thresholds for Low Logarithmic Mean in Gene Expression}
\usage{
setQCthresh_LowLogMean(msfe, threshold = 1, sample_id = TRUE)
}
\arguments{
\item{msfe}{A SpatialFeatureExperiment or a MetaSpatialFeatureExperiment
object.}

\item{threshold}{Numeric, the threshold for the logarithmic mean value.
Genes with mean values below this threshold will be flagged.}

\item{sample_id}{A character string, \code{TRUE}, or \code{NULL} specifying
sample/image identifier(s). If \code{TRUE}, all samples/images are
considered. If \code{NULL}, the first available entry is considered.}
}
\value{
A SpatialFeatureExperiment or a MetaSpatialFeatureExperiment
object with added column in rowData indicating whether each gene passes the
specified QC threshold for low logarithmic mean values.
}
\description{
This function sets quality control (QC) thresholds for low logarithmic mean
values in gene expression data. It is used after log transforming the gene
count matrix
}
\details{
The function sets QC thresholds for low logarithmic mean values in
gene expression data for each sample in the provided object. Genes failing
to meet the thresholds are flagged.
}
\examples{
# 'msfe' is a SpatialFeatureExoeriment or a MetaSpatialFeatureExoeriment,
# object

# Set QC thresholds for low logarithmic mean values of gene expression.
msfe <- setQCthresh_LowLogMean(msfe, threshold = 1, sample_id = TRUE)

}
\seealso{
\code{\link{.int_logLowMean}},
\code{\link[SpatialFeatureExperiment]{rowData}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{QC,}
\keyword{SpatialFeatureExoeriment}
\keyword{expression,}
\keyword{gene}
