% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setQCthresh.R
\name{setQCthresh_Mito}
\alias{setQCthresh_Mito}
\alias{setQCthresh}
\title{Set QC Threshold for Mitochondrial Content}
\usage{
setQCthresh_Mito(sfe, sample_id = TRUE, min_t, max_t)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object containing mitochondrial
content data (a colData column named subset_mito_percent)}

\item{sample_id}{A character string, \code{TRUE}, or \code{NULL} specifying
sample/image identifier(s). If \code{TRUE}, all samples/images are
considered. If \code{NULL}, the first available entry is considered.}

\item{min_t}{Numeric, the minimum threshold for mitochondrial content
percentage. Samples with percentages below this threshold will be flagged.
Use NA for no minimum threshold.}

\item{max_t}{Numeric, the maximum threshold for mitochondrial content
percentage. Samples with percentages above this threshold will be flagged.
Use NA for no maximum threshold.}
}
\value{
A modified SpatialFeatureExperiment object with an added column in
the colData indicating whether each location passes the QC thresholds for
mitochondrial content.
}
\description{
This function sets quality control (QC) thresholds for mitochondrial content
based on the percentage of mitochondrial gene expression form the total of
expressed genes in a location.
}
\details{
The function sets QC thresholds for mitochondrial content based on
the specified minimum and maximum percentage values. Locations failing to
meet the thresholds are flagged, and a summary table of filtered locations
is printed.
}
\examples{
# 'sfe' is a SpatialFeatureExoeriment object

# Set QC thresholds for mitochondrial content percentages above 30
sfe <- setQCthresh_Mito(sfe, sample_id = TRUE, max_t = 30)

}
\seealso{
\code{\link{.int_getSmplIDs}}, \code{\link[colData]{colData}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{QC,}
\keyword{SpatialFeatureExperiment}
\keyword{content,}
\keyword{mitochondrial}
