% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setQCthresh.R
\name{setQCthresh_NAs}
\alias{setQCthresh_NAs}
\title{Set QC Threshold for Missing Annotations}
\usage{
setQCthresh_NAs(sfe, sample_id = TRUE)
}
\arguments{
\item{sfe}{A SpatialFeatureExoeriment object containing annotation data.}

\item{sample_id}{A character string, \code{TRUE}, or \code{NULL} specifying
sample/image identifier(s). If \code{TRUE}, all samples/images are
considered. If \code{NULL}, the first available entry is considered.}
}
\value{
A modified SpatialFeatureExoeriment object with an added column  in
colData indicating whether each location has missing annotations.
}
\description{
This function sets quality control (QC) thresholds for missing annotations
based on the presence of NAs in the annotation column. This QC function is
mostly useful when there are relatively few parts of the tissue sample that
couldn't be annotated by an expert or another approach.
}
\details{
The function sets QC thresholds for missing annotations based on
the presence of NAs in the annotation column in colData. Locations with
missing annotations are flagged, and a summary table of filtered locations
is printed. his QC function is
mostly useful when there are relatively few parts of the tissue sample that
couldn't be annotated by an expert or another approach.
}
\examples{
# 'sfe' is a SpatialFeatureExoeriment object

# Set QC thresholds with for missing annotations
sfe <- setQCthresh_NAs(sfe, sample_id = TRUE)

}
\seealso{
\code{\link{.int_getSmplIDs}}, \code{\link[SpatialFeatureExperiment]{colData}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{QC,}
\keyword{SpatialFeatureExoeriment}
\keyword{annotations,}
\keyword{missing}
