% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setQCthresh.R
\name{setQCthresh_ZeroExpr}
\alias{setQCthresh_ZeroExpr}
\alias{setQCthresh,}
\title{Set QC Thresholds for Zero Expression}
\usage{
setQCthresh_ZeroExpr(msfe, sample_id = TRUE)
}
\arguments{
\item{msfe}{A SpatialFeatureExperiment or a MetaSpatialFeatureExperiment
object.}

\item{sample_id}{A character string, \code{TRUE}, or \code{NULL} specifying
sample/image identifier(s). If \code{TRUE}, all samples/images are
considered. If \code{NULL}, the first available entry is considered.}
}
\value{
A SpatialFeatureExperiment or a MetaSpatialFeatureExperiment
object with added columns in the rowData indicating whether each gene passes
the specified QC thresholds.
}
\description{
These functions set quality control (QC) thresholds for gene expression data
based on zero expression values.
}
\details{
This function, identifies genes without expression in each sample.
}
\examples{
# 'sfe' is a SpatialFeatureExoeriment object

# Set QC thresholds for zero expression
msfe <- setQCthresh_ZeroExpr(msfe, sample_id = TRUE)

}
\seealso{
\code{\link{.int_isZero}}, \code{\link{.int_logLowMean}},
\code{\link[colData]{colData}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{QC,}
\keyword{SpatialFeatureExoeriment}
\keyword{expression,}
\keyword{gene}
