% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_setQCthresh.R
\name{setQCtoDiscard_feat}
\alias{setQCtoDiscard_feat}
\alias{setQCtoDiscard_feat.SpatialFeatureExperiment}
\alias{setQCtoDiscard_feat.MetaSpatialFeatureExperiment}
\title{Set QC Thresholds for Discarding Features}
\usage{
setQCtoDiscard_feat(msfe, sample_id = TRUE, filters = TRUE)

\method{setQCtoDiscard_feat}{SpatialFeatureExperiment}(m_sfe, sample_id = TRUE, filters = TRUE)

\method{setQCtoDiscard_feat}{MetaSpatialFeatureExperiment}(m_sfe, sample_id = TRUE, filters = TRUE)
}
\arguments{
\item{sample_id}{A character string, \code{TRUE}, or \code{NULL} specifying
sample/image identifier(s). If \code{TRUE}, all samples/images are
considered. If \code{NULL}, the first available entry is considered.}

\item{filters}{Either TRUE to select all calculated QC metrics or a
character vector with the column names of the QC metrics to consider for
filtering features. QC metric columns begin with "qc_".}

\item{m_sfe}{A SpatialFeatureExperiment or a MetaSpatialFeatureExperiment.}
}
\value{
A SpatialFeatureExoeriment or a MetaSpatialFeatureExoeriment object
with features flagged for discard based on specified QC thresholds.
}
\description{
This function marks the features (genes) flagged for discarding based on
specified QC metrics.
}
\details{
The function marks the features (genes) flagged for discarding
based on the specified QC metrics. A feature failing to meet any of the
thresholds applied earlier is flagged.

A summary table of filtered features is printed. Users can choose to filter
based on all available QC metrics or specify a subset using the \code{filters}
argument.
}
\examples{
# 'msfe' is a SpatialFeatureExoeriment or a MetaSpatialFeatureExoeriment,
# object

# Mark the features (genes) flagged for discarding
msfe <- setQCtoDiscard_feat(msfe, filters = TRUE)

}
\seealso{
\code{\link{.int_featToDiscard}}, \code{\link{.int_summaryTable}},
\code{\link[SpatialFeatureExperiment]{colData}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{QC,}
\keyword{SpatialFeatureExoeriment}
\keyword{feature}
\keyword{filtering,}
