% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_setQCthresh.R
\name{setQCtoDiscard_loc}
\alias{setQCtoDiscard_loc}
\title{Select Locations to Discard}
\usage{
setQCtoDiscard_loc(sfe, sample_id = TRUE, filters = TRUE)
}
\arguments{
\item{sfe}{A SpatialFeatureExoeriment object containing QC metrics.}

\item{sample_id}{A character string, \code{TRUE}, or \code{NULL} specifying
sample/image identifier(s). If \code{TRUE}, all samples/images are
considered. If \code{NULL}, the first available entry is considered.}

\item{filters}{Either TRUE to select all calculated QC metrics or a
character vector with the column names of the QC metrics to consider for
filtering locations. QC metric columns begin with "qc_".}
}
\value{
A modified SpatialFeatureExoeriment object with an added column indicating whether each location passes the specified QC thresholds.
}
\description{
This function selects locations to discard based on specified QC metrics.
}
\details{
The function sets QC thresholds for discarding locations based on the specified QC metrics. Locations failing to meet the thresholds are flagged, and a summary table of filtered locations is printed. Users can choose to filter based on all available QC metrics or specify a subset using the \code{filters} argument.
}
\examples{
# Set QC thresholds with sample IDs for discarding locations based on all available QC metrics
se <- setQCtoDiscard_loc(sfe, sample_id = TRUE, filters = TRUE)

}
\seealso{
\code{\link{.int_getSmplIDs}}, \code{\link[SpatialFeatureExperiment]{colData}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{QC,}
\keyword{SpatialFeatureExoeriment}
\keyword{filtering,}
\keyword{location}
