% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sfe}
\alias{sfe}
\title{10X Visium Steatotic Human Liver - processed}
\format{
\subsection{\code{sfe}}{

A SpatialFeatureExperiment S4 class object with 8535 elements. Specifically,
8535 features and 1161 spots. Includes rowData with feature metadata and
statistics. Includes colData with spot metadata and statistics. Includes
colGeometry data with sf geometry objects. Includes raw and log2-normalised
UMI count data.

The QC filters used:
\itemize{
\item Spot-level QC
\itemize{
\item qc_lib_size: removed spots with library size (total UMI counts
in the spot) of < 1000 or > 45000.
\item qc_detected: removed spots with number of detected genes < 500
or > 6000.
\item qc_mito: removed spots with percentage of mitochondrial
expression > 22\%.
}
\item Gene-level QC
\itemize{
\item is_zero: removed any gene with no expression in the dataset.
\item is_mito: removed all mitochondrial genes.
\item is_logLow: remove genes with a sample mean of log2 UMI counts
<1. Sample mean is the mean of counts over the locations a gene is
present.
}
}
}
}
\source{
\url{https://www.livercellatlas.org/download.php}
}
\usage{
data(sfe)
}
\description{
A SpatialFeatureExperiment object containing a 10X Visium liver dataset from
the Liver Cell Atlas consortium. The sample used is the JBO019. The dataset
has passed QC for spot and feature selection.
}
\keyword{datasets}
