% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sfe_nmf}
\alias{sfe_nmf}
\title{NMF for the subset of Fibrotic Lung data}
\format{
\subsection{\code{sfe_nmf}}{

A dataframe containing the NMF result for the Fibrotic sample subset.
}
}
\source{
\url{https://doi.org/10.1038/s41588-024-01819-2}
}
\usage{
sfe_nmf
}
\description{
The lung dataset comes from the Franzen et. al., 2024, Nat. Genet.
publication \url{https://doi.org/10.1038/s41588-024-01819-2}. The Healthy and
Fibrotic samples here is the sample V19S23-092-D1 from the publication.
}
\details{
This object contains the results of running NMF dimensionality reduction to
be used for FGWC.

It is generated using the below command:
sfe_nmf <- fgwc_nmf(sfe, sample_id = "Fibrotic",
top_hvgs = top_hvgs[\link{"Fibrotic"}], ncomponents = 7)

To find the top_hvgs we used the whole Fibrotic sample data and the default
parameters in modeling gene variance and selecting topHVGs using
STExplorer's functions.
}
\keyword{datasets}
