% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sfe_raw}
\alias{sfe_raw}
\title{10X Visium Steatotic Human Liver - unprocessed}
\format{
\subsection{\code{sfe}}{

A SpatialFeatureExperiment S4 class object with 32738 elements. Specifically,
8535 features and 1185 spots.
}
}
\source{
\url{https://www.livercellatlas.org/download.php}
}
\usage{
data(sfe_raw)
}
\description{
A SpatialFeatureExperiment object containing a 10X Visium liver dataset from
the Liver Cell Atlas consortium. The sample used is the JBO019. The dataset
has NOT passed any QC and any preprocessing.
}
\keyword{datasets}
