% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_FunctionalClust.R
\name{viewCollections}
\alias{viewCollections}
\title{Look at Gene Set catagories and their subcategories}
\usage{
viewCollections()
}
\description{
This function is a wrapper around the "msigdbr_collections" function that
prints the possible combinations of categories and subcategories. It is
useful because if a category is matched with a subcategory that it shouldn't
then GSEA will return no results.
}
\examples{
# Check available collections
viewCollections()

}
