% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_STE.R
\name{wpca.ste}
\alias{wpca.ste}
\title{A basic WPCA with distance weighting.}
\usage{
wpca.ste(x, wt, ...)
}
\arguments{
\item{x}{Data matrix.}

\item{wt}{Weight vector.}

\item{...}{Additional arguments passed to the svd function.}
}
\value{
Returns the result of the singular value decomposition.
}
\description{
IMPORTANT: this function is not exported to be used alone.
Performs a basic Weighted-PCA with distance weighting.
}
\details{
This function performs a basic Weighted Principal Component Analysis (WPCA)
with distance weighting. The gene counts are centred by subtracting a
distance-weighted mean, and then the centred matrix is multiplied by the
square root of the weights. Finally, the Singular Value Decomposition (SVD)
is applied to the weighted and centred matrix.
}
\seealso{
\code{\link{svd}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{internal}
