% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpca_STE.R
\name{wt.median.ste}
\alias{wt.median.ste}
\title{Calculate a distance weighted median for use in Robust WPCA.}
\usage{
wt.median.ste(x, wt)
}
\arguments{
\item{x}{Data matrix.}

\item{wt}{Weight vector.}
}
\value{
Returns a vector of weighted medians, one for each column of the
input matrix.
}
\description{
IMPORTANT: this function is not exported to be used alone.
Calculates a weighted median for use in Robust Weighted Principal Component
Analysis (WPCA) with distance weighting.
}
\details{
This function calculates a weighted median for each column of the input data
matrix. The weighted median is computed by rearranging the data in ascending
order, computing the cumulative sum of weights, and identifying the index of
the median. The weighted median is then returned for each column.
}
\seealso{
\code{\link{order}}
}
\keyword{internal}
