% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_addPerGeneQC.R
\name{addPerGeneQC}
\alias{addPerGeneQC}
\title{Add feature-related QC metrics}
\usage{
addPerGeneQC(
  m_sfe,
  sample_id,
  assay = "counts",
  organism = "human",
  version = NULL,
  mirror = NULL,
  add = c("sparsity", "zeroexpr", "coefofvar", "exprstats", "none"),
  ...
)
}
\arguments{
\item{m_sfe}{An object of class SpatialFeatureExperiment or
MetaSpatialFeatureExperiment.}

\item{sample_id}{A character vector specifying the sample IDs to include
(only relevant if a MetaSpatialFeatureExperiment has been provided in the
\code{m_sfe} argument).}

\item{assay}{the name of the assay to use. Defaults to 'counts'.}

\item{organism}{The organism for which the annotation is used. Defaults to
"human". Other names that can be used: mouse, zebrafish, rat, fruitfly, pig,
worm, yeast.}

\item{version}{The ENSEMBL version of the annotation you want to use. It is
advised to use the annotation version that was used to create the .bam files.
If you don't want to add annotation to the genes leave it as NULL.}

\item{mirror}{Specify an Ensembl mirror to connect to. The valid options here
are 'www', 'uswest', 'useast', 'asia'. If no mirror is specified the primary
site at www.ensembl.org will be used. Mirrors are not available for the
Ensembl Genomes databases. The mirror and the version arguments cannot be
used together.}

\item{add}{Character string or a vector of characters. The vector needs to
contain one or any combination of "sparsity", "zeroexpr", "coefofvar",
"exprstats", and "none". Indicates which additional QC metrics to be
calculated. Defaults to "zeroexpr". Leave it as is since the other metrics
are not used downstream at the moment and can only increase size and running
times.}

\item{...}{further arguments passed to \code{addPerFeatureQC}, to pass
to \code{perFeatureQCMetrics}.}
}
\description{
A function to add a series of feature (gene)-related QC metrics.
}
\examples{
data(sfe_raw)
sfe <- addPerGeneQC(sfe_raw, assay = "counts", organism = "human",
version = NULL, mirror = NULL)

}
\seealso{
\code{\link{addPerFeatureQC}}, \code{\link{get.QC.Sparsity}},
\code{\link{get.QC.FindZeroExpr}}, \code{\link{get.QC.ExprStats}},
\code{\link{get.QC.CoefficientOfVar}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
