% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyQCthresh_.R, R/filterInTissue.R
\name{.int_cleanMetaData}
\alias{.int_cleanMetaData}
\title{Internal Function: Clean Up Duplicate Entries in the Metadata Slot of a
SpatialFeatureExperiment}
\usage{
.int_cleanMetaData(sfeOut)

.int_cleanMetaData(sfeOut)
}
\arguments{
\item{sfeOut}{A \code{SpatialFeatureExperiment} object.}
}
\value{
A modified \code{SpatialFeatureExperiment} object with duplicate
entries removed from the metadata slot.

A \code{SpatialFeatureExperiment} object with cleaned metadata
entries.
}
\description{
This internal function cleans up duplicate entries in the metadata slot of a
\code{SpatialFeatureExperiment} object.

This internal function cleans metadata entries in a SpatialFeatureExperiment
object. It ensures that duplicate entries in the metadata slot are removed,
preserving the original names. The function takes a SpatialFeatureExperiment
object (sfeOut) as input and returns the cleaned object.
}
\details{
This internal function ensures that the metadata slot of the
\code{SpatialFeatureExperiment} object does not contain duplicate entries.
It makes the names of the metadata unique and keeps the slots with the
original names.

This internal function efficiently cleans metadata entries in a
SpatialFeatureExperiment object. It takes a SpatialFeatureExperiment object
(sfeOut) as input and ensures that duplicate entries in the metadata slot
are removed. The function maintains the original names while making them
unique.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{function,}
\keyword{internal}
\keyword{metadata}
\keyword{spatial}
\keyword{transcriptomics}
\keyword{transcriptomics,}
