% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{.int_findElbowPointKnee}
\alias{.int_findElbowPointKnee}
\title{Internal Function: Identify Elbow Point Using the Knee Method}
\usage{
.int_findElbowPointKnee(index_data)
}
\arguments{
\item{index_data}{A data frame containing columns \code{k} (number of clusters)
and \code{IndexValue} (metric value).}
}
\value{
An integer representing the optimal number of clusters (\code{k}).
}
\description{
This function identifies the elbow point of a curve using the Knee method.
}
\details{
This function identifies the elbow point by calculating the distance of each
point from the line connecting the first and last points. The point with the
maximum distance is considered the elbow point.
}
\seealso{
\code{\link{fgwc_findOptimumK}} for finding the optimal number of
clusters.
}
\keyword{elbow-point}
\keyword{internal}
