% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_plot.R
\name{.int_getAnnotsAndColoursSFE}
\alias{.int_getAnnotsAndColoursSFE}
\title{Internal: Get Annotations and Colours}
\usage{
.int_getAnnotsAndColoursSFE(sfe, loc_annot, order_col)
}
\arguments{
\item{fgwc}{gwc class object as generated by \code{fgwcSTE}}

\item{loc-annot}{which annotations to add}

\item{order_rows}{order rows based on annotation or cluster}
}
\value{
a list
}
\description{
This function returns annotations label and colours for the heatmap
}
\author{
Eleftherios (Lefteris) Zormpas
}
