% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNeighbourGraph.R
\name{.int_getImgDtMSFE}
\alias{.int_getImgDtMSFE}
\title{Internal Function: Fetch Image Data and Transform to Raster}
\usage{
.int_getImgDtMSFE(sfe, image_id)
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object.}

\item{image_id}{Character vector specifying the image ID.}
}
\value{
A raster object containing the image data.
}
\description{
This internal function retrieves image data for a given
SpatialFeatureExperiment and transforms it into a raster format.
}
\details{
This function is an internal utility to fetch image data from a
SpatialFeatureExperiment using the specified image ID. The retrieved image
is then transformed into a raster format for further processing.
}
\seealso{
\code{\link{getImg}}, \code{\link{imgRaster}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{image}
\keyword{raster}
\keyword{spatial}
