% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metaSFE.R
\name{.int_getMSFEsmplID}
\alias{.int_getMSFEsmplID}
\title{Internal Function: Get SpatialFeatureExperiment Sample IDs}
\usage{
.int_getMSFEsmplID(msfe, sample_id)
}
\arguments{
\item{msfe}{A SpatialFeatureExperiment object.}

\item{sample_id}{Either a logical vector indicating the samples to include
(if TRUE, all samples are included), or a character vector specifying the
sample IDs to include.}
}
\value{
A character vector of selected sample IDs.
}
\description{
This internal function retrieves the sample IDs based on the provided
criteria.
}
