% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{fgwc_nmfFactorNumber}
\alias{fgwc_nmfFactorNumber}
\title{Identify the optimal number of factors}
\usage{
fgwc_nmfFactorNumber()
}
\arguments{
\item{name}{description}
}
\value{
An integer. The number of optimal factors.
}
\description{
This function uses different statistical methods to select the optimal
number of factors for a given dataset.
}
\seealso{
Other clustering functions: 
\code{\link{fgwc_nmf}()},
\code{\link{fgwc_params}()}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\concept{clustering functions}
\keyword{clustering}
\keyword{nmf}
\keyword{spatial-expression}
