% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_STE.R
\name{fgwc_params}
\alias{fgwc_params}
\alias{classic_params}
\alias{abc_params}
\title{Generate Parameters for Fuzzy Geographically Weighted Clustering}
\usage{
fgwc_params(
  algorithm = c("classic", "abc", "fpa", "gsa", "hho", "ifa", "pso", "tlbo"),
  ...
)

classic_params(
  ncluster,
  kind = "u",
  m = 1.5,
  distance = "manhattan",
  order = 1,
  alpha = 0.5,
  a = 2,
  b = 1,
  max.iter = 500,
  error = 1e-05,
  randomN = 1,
  uij = NA,
  vi = NA
)

abc_params(
  ncluster,
  m = 1.6,
  distance = "manhattan",
  order = 1,
  alpha = 0.5,
  a = 2,
  b = 1,
  error = 1e-05,
  max.iter = 100,
  randomN = 0,
  vi.dist = "uniform",
  nfood = 10,
  n.onlooker = 5,
  limit = 4,
  pso = FALSE,
  abc.same = 10
)
}
\arguments{
\item{algorithm}{The type of FGWC or optimization algorithm to use. Choose
from "classic", "abc", "fpa", "gsa", "hho", "ifa", "pso", "tlbo".}

\item{...}{arguments passed down to the specific algorithm function}

\item{ncluster}{The number of clusters.}

\item{kind}{Use \code{'u'} for the membership approach and \code{'v'} for
the centroid approach.}

\item{m}{Degree of fuzziness or fuzzifier (default is 2).}

\item{distance}{The distance metric between data and centroid (default is
"euclidean").}

\item{order}{Minkowski order (default is 1).}

\item{alpha}{The old membership effect with \verb{[0,1]}. If \code{alpha} equals
1, it will be the same as Fuzzy C-Means. If 0, it equals the neighborhood
effect.}

\item{a}{Spatial magnitude of distance (default is 1).}

\item{b}{Spatial magnitude of population (default is 1).}

\item{max.iter}{maximum iteration. Default is 500.}

\item{error}{error tolerance. Default is 1e-5.}

\item{randomN}{random seed for initialisation (if uij or vi is NA).
Default is 0.}

\item{uij}{Membership matrix initialization.}

\item{vi}{Centroid matrix initialization.}

\item{vi.dist}{a string of centroid population distribution between
"uniform" (default) and "normal". Can be defined as vi.dist= in opt_param.}

\item{nfood}{number of foods population. Can be defined as npar= in
opt_param.}

\item{n.onlooker}{number of onlooker bees, Can be defined as n.onlooker in
opt_param.}

\item{limit}{number of turns to eliminate food with no solutions. Can be
defined as limit in opt_param.}

\item{pso}{whether to add PSO term in bee's movement. Either TRUE or FALSE.
Can be defined as pso in opt_param.}

\item{abc.same}{number of consecutive unchanges to stop the iteration. Can
be defined as same= in opt_param. description}
}
\value{
A list of FGWC parameters or optimization algorithm
parameters.
}
\description{
This function generates parameters for Fuzzy Geographically Weighted
Clustering (FGWC) or optimization algorithms for FGWC.
}
\details{
depending on the algorithm the function calls internally the respective
function.
}
\examples{
\dontrun{
# Example usage:
params <- fgwc_params(algorithm = "classic", ncluster = 3, kind = "u", m = 2)
}

}
\seealso{
\code{\link[naspaclust]{fgwcuv}}, \code{\link[naspaclust]{abcfgwc}},
\code{\link[naspaclust]{fpafgwc}}, \code{\link[naspaclust]{gsafgwc}},
\code{\link[naspaclust]{hhofgwc}}, \code{\link[naspaclust]{ifafgwc}},
\code{\link[naspaclust]{psofgwc}}, \code{\link[naspaclust]{tlbofgwc}}

Other clustering functions: 
\code{\link{fgwc_nmf}()},
\code{\link{fgwc_nmfFactorNumber}()}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\concept{clustering functions}
\keyword{clustering}
\keyword{fgwc}
\keyword{fuzzy-logic}
\keyword{optimization}
