% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_gearyC.R
\name{gearyGlobalCTest}
\alias{gearyGlobalCTest}
\title{Geary's C Test for Spatial Autocorrelation}
\usage{
gearyGlobalCTest(
  m_sfe,
  sample_id = NULL,
  genes = TRUE,
  randomisation = TRUE,
  zero.policy = NULL,
  alternative = "greater",
  spChk = NULL,
  adjust.n = TRUE,
  mc.cores = getOption("mc.cores", 2L)
)
}
\arguments{
\item{m_sfe}{An object of class SpatialFeatureExperiment or
MetaSpatialFeatureExperiment.}

\item{sample_id}{A character vector specifying the sample IDs to include
(only relevant if a MetaSpatialFeatureExperiment has been provided in the
\code{m_sfe} argument).}

\item{genes}{TRUE or a named character vector with gene names for which the
SA statistic needs to be calculated. If left to TRUE then the SA statistic
is calculated for every gene.}

\item{randomisation}{Variance of I calculated under the assumption of
randomisation. If FALSE, normality.}

\item{zero.policy}{Default is NULL. If not changed then internally, it is
set to \code{attr(listw, "zero.policy")} as set when \code{listw} was
created. If TRUE, assign zero to the lagged value of zones without
neighbours. If FALSE, assign NA. If attribute not set, use the global option
value.}

\item{alternative}{A character string specifying the alternative hypothesis.
Must be one of "greater" (default), "less" or "two.sided".}

\item{spChk}{Should the data vector names be checked against the spatial
objects for identity integrity? TRUE or FALSE, default NULL to use
\code{get.spChkOption()}.}

\item{adjust.n}{Default TRUE. If FALSE, the number of observations is not
adjusted for no-neighbour observations. If TRUE, the number of observations
is adjusted.}

\item{mc.cores}{Argument from \code{link[parallel]{mclapply}}. The number of
cores to use, i.e., at most how many child processes will be run
simultaneously. The option is initialized from environment variable MC_CORES
if set. Must be at least one, and parallelisation requires at least two
cores.}
}
\value{
An SFE object with the below two columns added in the \code{rowData}:
\item{gearyC_test}{the value of the standard deviate of Geary's C.}
\item{gearyCPval_test}{the p-value of the test.}
}
\description{
This function is a wrapper of the \code{\link[spdep]{geary.test}} function
from the \code{spdep} package developed by Roger Bivand. It performs Geary's test
for spatial autocorrelation using a spatial weights matrix in weights list
form. The assumptions underlying the test are sensitive to the form of the
graph of neighbour relationships and other factors, and results may be
checked against those of \code{\link[spdep]{geary.mc}} permutations.
}
\details{
The derivation of the test (Cliff and Ord, 1981, p. 18) assumes
that the weights matrix is symmetric. For inherently non-symmetric matrices,
such as k-nearest neighbour matrices, \code{listw2U()} can be used to make
the matrix symmetric. In non-symmetric weights matrix cases, the variance of
the test statistic may be negative. Geary's C is affected by non-symmetric
weights under normality much more than Moran's I. From 0.4-35, the sign of
the standard deviate of C is changed to match Cliff and Ord (1973, p. 21).
}
\examples{
\dontrun{
# Export an SFE object from an MSFE object
sfe <- getSFE(msfe, "JBO019")
# A vector of gene expression of one gene over all locations
gene_exp <- assay(sfe, "logcounts")[1,]
# A listw object containing distance-based spatial weights for neighbours
listw <- listw
# Calculate Geary's C and statistical significance for this gene
geary <- gearyGlobalCTest(x = gene_exp, listw = listw, zero.policy = TRUE)
}

}
\references{
Cliff, A. D., Ord, J. K. 1981 Spatial processes, Pion, p. 21,
Cliff, A. D., Ord, J. K. 1973 Spatial Autocorrelation, Pion, pp. 15-16, 21;
Bivand RS, Wong DWS 2018 Comparing implementations of global and local
indicators of spatial association. TEST, 27(3), 716--748 \doi{10.1007/s11749
-018-0599-x}
}
\seealso{
\code{\link{geary}}, \code{\link{geary.mc}}, \code{\link{listw2U}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{autocorrelation}
\keyword{geary}
\keyword{listw}
\keyword{spatial}
