% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_geary.R
\name{gearyLocalC}
\alias{gearyLocalC}
\alias{gearyLocalCPerm}
\title{Local Geary's C Statistic for Spatial Autocorrelation}
\usage{
gearyLocalC(x, listw, ..., zero.policy = attr(listw, "zero.policy"))

gearyLocalCPerm(
  x,
  listw,
  nsim = 499,
  alternative = "two.sided",
  ...,
  zero.policy = attr(listw, "zero.policy"),
  iseed = NULL,
  no_repeat_in_row = FALSE
)
}
\arguments{
\item{x}{A numeric vector, list, or matrix. In univariate
context, \code{x} can be a numeric vector. In multivariate
context, provide a list or matrix.}

\item{listw}{A \code{listw} object created, e.g., by
\code{nb2listw}.}
}
\value{
Numeric vector containing Local Geary statistic
with attribute \code{pseudo-p} when \code{localC_perm()} is
used. \code{pseudo-p} is an 8-column matrix containing:
\item{E.Ci}{expectation of Local Geary statistic based on
permutation sample}
\item{Var.Ci}{variance of Local Geary based on permutation
sample}
\item{Z.Ci}{standard deviate of Local Geary based on
permutation sample}
\item{Pr()}{p-value of Local Geary statistic using
\code{pnorm()} based on permutation sample means and standard
deviations}
\item{Pr() Sim}{\code{rank()} and \code{punif()} of
observed statistic rank for \link{0, 1} p-values using
\code{alternative=}}
\item{Pr(folded) Sim}{simulation folded \link{0, 0.5} range
ranked p-value (based on
\url{https://github.com/pysal/esda/blob/4a63e0b5df1e754b17b5f1205b8cadcbecc
5e061/esda/crand.py#L211-L213})}
\item{Skewness}{output of \code{e1071::skewness()} for
permutation samples underlying standard deviates}
\item{Kurtosis}{output of \code{e1071::kurtosis()} for
permutation samples underlying standard deviates}
}
\description{
The Local Geary is a local adaptation of Geary's C
statistic of spatial autocorrelation. It uses squared
differences to measure dissimilarity, unlike the Local
Moran. Low values indicate positive spatial autocorrelation,
and large values refer to negative autocorrelation.
}
\details{
Inference for the Local Geary is based on permutation
approach, comparing the observed value to the reference
distribution under spatial randomness. \code{localC_perm()}
returns a pseudo p-value. This is not an analytical p-value
and should be used with care, based on permutations.

For more information about the usage cases and the
rest of the arguments visit the help page of \code{\link[spdep]{localC}}
Local Geary can be extended to a multivariate
context. When \code{x} is numeric, univariate Local Geary
is calculated. For multivariate, provide a list or matrix.

While not required in univariate context, standardized
Local Geary is calculated. Multivariate Local Geary is
\emph{always} standardized.

The univariate Local Geary is calculated as \eqn{c_i = \sum_j
w_{ij}(x_i - x_j)^2} and the multivariate Local Geary is
calculated as \eqn{c_{k,i} = \sum_{v=1}^{k} c_{v,i}} as
described in Anselin (2019).
}
\examples{
\dontrun{
# Export an SFE object from an MSFE object
sfe <- getSFE(msfe, "JBO019")
# A vector of gene expression of one gene over all locations
gene_exp <- assay(sfe, "logcounts")[1,]
# A listw object containing distance-based spatial weights for neighbours
listw <- listw
# Calculate Geary's C for this gene
geary <- gearyLocalC(x = gene_exp, listw = listw, zero.policy = TRUE)
}

\dontrun{
# Calculate  Geary's C and statistical significance for this gene
geary <- gearyLocalCPerm(x = gene_exp, listw = listw, nsim = 499,
zero.policy = TRUE)
}
}
\references{
Anselin, L. (2019). Local indicators of spatial
association—LISA. Spatial Statistics, 29, 10-27.
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{autocorrelation}
\keyword{geary}
\keyword{listw}
\keyword{spatial}
