% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_getisG.R
\name{getisGlobalGTest}
\alias{getisGlobalGTest}
\title{Global G test for spatial autocorrelation}
\usage{
getisGlobalGTest(
  x,
  listw,
  zero.policy = NULL,
  alternative = "greater",
  spChk = NULL,
  adjust.n = TRUE,
  B1correct = TRUE,
  adjust.x = TRUE,
  Arc_all_x = FALSE
)
}
\arguments{
\item{x}{Numeric vector the same length as the neighbors
list in listw}

\item{listw}{\code{listw} object created, for example,
by \code{nb2listw}; if a sequence of distance bands is
to be used, it is recommended that the weights style be
binary (one of \code{c("B", "C", "U")}).}

\item{zero.policy}{Default \code{attr(listw,
  "zero.policy")} as set when \code{listw} was created. If
attribute not set, use global option value. If TRUE,
assign zero to the lagged value of zones without
neighbors; if FALSE, assign NA.}

\item{alternative}{A character string specifying the
alternative hypothesis, must be one of "greater"
(default), "less" or "two.sided".}

\item{spChk}{Should the data vector names be checked
against the spatial objects for identity integrity,
TRUE, or FALSE, default NULL to use
\code{get.spChkOption()}.}

\item{adjust.n}{Default TRUE, if FALSE the number of
observations is not adjusted for no-neighbor
observations, if TRUE, the number of observations is
adjusted.}

\item{B1correct}{Default TRUE, if TRUE, the erratum
referenced below: "On page 195, the coefficient of W2
in B1, (just below center of the page) should be 6,
not 3." is applied; if FALSE, 3 is used (as in CrimeStat
IV).}

\item{adjust.x}{Default TRUE, if TRUE, x values of
observations with no neighbors are omitted in the
denominator of G.}

\item{Arc_all_x}{Default FALSE, if Arc_all_x=TRUE and
adjust.x=TRUE, use the full x vector in part of the
denominator term for G.}
}
\value{
A list with class \code{htest} containing the
following components:
\item{statistic}{the value of the standard deviate
of Getis-Ord G.}
\item{p.value}{the p-value of the test.}
\item{estimate}{the value of the observed statistic,
its expectation and variance.}
\item{alternative}{a character string describing the
alternative hypothesis.}
\item{data.name}{a character string giving the
name(s) of the data.}
}
\description{
This function is a wrapper of the \code{\link[spdep]{moran}} function from
the \code{spdep} package developed by Roger Bivand. The global G statistic for
spatial autocorrelation, complementing the local Gi LISA measures:
\code{\link{localG}}.
}
\examples{
\dontrun{
# Example usage:

# Export an SFE object from an MSFE object
sfe <- getSFE(msfe, "JBO019")

# A vector of gene expression of one gene over all locations
gene_exp <- assay(sfe, "logcounts")[1,]

# A listw object containing distance-based spatial weights for neighbours
listw <- listw

# Calculate Moran's I for this gene
getis <- getisGlobalGTest(x = gene_exp, listw = listw,zero.policy = TRUE)
}

}
\references{
Getis. A, Ord, J. K. 1992 The analysis of spatial association by
use of distance statistics, \emph{Geographical Analysis}, 24, p. 195; see
also Getis. A, Ord, J. K. 1993 Erratum, \emph{Geographical Analysis}, 25,
p. 276; Bivand RS, Wong DWS 2018 Comparing implementations of global and
local indicators of spatial association. TEST, 27(3), 716--748 \doi{10.1007/
s11749-018-0599-x}
}
\seealso{
\code{\link{localG}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
