% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metaSFE.R
\name{getMultiSFE}
\alias{getMultiSFE}
\title{Method to retrieve multiple SFE objects as a single SFE object from within
an MSFE object}
\usage{
getMultiSFE(x, sample_ids)
}
\arguments{
\item{x}{An instance of the MetaSpatialFeatureExperiment class.}

\item{sample_ids}{Character string. The sample IDs from the SFE objects to
be combined and retrieved as one SFE object}
}
\value{
An SFE object with multiple samples inside.
}
\description{
Method to retrieve multiple SFE objects as a single SFE object from within
an MSFE object
}
\details{
During merging the SFE objects, the common columns from the rowData are
removed to allow merging. However, the unique rowData columns from each SFE
object are kept. Additionally, the "gene_name", "id", "mean", "detected",
and "total" columns are generated again. Only this time, the "mean",
"detected", and "total" values are calculated based on the gene counts in
ALL samples and NOT in a per sample way.

The colData are easier to merge since there we are performing an rbind
instead of the cbind we have to do in the rowData. As a result all columns
are kept.
}
