% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metaSFE.R
\name{getSFE}
\alias{getSFE}
\title{Method to access a SpatialFeatureExperiment by name}
\usage{
getSFE(x, sample_id = NULL)
}
\arguments{
\item{x}{An instance of the MetaSpatialFeatureExperiment class.}

\item{sample_id}{Character string. The name of the experiment/sample for
the SFE to be accessed. If left to NULL, then the first sfe object in the
list is selected.}
}
\value{
The SFE object associated with the provided name.
}
\description{
Method to access a SpatialFeatureExperiment by name
}
