% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_STE.R
\name{gwr_stats}
\alias{gwr_stats}
\alias{gwr_stats.gwrm}
\alias{gwr_stats.SF}
\title{GWR summary table}
\usage{
gwr_stats(gwr, stat = NULL)

\method{gwr_stats}{gwrm}(gwr)

\method{gwr_stats}{SF}(gwr, stat)
}
\arguments{
\item{gwr}{A GWR object as created by \code{gwrSTE}, or the SDF data frame from
within the GWR object converted to an SF class object using
\code{gwr_toSF}.}

\item{stat}{Character string. It must be a column name from the \code{gwr$SDF}
data frame. It is essentially, the statistic for which to
provide a summary.}
}
\value{
A summary table containing descriptive statistics for the GWR object.
}
\description{
This function generates a summary table for a Geographically Weighted
Regression (GWR) object.
}
\author{
Eleftherios Zormpas
}
