% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_STE.R
\name{gwr_stats.gwr}
\alias{gwr_stats.gwr}
\title{GWR summary table}
\usage{
gwr_stats.gwr(gwr)
}
\arguments{
\item{gwr}{A GWR object as created by \code{gwrSTE}, or the SDF data frame from
within the GWR object converted to an SF class object using
\code{gwr_toSF}.}

\item{stat}{Character string. It must be a column name from the \code{gwr$SDF}
data frame. It is essentially, the statistic for which to
provide a summary.}
}
\value{
A summary table containing descriptive statistics for the GWR object.
}
\description{
This function generates a summary table for a Geographically Weighted
Regression (GWR) object.
}
\author{
Eleftherios Zormpas
}
