% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers}
\alias{markers}
\title{Set of liver marker genes}
\format{
\subsection{\code{data.frame}}{

A 4-column data frame with marker genes and the cell types. Includes the
below columns:
\itemize{
\item gene.name: the gene names
\item ensg.ID: the ENSGene IDs
\item Type: the cell type
\item Subtype: the cell subtype
}
Each column is \code{character}.
}
}
\usage{
data(markers)
}
\description{
Marker genes for liver cell types and subtypes. Used for the heatmaps and
subclustering of FGWC
}
\keyword{datasets}
