% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_moran.R
\name{moranGlobalI}
\alias{moranGlobalI}
\title{Compute Moran's I}
\usage{
moranGlobalI(x, listw, n = NULL, S0 = NULL, zero.policy = NULL, NAOK = FALSE)
}
\arguments{
\item{x}{a numeric vector same length as neighbours in listw}

\item{listw}{\code{listw} object, e.g., by \code{nb2listw}}

\item{n}{number of zones}

\item{S0}{global sum of weights}

\item{zero.policy}{default \code{attr(listw, "zero.policy")}}

\item{NAOK}{if 'TRUE' pass 'NA' or 'NaN' or 'Inf' to foreign function}
}
\value{
list of
\item{I}{Moran's I}
\item{K}{sample kurtosis of x}
}
\description{
This function is a wrapper of the \code{\link[spdep]{moran}} function from
the \code{spdep} package developed by Roger Bivand. It is a function to compute
Moran's I, called by \code{moran.test} and \code{moran.mc}.
}
\examples{
\dontrun{
# Example usage:

# Export an SFE object from an MSFE object
sfe <- getSFE(msfe, "JBO019")

# A vector of gene expression of one gene over all locations
gene_exp <- assay(sfe, "logcounts")[1,]

# A listw object containing distance-based spatial weights for neighbours
listw <- listw

# Calculate Moran's I for this gene
moran <- moranGlobalI(x = gene_exp, listw = listw,zero.policy = TRUE)
}

}
\references{
Cliff, A. D., Ord, J. K. 1981 Spatial processes, Pion, p. 17.
}
\seealso{
\code{\link{moran.test}}, \code{\link{moran.mc}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
