% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_moranI.R
\name{moranGlobalIPerm}
\alias{moranGlobalIPerm}
\title{Permutation test for Moran's I statistic}
\usage{
moranGlobalIPerm(
  m_sfe,
  sample_id = NULL,
  genes = TRUE,
  nsim = 999,
  zero.policy = NULL,
  alternative = "greater",
  na.action = na.fail,
  spChk = NULL,
  return_boot = FALSE,
  adjust.n = TRUE,
  mc.cores = getOption("mc.cores", 2L)
)
}
\arguments{
\item{m_sfe}{An object of class SpatialFeatureExperiment or
MetaSpatialFeatureExperiment.}

\item{sample_id}{A character vector specifying the sample IDs to include
(only relevant if a MetaSpatialFeatureExperiment has been provided in the
\code{m_sfe} argument).}

\item{genes}{TRUE or a named character vector with gene names for which the
SA statistic needs to be calculated. If left to TRUE then the SA statistic
is calculated for every gene.}

\item{nsim}{Number of permutations}

\item{zero.policy}{Default is NULL. If not changed then internally, it is
set to \code{attr(listw, "zero.policy")} as set when \code{listw} was
created. If TRUE, assign zero to the lagged value of zones without
neighbours. If FALSE, assign NA. If attribute not set, use the global option
value.}

\item{alternative}{Character string specifying the alternative hypothesis.
Options: "greater" (default), "two.sided", or "less".}

\item{na.action}{Function (default \code{na.fail}). Options: \code{na.omit},
\code{na.exclude}. The weights list is subsetted to remove NAs in the data.
\code{na.pass} is not permitted because it is meaningless in a permutation
test.}

\item{spChk}{Check data vector names against spatial objects for identity
integrity. TRUE or FALSE, default NULL to use \code{get.spChkOption()}}

\item{return_boot}{Return an object of class \code{boot} from the equivalent
permutation bootstrap.}

\item{adjust.n}{Default TRUE. If FALSE, the number of observations is not
adjusted for no-neighbour observations. If TRUE, the number of observations
is adjusted.}

\item{mc.cores}{Argument from \code{link[parallel]{mclapply}}. The number of
cores to use, i.e., at most how many child processes will be run
simultaneously. The option is initialized from environment variable MC_CORES
if set. Must be at least one, and parallelisation requires at least two
cores.}
}
\value{
An SFE object with the below two columns added in the \code{rowData}:
\item{I}{the value of the observed Moran's I. - replaces other global I
calculations since the I is always going to be the same.}
\item{moranPval_perm}{the pseudo p-value of the test.}
}
\description{
This function is a wrapper of the \code{\link[spdep]{moran.mc}} function from
the \code{spdep} package developed by Roger Bivand. A permutation test for
Moran's I statistic calculated by using nsim random
permutations of x for the given spatial weighting scheme, establishing the
rank of the observed statistic in relation to the nsim simulated values.
}
\examples{
\dontrun{
# Example usage:

# Export an SFE object from an MSFE object
sfe <- getSFE(msfe, "JBO019")

# A vector of gene expression of one gene over all locations
gene_exp <- assay(sfe, "logcounts")[1,]

# A listw object containing distance-based spatial weights for neighbours
listw <- listw

# Calculate Moran's I for this gene
moran <- moranGlobalIPerm(x = gene_exp, listw = listw, nsim = 499,
zero.policy = TRUE)
}

}
\references{
Cliff, A. D., Ord, J. K. 1981 Spatial processes, Pion, p. 63-5.
}
\seealso{
\code{\link{moran}}, \code{\link{moran.test}}

\code{\link{moran.test}}, \code{\link{moran}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
