% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{normaliseCounts}
\alias{normaliseCounts}
\alias{normaliseCounts.SpatialFeatureExperiment}
\alias{normaliseCounts.MetaSpatialFeatureExperiment}
\title{Normalise Counts in a Spatial Transcriptomics Experiment}
\usage{
normaliseCounts(m_sfe, sample_id, method = c("log"), ...)

\method{normaliseCounts}{SpatialFeatureExperiment}(m_sfe, method = c("log"), ...)

\method{normaliseCounts}{MetaSpatialFeatureExperiment}(m_sfe, method = c("log"), ...)
}
\arguments{
\item{m_sfe}{An object of class SpatialFeatureExperiment or
MetaSpatialFeatureExperiment.}

\item{sample_id}{A character string specifying unique sample identifiers,
one for each directory specified via samples. This parameter is ignored if
the names of samples are not null (!is.null(names(samples))).}

\item{method}{The normalisation method to be applied. Currently, only "log"
transformation is supported.}

\item{...}{Additional arguments to be passed to the normalisation function.}
}
\value{
A normalised SpatialFeatureExperiment object.
}
\description{
This function normalises counts in a spatial transcriptomics experiment
using specified methods.
}
\details{
The function normalises counts in a spatial transcriptomics
experiment using specified methods. Currently, only the "log" transformation
method is supported. Additional arguments can be passed to the underlying
normalisation function.
}
\examples{
\dontrun{
# Normalise counts using log transformation
# msfe <- # MetaSpatialFeatureExperiment object
msfe <- normaliseCounts(msfe, method = "log", base = 2)
}

}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{SpatialFeatureExperiment}
\keyword{log}
\keyword{normalisation}
\keyword{transformation}
