% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgwc_plot.R
\name{plotFGWC_multiViolin}
\alias{plotFGWC_multiViolin}
\title{Plot Membership percentages from FGWC per annotation or winning cluster}
\usage{
plotFGWC_multiViolin(fgwc)
}
\arguments{
\item{fgwc}{The result of FGWC clustering.}
}
\value{
A ggplot object.
}
\description{
This function plots the results of Fuzzy Geographically Weighted Clustering
(FGWC) for multiple clusters. It plots the membership percentage of each
cluster in each location as violin plots.
}
\details{
The function produces a set of violin plots. One set per annotation or
winning cluster
}
\seealso{
\code{\link{fgwc_nmf}}, \code{\link{plotFGWC_single}},
\code{\link{plotFGWC_heatmap}}

Other plotting functions: 
\code{\link{plotFGWC_markersHeatmap}()},
\code{\link{plotFGWC_multiMap}()},
\code{\link{plotFGWC_nmfFactorsMap}()},
\code{\link{plotFGWC_singleMap}()}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\concept{plotting functions}
\keyword{clustering}
\keyword{fgwc}
\keyword{fuzzy-logic}
\keyword{ggplot2}
\keyword{spatial}
\keyword{visualization}
