% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{plotQC_filtered}
\alias{plotQC_filtered}
\title{Plot Quality Control Filtered Locations Map in a SpatialFeatureExperiment}
\usage{
plotQC_filtered(
  m_sfe,
  metric = c("libsize", "mito", "NAs", "detected", "cellCount", "discard", "custom"),
  sample_id = TRUE,
  type = c("spot", "hex", "cntd"),
  metric_name = NULL,
  metric_lab = NULL,
  ...
)
}
\arguments{
\item{m_sfe}{A SpatialFeatureExperiment or MetaSFE object.}

\item{metric}{Character vector specifying the metric, either "libsize",
"mito", "NAs", "detected", "cellCount", "discard", or custom. If custom is
selected then you have to provide the \code{metric_name} and \code{metric_lab}
arguments too.}

\item{sample_id}{Character string, TRUE, or NULL specifying sample/image
identifier(s). TRUE is equivalent to all samples/images, and NULL specifies
the first available entry.}

\item{type}{Character vector specifying the spot type, either "spot", "cntd"
or "hex".}

\item{metric_name}{Character string. The custom metric's column name as
found in the \code{colData}.}

\item{metric_lab}{Character string. The title you want for the map's legend.}

\item{...}{Additional arguments to be passed to the underlying \code{geom_sf}
plotting function.}
}
\description{
This function plots the quality control filtered locations map in a
SpatialFeatureExperiment, providing options for different metrics and
sample IDs.
}
\details{
This function plots the quality control filtered locations map in a
SpatialFeatureExperiment. Users can specify the metric (libsize, mito, NAs,
detected, cellCount, or discard) and sample ID.
}
\examples{
\dontrun{
# Example usage:
data(sfe)
plotQC_filtered(sfe, metric = "libsize")
}

}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{filtering}
\keyword{ggplot2}
\keyword{mapping}
\keyword{quality-control}
\keyword{spatial}
