% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setQCthresh.R
\name{setQCtoDiscard_feat}
\alias{setQCtoDiscard_feat}
\alias{setQCtoDiscard}
\title{Set QC Thresholds for Discarding Features}
\usage{
setQCtoDiscard_feat(msfe, filters = TRUE)
}
\arguments{
\item{msfe}{A SpatialFeatureExoeriment or a MetaSpatialFeatureExoeriment.}

\item{filters}{Either TRUE to select all calculated QC metrics or a
character vector with the column names of the QC metrics to consider for
filtering features. QC metric columns begin with "qc_".}
}
\value{
A SpatialFeatureExoeriment or a MetaSpatialFeatureExoeriment object
with features flagged for discard based on specified QC thresholds.
}
\description{
This function marks the features (genes) flagged for discarding based on
specified QC metrics.
}
\details{
The function marks the features (genes) flagged for discarding
based on the specified QC metrics. A feature failing to meet any of the
thresholds applied earlier is flagged.

A summary table of filtered features is printed. Users can choose to filter
based on all available QC metrics or specify a subset using the \code{filters}
argument.
}
\examples{
# 'msfe' is a SpatialFeatureExoeriment or a MetaSpatialFeatureExoeriment,
# object

# Mark the features (genes) flagged for discarding
msfe <- setQCtoDiscard_feat(msfe, filters = TRUE)

}
\seealso{
\code{\link{.int_featToDiscard}}, \code{\link{.int_summaryTable}},
\code{\link[colData]{colData}}
}
\author{
Eleftherios (Lefteris) Zormpas
}
\keyword{QC,}
\keyword{SpatialFeatureExoeriment}
\keyword{feature}
\keyword{filtering,}
