% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sfe}
\alias{sfe}
\title{Subset of Fibrotic Lung data}
\format{
\subsection{\code{sfe}}{

A SpatialFeatureExperiment class object containing the Fibrotic sample subset.
}
}
\source{
\url{https://doi.org/10.1038/s41588-024-01819-2}
}
\usage{
sfe
}
\description{
The lung dataset comes from the Franzen et. al., 2024, Nat. Genet.
publication \url{https://doi.org/10.1038/s41588-024-01819-2}. The Healthy and
Fibrotic samples here is the sample V19S23-092-D1 from the publication.
}
\details{
This object contains a subset of the Fibrotic sample. Specifically the top
left corner as defined by the below selection:

spatialCoords(sfe)\link{, "pxl_row_in_fullres"} > 8000 & spatialCoords(sfe)\link{, "pxl_col_in_fullres"} < 7000

It is done so that the vignettes render faster and to be used for the
function examples.
}
\keyword{datasets}
